% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_att_to_sf}
\alias{add_att_to_sf}
\title{Add attribute to simple features object}
\usage{
add_att_to_sf(area_sf, att_data_tb, att_data_desc_1L_chr)
}
\arguments{
\item{area_sf}{Area (a simple features object)}

\item{att_data_tb}{Attribute data (a tibble)}

\item{att_data_desc_1L_chr}{Attribute data description (a character vector of length one)}
}
\value{
Updated area (a simple features object)
}
\description{
add_att_to_sf() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add attribute to simple features object. Function argument area_sf specifies the object to be updated. The function returns Updated area (a simple features object).
}
\keyword{internal}
