% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attr_to_sf.R
\name{add_attr_list_to_sf}
\alias{add_attr_list_to_sf}
\title{add_attr_list_to_sf}
\usage{
add_attr_list_to_sf(x, y, area_unit, boundary_year, data_lookup_tb,
  sub_div_unit)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{area_unit}{PARAM_DESCRIPTION}

\item{boundary_year}{PARAM_DESCRIPTION}

\item{data_lookup_tb}{PARAM_DESCRIPTION}

\item{sub_div_unit}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready.utils]{data_get}}
}
