% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_attr_recrly_to_sf}
\alias{add_attr_recrly_to_sf}
\title{Add attr recrly to}
\usage{
add_attr_recrly_to_sf(
  input_ls,
  sub_div_unit = NULL,
  area_unit,
  boundary_year,
  attribute_data
)
}
\arguments{
\item{input_ls}{Input (a list)}

\item{sub_div_unit}{PARAM_DESCRIPTION, Default: NULL}

\item{area_unit}{PARAM_DESCRIPTION}

\item{boundary_year}{PARAM_DESCRIPTION}

\item{attribute_data}{PARAM_DESCRIPTION}
}
\description{
add_attr_recrly_to_sf() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add attr recrly to simple features object. Function argument input_ls specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
