% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_attr_tb_to_data_pack_lup}
\alias{add_attr_tb_to_data_pack_lup}
\title{Add attr tibble to data pack}
\usage{
add_attr_tb_to_data_pack_lup(
  data_pack_lup,
  attr_tb,
  object_name,
  area_type,
  area_bound_yr,
  region,
  year,
  year_start,
  year_end,
  main_feature
)
}
\arguments{
\item{data_pack_lup}{Data pack (a lookup table)}

\item{attr_tb}{Attr (a tibble)}

\item{object_name}{PARAM_DESCRIPTION}

\item{area_type}{PARAM_DESCRIPTION}

\item{area_bound_yr}{PARAM_DESCRIPTION}

\item{region}{PARAM_DESCRIPTION}

\item{year}{PARAM_DESCRIPTION}

\item{year_start}{PARAM_DESCRIPTION}

\item{year_end}{PARAM_DESCRIPTION}

\item{main_feature}{PARAM_DESCRIPTION}
}
\description{
add_attr_tb_to_data_pack_lup() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add attr tibble to data pack lookup table. Function argument data_pack_lup specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
