% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_attr_tb_to_data_pack_lup_from_arg_list}
\alias{add_attr_tb_to_data_pack_lup_from_arg_list}
\title{Add attr tibble to data pack lookup table from argument list}
\usage{
add_attr_tb_to_data_pack_lup_from_arg_list(x, y)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}
}
\description{
add_attr_tb_to_data_pack_lup_from_arg_list() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add attr tibble to data pack lookup table from argument list. Function argument x specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
