% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_attr_to_global}
\alias{add_attr_to_global}
\title{Add attr to global}
\usage{
add_attr_to_global(combined_ste_ppr_ls, object_name_stub)
}
\arguments{
\item{combined_ste_ppr_ls}{Combined ste ppr (a list)}

\item{object_name_stub}{PARAM_DESCRIPTION}
}
\description{
add_attr_to_global() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add attr to global. Function argument combined_ste_ppr_ls specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
