% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attr_to_sf.R
\name{add_attr_to_sf}
\alias{add_attr_to_sf}
\title{Adds area attributes to a spatial object.}
\usage{
add_attr_to_sf(area_unit, area_sf, attr_data_tb, attr_data_desc,
  seifa_deciles_by_unit, boundary_year = "2016")
}
\arguments{
\item{area_unit}{A String specifying that the area type is either "LGA" or "SA2".}

\item{area_sf}{A SF opject, with the boundary geometeries of the specified spatial unit type.}
}
\value{
Returns a SF.
}
\description{
Merges a SF file with population data relating to areas described in that file.
}
\details{
Makes data transformations (variable name, changes from strings to factors) necessary
to ensure that merged objects are compatible.
}
\examples{

}
