% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_attribute_to_data_pack}
\alias{add_attribute_to_data_pack}
\title{Add attribute to data pack}
\usage{
add_attribute_to_data_pack(combined_ste_ppr_ls, object_name_stub)
}
\arguments{
\item{combined_ste_ppr_ls}{Combined ste ppr (a list)}

\item{object_name_stub}{PARAM_DESCRIPTION}
}
\description{
add_attribute_to_data_pack() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add attribute to data pack. Function argument combined_ste_ppr_ls specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
