% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attr_to_data_pack.R
\name{add_attribute_to_data_pack_from_tb}
\alias{add_attribute_to_data_pack_from_tb}
\title{add_attribute_to_data_pack_from_tb}
\usage{
add_attribute_to_data_pack_from_tb(attr_tb, object_name)
}
\arguments{
\item{attr_tb}{PARAM_DESCRIPTION}

\item{object_name}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
