% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_pack.R
\name{add_data_pack_from_script}
\alias{add_data_pack_from_script}
\title{add_data_pack_from_script}
\usage{
add_data_pack_from_script(x, lookup_tbs_r4, merge_sfs_vec, processed_dir,
  raw_data_dir, pckg_name, overwrite_lgl)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{lookup_tbs_r4}{PARAM_DESCRIPTION}

\item{merge_sfs_vec}{PARAM_DESCRIPTION}

\item{processed_dir}{PARAM_DESCRIPTION}

\item{raw_data_dir}{PARAM_DESCRIPTION}

\item{pckg_name}{PARAM_DESCRIPTION}

\item{overwrite_lgl}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{pull}}
}
