% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_data_pack_lup}
\alias{add_data_pack_lup}
\title{Add data pack}
\usage{
add_data_pack_lup(
  lookup_tbs_r4,
  tb_data_type = "Geometry",
  template_ls = NULL,
  pckg_name
)
}
\arguments{
\item{lookup_tbs_r4}{Lookup tibbles (a ready4 S4)}

\item{tb_data_type}{PARAM_DESCRIPTION, Default: 'Geometry'}

\item{template_ls}{Template (a list), Default: NULL}

\item{pckg_name}{PARAM_DESCRIPTION}
}
\description{
add_data_pack_lup() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add data pack lookup table. Function argument lookup_tbs_r4 specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
