% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_dynamic_sp_vars_to_sf}
\alias{add_dynamic_sp_vars_to_sf}
\title{Add dynamic sp vars to}
\usage{
add_dynamic_sp_vars_to_sf(
  dynamic_sp_vars_sf,
  pop_attr_sf,
  age_sex_pop_resolution,
  age_sex_var_name,
  popl_var_prefix,
  data_year,
  crs_nbr_vec
)
}
\arguments{
\item{dynamic_sp_vars_sf}{Dynamic sp vars (a simple features object)}

\item{pop_attr_sf}{Pop attr (a simple features object)}

\item{age_sex_pop_resolution}{PARAM_DESCRIPTION}

\item{age_sex_var_name}{PARAM_DESCRIPTION}

\item{popl_var_prefix}{PARAM_DESCRIPTION}

\item{data_year}{PARAM_DESCRIPTION}

\item{crs_nbr_vec}{PARAM_DESCRIPTION}
}
\description{
add_dynamic_sp_vars_to_sf() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add dynamic sp vars to simple features object. Function argument dynamic_sp_vars_sf specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
