% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_sfs_keep_counts.R
\name{add_feature_areas}
\alias{add_feature_areas}
\title{add_feature_areas}
\usage{
add_feature_areas(sf, data_type, data_year, feature,
  popl_var_prefix = NULL)
}
\arguments{
\item{sf}{PARAM_DESCRIPTION}

\item{data_type}{PARAM_DESCRIPTION}

\item{data_year}{PARAM_DESCRIPTION}

\item{feature}{PARAM_DESCRIPTION}

\item{popl_var_prefix}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{funs}}
 \code{\link[rlang]{sym}}
 \code{\link[sf]{geos_measures}}
 \code{\link[units]{set_units}}
}
