% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_km_sqd_by_group}
\alias{add_km_sqd_by_group}
\title{Add kilometre squared by group}
\usage{
add_km_sqd_by_group(
  geometry_sf,
  feature_nm_1L_chr,
  group_by_var_1L_chr,
  prefix_1L_chr = "whl_",
  suffix_1L_chr = "_area"
)
}
\arguments{
\item{geometry_sf}{Geometry (a simple features object)}

\item{feature_nm_1L_chr}{Feature name (a character vector of length one)}

\item{group_by_var_1L_chr}{Group by variable (a character vector of length one)}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'whl_'}

\item{suffix_1L_chr}{Suffix (a character vector of length one), Default: '_area'}
}
\value{
Geometry (a simple features object)
}
\description{
add_km_sqd_by_group() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add kilometre squared by group. Function argument geometry_sf specifies the object to be updated. The function returns Geometry (a simple features object).
}
\keyword{internal}
