% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_sfs_keep_counts.R
\name{add_kmsq_area_all_features}
\alias{add_kmsq_area_all_features}
\title{add_kmsq_area_all_features}
\usage{
add_kmsq_area_all_features(sf, feature_nm)
}
\arguments{
\item{sf}{PARAM_DESCRIPTION}

\item{feature_nm}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
 \code{\link[rlang]{sym}}
 \code{\link[sf]{geos_measures}}
 \code{\link[units]{set_units}}
}
