% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_kmsq_area_by_group}
\alias{add_kmsq_area_by_group}
\title{Add kmsq area by group}
\usage{
add_kmsq_area_by_group(
  sf,
  group_by_var,
  feature_nm,
  prefix = "whl_",
  suffix = "_area"
)
}
\arguments{
\item{sf}{Simple features object (a simple features object)}

\item{group_by_var}{PARAM_DESCRIPTION}

\item{feature_nm}{PARAM_DESCRIPTION}

\item{prefix}{PARAM_DESCRIPTION, Default: 'whl_'}

\item{suffix}{PARAM_DESCRIPTION, Default: '_area'}
}
\description{
add_kmsq_area_by_group() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add kmsq area by group. Function argument sf specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
