% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_names}
\alias{add_names}
\title{Add names}
\usage{
add_names(ds_tb)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}
}
\value{
Dataset (a tibble)
}
\description{
add_names() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add names. Function argument ds_tb specifies the object to be updated. The function returns Dataset (a tibble).
}
\keyword{internal}
