% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_add_path_col.R
\name{add_path_col}
\alias{add_path_col}
\alias{add_path_col,ready4_sp_data_pack_lup-method}
\title{add_path_col}
\usage{
add_path_col(x, ...)

add_path_col(x, ...)

\S4method{add_path_col}{ready4_sp_data_pack_lup}(x, r_data_dir_chr)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{r_data_dir_chr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

Generic method for getting data referenced in a lookup table
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
