% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_add_path_col.R
\name{add_path_col.ready4_sp_data_pack_lup}
\alias{add_path_col.ready4_sp_data_pack_lup}
\title{add_path_col.ready4_sp_data_pack_lup}
\usage{
add_path_col.ready4_sp_data_pack_lup(x, r_data_dir_chr)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{r_data_dir_chr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
 \code{\link[purrr]{map}},\code{\link[purrr]{pluck}}
 \code{\link[stringr]{str_locate}},\code{\link[stringr]{str_sub}}
}
