% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_popl_counts}
\alias{add_popl_counts}
\title{Add population counts}
\usage{
add_popl_counts(
  profiled_sf,
  group_by_var_1L_chr,
  nse_objs_ls,
  convert_sfx_to_pfx_1L_lgl = FALSE,
  top_level_1L_lgl = FALSE
)
}
\arguments{
\item{profiled_sf}{Profiled (a simple features object)}

\item{group_by_var_1L_chr}{Group by variable (a character vector of length one)}

\item{nse_objs_ls}{Non-standard evaluation objects (a list)}

\item{convert_sfx_to_pfx_1L_lgl}{Convert suffix to prefix (a logical vector of length one), Default: FALSE}

\item{top_level_1L_lgl}{Top level (a logical vector of length one), Default: FALSE}
}
\value{
Profiled (a simple features object)
}
\description{
add_popl_counts() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add population counts. Function argument profiled_sf specifies the object to be updated. The function returns Profiled (a simple features object).
}
\keyword{internal}
