% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_ppr_ls_to_data_pack_lup}
\alias{add_ppr_ls_to_data_pack_lup}
\title{Add ppr list to data pack}
\usage{
add_ppr_ls_to_data_pack_lup(x, y)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}
}
\description{
add_ppr_ls_to_data_pack_lup() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add ppr list to data pack lookup table. Function argument x specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
