% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attr_to_data_pack.R
\name{add_ppr_to_data_pack_lup}
\alias{add_ppr_to_data_pack_lup}
\title{add_ppr_to_data_pack_lup}
\usage{
add_ppr_to_data_pack_lup(data_pack_lup, combined_ste_ppr_ls,
  object_name_stub, area_type, region)
}
\arguments{
\item{data_pack_lup}{PARAM_DESCRIPTION}

\item{combined_ste_ppr_ls}{PARAM_DESCRIPTION}

\item{object_name_stub}{PARAM_DESCRIPTION}

\item{area_type}{PARAM_DESCRIPTION}

\item{region}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
 \code{\link[stringr]{str_sub}}
 \code{\link[dplyr]{bind}}
}
