% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prevalence.R
\name{add_prevalence_col}
\alias{add_prevalence_col}
\title{add_prevalence_col}
\usage{
add_prevalence_col(sp_data_sf, col_name, param_tb_slimmed,
  par_name_var = "parameter_name", it_nbr)
}
\arguments{
\item{sp_data_sf}{PARAM_DESCRIPTION}

\item{col_name}{PARAM_DESCRIPTION}

\item{param_tb_slimmed}{PARAM_DESCRIPTION}

\item{par_name_var}{PARAM_DESCRIPTION, Default: 'parameter_name'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_sub}}
 \code{\link[dplyr]{pull}},\code{\link[dplyr]{mutate}}
 \code{\link[rlang]{sym}}
 \code{\link[ready.data]{data_get}}
}
