% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add_sp_resolution.R
\name{add_sp_resolution}
\alias{add_sp_resolution}
\title{add_sp_resolution}
\usage{
add_sp_resolution(lookup_tbs_r4, processed_dir)
}
\arguments{
\item{lookup_tbs_r4}{PARAM_DESCRIPTION}

\item{processed_dir}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{pull}},\code{\link[dplyr]{arrange}}
 \code{\link[stringr]{str_subset}}
 \code{\link[purrr]{map2}}
 \code{\link[tibble]{tibble}}
}
