% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_starter_sf_to_lups}
\alias{add_starter_sf_to_lups}
\title{Add starter simple features object to lups}
\usage{
add_starter_sf_to_lups(lookup_tbs_r4, path_to_starter_sf_chr)
}
\arguments{
\item{lookup_tbs_r4}{Lookup tibbles (a ready4 S4)}

\item{path_to_starter_sf_chr}{Path to starter simple features object (a character vector)}
}
\description{
add_starter_sf_to_lups() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add starter simple features object to lups. Function argument lookup_tbs_r4 specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
