% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_uid_lup}
\alias{add_uid_lup}
\title{Add unique identifier lookup table}
\usage{
add_uid_lup(x_VicinityLookup)
}
\arguments{
\item{x_VicinityLookup}{PARAM_DESCRIPTION}
}
\value{
X (Look up tables for spatiotemporal data)
}
\description{
add_uid_lup() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add unique identifier lookup table. Function argument x_VicinityLookup specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
