% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_by_yearly_age_sex.R
\name{adjust_pop_proj_for_pe}
\alias{adjust_pop_proj_for_pe}
\title{adjust_pop_proj_for_pe
ADD DESCRIPTIVE INFO HERE}
\usage{
adjust_pop_proj_for_pe(t0_pop, t1_pop, n_periods, age_sex_band, param_tb,
  it_nbr)
}
\arguments{
\item{t0_pop}{PARAM_DESCRIPTION}

\item{t1_pop}{PARAM_DESCRIPTION}

\item{n_periods}{PARAM_DESCRIPTION}

\item{age_sex_band}{PARAM_DESCRIPTION}

\item{param_tb}{PARAM_DESCRIPTION}

\item{it_nbr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
adjust_pop_proj_for_pe
ADD DESCRIPTIVE INFO HERE
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_sub}}
 \code{\link[ready.utils]{data_get}}
}
