% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suburb_based_effective_catchment.R
\name{apply_naming_convention}
\alias{apply_naming_convention}
\title{apply_naming_convention}
\usage{
apply_naming_convention(x, names_vector, state_territory)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{names_vector}{PARAM_DESCRIPTION}

\item{state_territory}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{case}}
}
