% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_area.R
\name{area}
\alias{area}
\alias{area-ready4_meso_area}
\title{area}
\usage{
area(x)

\S4method{area}{ready4_meso_area}(x)
}
\arguments{
\item{x}{An object of class ready4_meso_area}
}
\description{
S4 Generic function to get the value of the slot area

Get the value of the slot area for S4 objects of class ready4_meso_area
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
