% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_area_bound_year.R
\name{area_bound_year}
\alias{area_bound_year}
\alias{area_bound_year-ready4_meso_area}
\alias{area_bound_year,ready4_meso_area-method}
\title{area_bound_year}
\usage{
area_bound_year(x)

\S4method{area_bound_year}{ready4_meso_area}(x)
}
\arguments{
\item{x}{An object of class ready4_meso_area}
}
\description{
S4 Generic function to get the value of the slot area_bound_year

Get the value of the slot area_bound_year for S4 objects of class ready4_meso_area
}
