% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_area_bound_year.R,
%   R/s4_ready4_profiled_area.R
\name{area_bound_year}
\alias{area_bound_year}
\alias{area_bound_year-ready4_profiled_area}
\title{area_bound_year}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_profiled_area}
}
\description{
S4 Generic function to get the value of the slot area_bound_year

Get the value of the slot area_bound_year for S4 objects of class ready4_profiled_area
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
