% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_area_bound_year.R
\name{area_bound_year<-}
\alias{area_bound_year<-}
\alias{area_bound_year<--ready4_meso_area}
\alias{area_bound_year<-,ready4_meso_area-method}
\title{area_bound_year<-}
\usage{
area_bound_year(x) <- value

\S4method{area_bound_year}{ready4_meso_area}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_meso_area}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot area_bound_year

Set the value of the slot area_bound_year for S4 objects of class ready4_meso_area
}
