% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_area.R
\name{area<-}
\alias{area<-}
\alias{area<--ready4_meso_area}
\alias{area<-,ready4_meso_area-method}
\title{area<-}
\usage{
area(x) <- value

\S4method{area}{ready4_meso_area}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_meso_area}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot area

Set the value of the slot area for S4 objects of class ready4_meso_area
}
