% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_area_type.R
\name{area_type}
\alias{area_type}
\alias{area_type-ready4_meso_area}
\alias{area_type,ready4_meso_area-method}
\title{area_type}
\usage{
area_type(x)

\S4method{area_type}{ready4_meso_area}(x)
}
\arguments{
\item{x}{An object of class ready4_meso_area}
}
\description{
S4 Generic function to get the value of the slot area_type

Get the value of the slot area_type for S4 objects of class ready4_meso_area
}
