% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_area_type.R
\name{area_type<-}
\alias{area_type<-}
\alias{area_type<--ready4_profiled_area}
\title{area_type<-}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_profiled_area}
}
\description{
S4 Generic function to set the value of the slot area_type

Set the value of the slot area_type for S4 objects of class ready4_profiled_area
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
