% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_area_type.R
\name{area_type<-}
\alias{area_type<-}
\alias{area_type<--ready4_meso_area}
\alias{area_type<-,ready4_meso_area-method}
\title{area_type<-}
\usage{
area_type(x) <- value

\S4method{area_type}{ready4_meso_area}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_meso_area}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot area_type

Set the value of the slot area_type for S4 objects of class ready4_meso_area
}
