% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_author.R
\name{author.vicinity_processed}
\alias{author.vicinity_processed}
\alias{author,vicinity_processed-method}
\alias{author.vicinity_raw}
\alias{author,vicinity_raw-method}
\alias{author-VicinityLocalProcessed}
\alias{author,VicinityLocalProcessed-method}
\alias{author-VicinityLocalRaw}
\alias{author,VicinityLocalRaw-method}
\alias{author-VicinityArguments}
\alias{author,VicinityArguments-method}
\title{Author and save files}
\usage{
author.vicinity_processed(x, path_1L_chr, crs_dbl)

\S4method{author}{vicinity_processed}(x, path_1L_chr, crs_dbl)

author.vicinity_raw(
  x,
  crs_dbl = numeric(0),
  imports_ls = NULL,
  match_vals_xx = NULL,
  merge_itms_chr = character(0),
  overwrite_1L_lgl = F,
  path_1L_chr = character(0),
  what_1L_chr = "raw"
)

\S4method{author}{vicinity_raw}(
  x,
  crs_dbl = numeric(0),
  imports_ls = NULL,
  match_vals_xx = NULL,
  merge_itms_chr = character(0),
  overwrite_1L_lgl = F,
  path_1L_chr = character(0),
  what_1L_chr = "raw"
)

\S4method{author}{VicinityLocalProcessed}(x, crs_nbr_dbl, return_r4_1L_lgl = T)

\S4method{author}{VicinityLocalRaw}(x, processed_fls_dir_1L_chr_chr, crs_nbr_dbl)

\S4method{author}{VicinityArguments}(x, ...)
}
\arguments{
\item{x}{An object of class VicinityArguments}

\item{path_1L_chr}{Path (a character vector of length one), Default: character(0)}

\item{crs_dbl}{Coordinates reference system (a double vector), Default: numeric(0)}

\item{imports_ls}{Imports (a list), Default: NULL}

\item{match_vals_xx}{Match values (an output object of multiple potential types), Default: NULL}

\item{merge_itms_chr}{Merge items (a character vector), Default: character(0)}

\item{overwrite_1L_lgl}{Overwrite (a logical vector of length one), Default: F}

\item{what_1L_chr}{What (a character vector of length one), Default: 'raw'}

\item{crs_nbr_dbl}{Coordinates reference system number (a double vector)}

\item{return_r4_1L_lgl}{Return ready4 module (a logical vector of length one), Default: T}

\item{processed_fls_dir_1L_chr_chr}{Processed files directory length one character vector (a character vector)}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects.

Files written (a logical vector)

Return (an output object of multiple potential types)

Y (Look up tables for spatiotemporal data)

NULL (An object)
}
\description{
author.vicinity_processed() is an author method that authors and saves files to local or remote locations. This method is implemented for the ready4 submodule class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data). The function is called for its side effects and does not return a value.

author.vicinity_raw() is an author method that authors and saves files to local or remote locations. This method is implemented for the ready4 submodule class for tibble object lookup table of metadata about raw (un-processed) spatial data to import. The function returns Files written (a logical vector).

author method applied to VicinityLocalProcessed

author method applied to VicinityLocalRaw

author method applied to VicinityArguments
}
