% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_authorData.R
\name{authorData.vicinity_raw}
\alias{authorData.vicinity_raw}
\alias{authorData,vicinity_raw-method}
\alias{authorData-VicinityLocalRaw}
\alias{authorData,VicinityLocalRaw-method}
\title{Author and document datasets}
\usage{
authorData.vicinity_raw(
  x,
  path_1L_chr,
  data_match_value_xx,
  match_var_nm_1L_chr = "name_chr",
  sub_dirs_chr = NULL,
  overwrite_1L_lgl = F,
  what_1L_chr = "outer"
)

\S4method{authorData}{vicinity_raw}(
  x,
  path_1L_chr,
  data_match_value_xx,
  match_var_nm_1L_chr = "name_chr",
  sub_dirs_chr = NULL,
  overwrite_1L_lgl = F,
  what_1L_chr = "outer"
)

\S4method{authorData}{VicinityLocalRaw}(x, return_r4_1L_lgl = F)
}
\arguments{
\item{x}{An object of class VicinityLocalRaw}

\item{path_1L_chr}{Path (a character vector of length one)}

\item{data_match_value_xx}{Data match value (an output object of multiple potential types)}

\item{match_var_nm_1L_chr}{Match variable name (a character vector of length one), Default: 'name_chr'}

\item{sub_dirs_chr}{Sub directories (a character vector), Default: NULL}

\item{overwrite_1L_lgl}{Overwrite (a logical vector of length one), Default: F}

\item{what_1L_chr}{What (a character vector of length one), Default: 'outer'}

\item{return_r4_1L_lgl}{Return ready4 module (a logical vector of length one), Default: F}
}
\value{
Files written (a logical vector of length one)

Return (an output object of multiple potential types)
}
\description{
authorData.vicinity_raw() is an authorData method that authors and saves files necessary for creating and documenting datasets. This method is implemented for the ready4 submodule class for tibble object lookup table of metadata about raw (un-processed) spatial data to import. The function returns Files written (a logical vector of length one).

authorData method applied to VicinityLocalRaw
}
