% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_bind.R
\name{bind_isochrone_bands}
\alias{bind_isochrone_bands}
\title{Bind isochrone bands}
\usage{
bind_isochrone_bands(isochrone_bands_ls, index_1L_int, travel_mode_1L_chr)
}
\arguments{
\item{isochrone_bands_ls}{Isochrone bands (a list)}

\item{index_1L_int}{Index (an integer vector of length one)}

\item{travel_mode_1L_chr}{Travel mode (a character vector of length one)}
}
\value{
Isochrone bands (a simple features object)
}
\description{
bind_isochrone_bands() is a Bind function that binds two objects together to create a composite object. Specifically, this function implements an algorithm to bind isochrone bands. The function returns Isochrone bands (a simple features object).
}
\keyword{internal}
