% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_by_yearly_age_sex.R
\name{by_age_sex_for_a_year}
\alias{by_age_sex_for_a_year}
\title{by_age_sex_for_a_year
FUNCTION_DESCRIPTION}
\usage{
by_age_sex_for_a_year(profiled_sf, age_lower, age_upper, t0_date,
  age_bands_ref, intervals, included_age_bands_num_all, included_prefix)
}
\arguments{
\item{profiled_sf}{PARAM_DESCRIPTION}

\item{age_lower}{PARAM_DESCRIPTION}

\item{age_upper}{PARAM_DESCRIPTION}

\item{t0_date}{PARAM_DESCRIPTION}

\item{age_bands_ref}{PARAM_DESCRIPTION}

\item{intervals}{PARAM_DESCRIPTION}

\item{included_age_bands_num_all}{PARAM_DESCRIPTION}

\item{included_prefix}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
by_age_sex_for_a_year
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{reduce}},\code{\link[purrr]{prepend}}
}
