% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_end_date}
\alias{calculate_end_date}
\title{Calculate end date}
\usage{
calculate_end_date(
  model_start_ymdhms_dtm,
  nbr_steps_start_to_end_1L_int,
  simulation_steps_ymwd_dtm
)
}
\arguments{
\item{model_start_ymdhms_dtm}{Model start years months days hours minutes seconds (a date vector)}

\item{nbr_steps_start_to_end_1L_int}{Number steps start to end (an integer vector of length one)}

\item{simulation_steps_ymwd_dtm}{Simulation steps years month weeks days (a date vector)}
}
\value{
End date (a date vector)
}
\description{
calculate_end_date() is a Calculate function that performs a numeric calculation. Specifically, this function implements an algorithm to calculate end date. The function returns End date (a date vector).
}
\keyword{internal}
