% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_check.R
\name{check_if_ppr}
\alias{check_if_ppr}
\title{Check if ppr}
\usage{
check_if_ppr(data_name_item, data_lookup_tb, pop_projs_str)
}
\arguments{
\item{data_name_item}{PARAM_DESCRIPTION}

\item{data_lookup_tb}{Data lookup (a tibble)}

\item{pop_projs_str}{PARAM_DESCRIPTION}
}
\description{
check_if_ppr() is a Check function that performs a validity check. Specifically, this function implements an algorithm to check if ppr. The function is called for its side effects and does not return a value.
}
\keyword{internal}
