% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suburb_based_effective_catchment.R
\name{client_by_suburb_one_year}
\alias{client_by_suburb_one_year}
\title{client_by_suburb_one_year}
\usage{
client_by_suburb_one_year(cluster, year, file_ref = NA)
}
\arguments{
\item{cluster}{PARAM_DESCRIPTION}

\item{year}{PARAM_DESCRIPTION}

\item{file_ref}{PARAM_DESCRIPTION, Default: NA}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{mutate}}
 \code{\link[stringr]{case}}
}
