% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spatial_area_within_xmin_drive_of_points.R
\name{cluster_isochrones}
\alias{cluster_isochrones}
\title{cluster_isochrones}
\usage{
cluster_isochrones(cluster_tbs_list, look_up_ref, time_min = 0,
  time_max = 60, nbr_time_steps = 5)
}
\arguments{
\item{cluster_tbs_list}{PARAM_DESCRIPTION}

\item{look_up_ref}{PARAM_DESCRIPTION}

\item{time_min}{PARAM_DESCRIPTION, Default: 0}

\item{time_max}{PARAM_DESCRIPTION, Default: 60}

\item{nbr_time_steps}{PARAM_DESCRIPTION, Default: 5}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{pluck}},\code{\link[purrr]{map}},\code{\link[purrr]{accumulate}},\code{\link[purrr]{prepend}}
 \code{\link[dplyr]{select}},\code{\link[dplyr]{pull}}
 \code{\link[stats]{setNames}}
 \code{\link[sf]{geos_combine}},\code{\link[sf]{geos_binary_ops}}
}
