% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_by_yearly_age_sex.R
\name{col_pair_growth_rate}
\alias{col_pair_growth_rate}
\title{col_pair_growth_rate
ADD DESCRIPTIVE INFO HERE}
\usage{
col_pair_growth_rate(profiled_sf, list_element, included_cols_pairs,
  included_time_intervals_start_end, param_tb, it_nbr)
}
\arguments{
\item{profiled_sf}{PARAM_DESCRIPTION}

\item{list_element}{PARAM_DESCRIPTION}

\item{included_cols_pairs}{PARAM_DESCRIPTION}

\item{included_time_intervals_start_end}{PARAM_DESCRIPTION}

\item{param_tb}{PARAM_DESCRIPTION}

\item{it_nbr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
col_pair_growth_rate
ADD DESCRIPTIVE INFO HERE
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{reduce}},\code{\link[purrr]{prepend}},\code{\link[purrr]{pluck}}
 \code{\link[dplyr]{mutate}}
 \code{\link[rlang]{sym}}
 \code{\link[stringr]{str_sub}},\code{\link[stringr]{str_replace}},\code{\link[stringr]{case}}
}
