% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_travel_time_app_calc.R
\name{convert_tt_polygon_to_sf}
\alias{convert_tt_polygon_to_sf}
\title{convert_tt_polygon_to_sf}
\usage{
convert_tt_polygon_to_sf(tt_polyline, mode_of_transport, travel_time_hours,
  crs)
}
\arguments{
\item{tt_polyline}{PARAM_DESCRIPTION}

\item{mode_of_transport}{PARAM_DESCRIPTION}

\item{travel_time_hours}{PARAM_DESCRIPTION}

\item{crs}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[googlePolylines]{decode}}
 \code{\link[purrr]{pluck}}
 \code{\link[sf]{st}},\code{\link[sf]{sfc}},\code{\link[sf]{sf}}
}
