% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_country.R
\name{country}
\alias{country}
\alias{country-ready4_macro}
\alias{country,ready4_macro-method}
\title{country}
\usage{
country(x)

\S4method{country}{ready4_macro}(x)
}
\arguments{
\item{x}{An object of class ready4_macro}
}
\description{
S4 Generic function to get the value of the slot country

Get the value of the slot country for S4 objects of class ready4_macro
}
