% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_country.R
\name{country<-}
\alias{country<-}
\alias{country<--ready4_macro}
\title{country<-}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_macro}
}
\description{
S4 Generic function to set the value of the slot country

Set the value of the slot country for S4 objects of class ready4_macro
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
