% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_country.R
\name{country}
\alias{country}
\alias{country-ready4_profiled_area}
\title{country}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_profiled_area}
}
\description{
S4 Generic function to get the value of the slot country

Get the value of the slot country for S4 objects of class ready4_profiled_area
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
