% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_country_bound_year.R
\name{country_bound_year}
\alias{country_bound_year}
\alias{country_bound_year-ready4_macro}
\alias{country_bound_year,ready4_macro-method}
\title{country_bound_year}
\usage{
country_bound_year(x)

\S4method{country_bound_year}{ready4_macro}(x)
}
\arguments{
\item{x}{An object of class ready4_macro}
}
\description{
S4 Generic function to get the value of the slot country_bound_year

Get the value of the slot country_bound_year for S4 objects of class ready4_macro
}
