% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_country_bound_year.R
\name{country_bound_year<-}
\alias{country_bound_year<-}
\alias{country_bound_year<--ready4_macro}
\alias{country_bound_year<-,ready4_macro-method}
\title{country_bound_year<-}
\usage{
country_bound_year(x) <- value

\S4method{country_bound_year}{ready4_macro}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_macro}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot country_bound_year

Set the value of the slot country_bound_year for S4 objects of class ready4_macro
}
