% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_country.R
\name{country<-}
\alias{country<-}
\alias{country<--ready4_macro}
\alias{country<-,ready4_macro-method}
\title{country<-}
\usage{
country(x) <- value

\S4method{country}{ready4_macro}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_macro}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot country

Set the value of the slot country for S4 objects of class ready4_macro
}
