% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aus_land_boundary.R
\name{create_australia_land_boundary}
\alias{create_australia_land_boundary}
\title{create_australia_land_boundary}
\usage{
create_australia_land_boundary(aus_boundary_sf)
}
\arguments{
\item{aus_boundary_sf}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Create a boundary SF of Australia
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}}
 \code{\link[sf]{geos_combine}}
}
