% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_state_territory_boundary.R
\name{create_state_territory_boundary}
\alias{create_state_territory_boundary}
\title{FUNCTION_TITLE}
\usage{
create_state_territory_boundary(aus_boundary_sf, s_t)
}
\arguments{
\item{aus_boundary_sf}{PARAM_DESCRIPTION}

\item{s_t}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}}
}
