% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_crs_nbr.R
\name{crs_nbr<-}
\alias{crs_nbr<-}
\alias{crs_nbr<--ready4_profiled_area}
\title{crs_nbr<-}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_profiled_area}
}
\description{
S4 Generic function to set the value of the slot crs_nbr

Set the value of the slot crs_nbr for S4 objects of class ready4_profiled_area
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
