% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_profiled_area.R, R/gnrc_crs_nbr.R
\name{crs_nbr-ready4_profiled_area}
\alias{crs_nbr-ready4_profiled_area}
\alias{crs_nbr}
\title{crs_nbr}
\arguments{
\item{x}{An object of class ready4_profiled_area}

\item{x}{An object}
}
\description{
Get the value of the slot crs_nbr for S4 objects of class ready4_profiled_area

S4 Generic function to get the value of the slot crs_nbr
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
