% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_crs_nbr.R
\name{crs_nbr<-}
\alias{crs_nbr<-}
\alias{crs_nbr<--ready4_macro}
\alias{crs_nbr<-,ready4_macro-method}
\title{crs_nbr<-}
\usage{
crs_nbr(x) <- value

\S4method{crs_nbr}{ready4_macro}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_macro}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot crs_nbr

Set the value of the slot crs_nbr for S4 objects of class ready4_macro
}
