% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{data_import_get_file_name_from_path}
\alias{data_import_get_file_name_from_path}
\title{data_import_get_file_name_from_path}
\usage{
data_import_get_file_name_from_path(path_str, with_ext = TRUE)
}
\arguments{
\item{path_str}{PARAM_DESCRIPTION}

\item{with_ext}{PARAM_DESCRIPTION, Default: TRUE}

\item{path_str}{PARAM_DESCRIPTION}

\item{with_ext}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_sub}}
 \code{\link[stringi]{stri_locate_all}}

\code{\link[stringr]{str_sub}}
 \code{\link[stringi]{stri_locate_all}}
}
