% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_import_data.R
\name{data_import_get_one_path}
\alias{data_import_get_one_path}
\title{data_import_get_one_path}
\usage{
data_import_get_one_path(downloaded_data_tb, lookup_reference, data_directory)
}
\arguments{
\item{downloaded_data_tb}{PARAM_DESCRIPTION}

\item{lookup_reference}{PARAM_DESCRIPTION}

\item{data_directory}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
 \code{\link[dplyr]{select}}
 \code{\link[ready4utils]{data_get}}
}
