% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{data_import_items}
\alias{data_import_items}
\title{data_import_items}
\usage{
data_import_items(included_items_names, item_data_type, data_directory, x)
}
\arguments{
\item{included_items_names}{PARAM_DESCRIPTION}

\item{item_data_type}{PARAM_DESCRIPTION}

\item{data_directory}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}}
 \code{\link[purrr]{map}}
 \code{\link[sf]{st_read}}
 \code{\link[stats]{setNames}}
}
