% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_import_data.R
\name{write_dirs_for_imp}
\alias{write_dirs_for_imp}
\title{write_dirs_for_imp}
\usage{
write_dirs_for_imp(directory_paths)
}
\arguments{
\item{directory_paths}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
}
