% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_import_data.R
\name{get_non_shape_items_for_imp}
\alias{get_non_shape_items_for_imp}
\title{get_non_shape_items_for_imp}
\usage{
get_non_shape_items_for_imp(path_str, x)
}
\arguments{
\item{path_str}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_sub}}
 \code{\link[stringi]{stri_locate_all}}
 \code{\link[ready4fun]{get_from_lup}}
 \code{\link[purrr]{map}}
}
