% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_import_data.R
\name{write_to_rnm_fls_for_imp}
\alias{write_to_rnm_fls_for_imp}
\title{write_to_rnm_fls_for_imp}
\usage{
write_to_rnm_fls_for_imp(
  x,
  data_lookup_ref,
  lookup_variable,
  directory_path,
  overwrite_lgl = F
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{data_lookup_ref}{PARAM_DESCRIPTION}

\item{lookup_variable}{PARAM_DESCRIPTION}

\item{directory_path}{PARAM_DESCRIPTION}

\item{overwrite_lgl}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready4fun]{get_from_lup}}
 \code{\link[purrr]{map2}}
}
