% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_import_data.R
\name{data_import_save_files}
\alias{data_import_save_files}
\title{data_import_save_files}
\usage{
data_import_save_files(x, data_lookup_ref, lookup_variable, directory_path,
  overwrite_lgl = F)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{data_lookup_ref}{PARAM_DESCRIPTION}

\item{lookup_variable}{PARAM_DESCRIPTION}

\item{directory_path}{PARAM_DESCRIPTION}

\item{overwrite_lgl}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
 \code{\link[ready4utils]{data_get}}
 \code{\link[utils]{download.file}},\code{\link[utils]{unzip}}
}
