% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{data_import_selected_downloads}
\alias{data_import_selected_downloads}
\title{data_import_selected_downloads}
\usage{
data_import_selected_downloads(required_data, destination_directory, x)
}
\arguments{
\item{required_data}{PARAM_DESCRIPTION}

\item{destination_directory}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
}
