% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{data_import_show_menu_names}
\alias{data_import_show_menu_names}
\title{data_import_show_menu_names}
\usage{
data_import_show_menu_names(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}},\code{\link[dplyr]{pull}}
}
