% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_import_data.R
\name{data_import_show_menu_of_type_detail}
\alias{data_import_show_menu_of_type_detail}
\title{data_import_show_menu_of_type_detail}
\usage{
data_import_show_menu_of_type_detail(x, lookup_ref)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{lookup_ref}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}
}
