% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_data_year.R, R/s4_ready4_profiled_area.R
\name{data_year}
\alias{data_year}
\alias{data_year-ready4_profiled_area}
\title{data_year}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_profiled_area}
}
\description{
S4 Generic function to get the value of the slot data_year

Get the value of the slot data_year for S4 objects of class ready4_profiled_area
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
