% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_data_year.R
\name{data_year<-}
\alias{data_year<-}
\alias{data_year<--ready4_profiled_area}
\alias{data_year<-,ready4_profiled_area-method}
\title{data_year<-}
\usage{
data_year(x) <- value

\S4method{data_year}{ready4_profiled_area}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_profiled_area}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot data_year

Set the value of the slot data_year for S4 objects of class ready4_profiled_area
}
