% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_profiled_area.R, R/gnrc_data_ymds.R
\name{data_ymds<--ready4_profiled_area}
\alias{data_ymds<--ready4_profiled_area}
\alias{data_ymds<-}
\title{data_ymds<-}
\arguments{
\item{x}{An object of class ready4_profiled_area}

\item{x}{An object}
}
\description{
Set the value of the slot data_ymds for S4 objects of class ready4_profiled_area

S4 Generic function to set the value of the slot data_ymds
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
