% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_data_ymds.R
\name{data_ymds<-}
\alias{data_ymds<-}
\alias{data_ymds<--ready4_profiled_area}
\alias{data_ymds<-,ready4_profiled_area-method}
\title{data_ymds<-}
\usage{
data_ymds(x) <- value

\S4method{data_ymds}{ready4_profiled_area}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_profiled_area}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot data_ymds

Set the value of the slot data_ymds for S4 objects of class ready4_profiled_area
}
