% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_by_yearly_age_sex.R
\name{demographic_compound_growth_rate}
\alias{demographic_compound_growth_rate}
\title{demographic_compound_growth_rate
FUNCTION_DESCRIPTION}
\usage{
demographic_compound_growth_rate(t0_pop, t1_pop, n_periods, age_sex_band,
  param_tb, it_nbr)
}
\arguments{
\item{t0_pop}{PARAM_DESCRIPTION}

\item{t1_pop}{PARAM_DESCRIPTION}

\item{n_periods}{PARAM_DESCRIPTION}

\item{age_sex_band}{PARAM_DESCRIPTION}

\item{param_tb}{PARAM_DESCRIPTION}

\item{it_nbr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
demographic_compound_growth_rate
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
