% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_deterministic.2016.2021.R
\name{demographic_deterministic_2016_2021}
\alias{demographic_deterministic_2016_2021}
\title{Deterministic five year population growth projections}
\usage{
demographic_deterministic_2016_2021(demographic.data,
  pop.growth.by.age.lga)
}
\arguments{
\item{demographic.data}{A .....}

\item{pop.growth.by.age.lga}{A .....}
}
\value{
A list.
}
\description{
This function creates a list......
}
\details{

}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{join}},\code{\link[dplyr]{mutate}}

Other demographic modelling functions.: \code{\link{demographic_growth_calcs}},
  \code{\link{demographic_growth_list_single}},
  \code{\link{demographic_growth_list}},
  \code{\link{demographic_growth_projs_slice}},
  \code{\link{demographic_inputs}},
  \code{\link{demographic_projections}},
  \code{\link{demographic_range_of_projections}}
}
\concept{demographic modelling functions.}
