% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_growth_calcs.R
\name{demographic_growth_calcs}
\alias{demographic_growth_calcs}
\title{Demographic growth calculations}
\usage{
demographic_growth_calcs(age.sex.ref, reference.year, base.year, fup.year,
  poppreds.list, demographic.data)
}
\arguments{
\item{age.sex.ref}{A .....}

\item{reference.year}{A .....}

\item{base.year}{A .....}

\item{fup.year}{A .....}

\item{poppreds.list}{A list .....}

\item{demographic.data}{A .....}
}
\value{
A tibble.
}
\description{
This function creates a tibble......
}
\details{

}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}},\code{\link[dplyr]{pull}},\code{\link[dplyr]{summarise_all}},\code{\link[dplyr]{funs}}
 \code{\link[sf]{st_geometry}}
 \code{\link[purrr]{pluck}}

Other demographic modelling functions.: \code{\link{demographic_deterministic_2016_2021}},
  \code{\link{demographic_growth_list_single}},
  \code{\link{demographic_growth_list}},
  \code{\link{demographic_growth_projs_slice}},
  \code{\link{demographic_inputs}},
  \code{\link{demographic_projections}},
  \code{\link{demographic_range_of_projections}}
}
\concept{demographic modelling functions.}
