% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_growth_list.R
\name{demographic_growth_list}
\alias{demographic_growth_list}
\title{Demographic growth}
\usage{
demographic_growth_list(demographic.projection.errors, demographic.data)
}
\arguments{
\item{demographic.projection.errors}{A .....}

\item{demographic.data}{A .....}
}
\value{
A list.
}
\description{
This function creates a list......
}
\details{

}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{st_geometry}}
 \code{\link[dplyr]{select}},\code{\link[dplyr]{reexports}}
 \code{\link[stringr]{str_replace}},\code{\link[stringr]{str_detect}}
 \code{\link[purrr]{map}},\code{\link[purrr]{keep}},\code{\link[purrr]{flatten}},\code{\link[purrr]{map2}}
 \code{\link[stats]{setNames}}

Other demographic modelling functions.: \code{\link{demographic_deterministic_2016_2021}},
  \code{\link{demographic_growth_calcs}},
  \code{\link{demographic_growth_list_single}},
  \code{\link{demographic_growth_projs_slice}},
  \code{\link{demographic_inputs}},
  \code{\link{demographic_projections}},
  \code{\link{demographic_range_of_projections}}
}
\concept{demographic modelling functions.}
