% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_growth_list_single.R
\name{demographic_growth_list_single}
\alias{demographic_growth_list_single}
\title{Demographic predictions list}
\usage{
demographic_growth_list_single(reference.year, base.year, fup.year,
  poppreds.list, demographic.data)
}
\arguments{
\item{reference.year}{A .....}

\item{base.year}{A .....}

\item{fup.year}{A .....}

\item{poppreds.list}{A list .....}

\item{demographic.data}{A .....}
}
\value{
A list.
}
\description{
This function creates a list......
}
\details{

}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
 \code{\link[stringr]{str_subset}}
 \code{\link[stats]{setNames}}

Other demographic modelling functions.: \code{\link{demographic_deterministic_2016_2021}},
  \code{\link{demographic_growth_calcs}},
  \code{\link{demographic_growth_list}},
  \code{\link{demographic_growth_projs_slice}},
  \code{\link{demographic_inputs}},
  \code{\link{demographic_projections}},
  \code{\link{demographic_range_of_projections}}
}
\concept{demographic modelling functions.}
