% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_inputs.R
\name{demographic_inputs}
\alias{demographic_inputs}
\title{Demographic modelling inputs......}
\usage{
demographic_inputs(error.data, demographic.data, error.listref,
  demographic.data.namestub, demographic.data.col.ref)
}
\arguments{
\item{error.data}{A .....}

\item{demographic.data}{A .....}

\item{error.listref}{A .....}

\item{demographic.data.namestub}{A .....}

\item{demographic.data.col.ref}{A .....}
}
\value{
A tibble.
}
\description{
This function creates a tibble......
}
\details{

}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{pluck}}
 \code{\link[rlang]{sym}}
 \code{\link[dplyr]{select}}

Other demographic modelling functions.: \code{\link{demographic_deterministic_2016_2021}},
  \code{\link{demographic_growth_calcs}},
  \code{\link{demographic_growth_list_single}},
  \code{\link{demographic_growth_list}},
  \code{\link{demographic_growth_projs_slice}},
  \code{\link{demographic_projections}},
  \code{\link{demographic_range_of_projections}}
}
\concept{demographic modelling functions.}
