% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_projections.R
\name{demographic_projections}
\alias{demographic_projections}
\title{Forecast population groth with prediction intervals.}
\usage{
demographic_projections(naive, n.replications)
}
\arguments{
\item{naive}{A logical value, denoting whether using a naive modelling approach (assumes
previous growth rates continue into future) or a theory based model.}

\item{n.replications}{A positive integer.}
}
\value{
A list.
}
\description{
This function creates a list......
}
\details{
Function implements calculations based on:
http://www.jstor.org.ezp.lib.unimelb.edu.au/stable/41110880?seq=1#page_scan_tab_contents
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
 \code{\link[dplyr]{summarise}},\code{\link[dplyr]{slice}},\code{\link[dplyr]{bind}},\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{pull}}
 \code{\link[purrr]{map2}}
 \code{\link[stats]{setNames}}

Other demographic modelling functions.: \code{\link{demographic_deterministic_2016_2021}},
  \code{\link{demographic_growth_calcs}},
  \code{\link{demographic_growth_list_single}},
  \code{\link{demographic_growth_list}},
  \code{\link{demographic_growth_projs_slice}},
  \code{\link{demographic_inputs}},
  \code{\link{demographic_range_of_projections}}
}
\concept{demographic modelling functions.}
