% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_distribution.R
\name{distribution<-}
\alias{distribution<-}
\title{distribution<-}
\arguments{
\item{x}{An object}
}
\description{
S4 Generic function to set the value of the slot distribution
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
