% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_download_data.R
\name{download_data.ready4_sp_import_lup}
\alias{download_data.ready4_sp_import_lup}
\alias{download_data,ready4_sp_import_lup-method}
\title{Download data method applied toeadyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import..}
\usage{
download_data.ready4_sp_import_lup(
  x,
  destination_directory,
  data_lookup_ref,
  lookup_variable = "name",
  directory_sub_divs = NULL,
  overwrite_lgl = F
)

\S4method{download_data}{ready4_sp_import_lup}(
  x,
  destination_directory,
  data_lookup_ref,
  lookup_variable = "name",
  directory_sub_divs = NULL,
  overwrite_lgl = F
)
}
\arguments{
\item{x}{An instance of Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.}

\item{destination_directory}{PARAM_DESCRIPTION}

\item{data_lookup_ref}{PARAM_DESCRIPTION}

\item{lookup_variable}{PARAM_DESCRIPTION, Default: 'name'}

\item{directory_sub_divs}{PARAM_DESCRIPTION, Default: NULL}

\item{overwrite_lgl}{Overwrite (a logical vector), Default: F}
}
\description{
download_data.ready4_sp_import_lup() is a Download Data method that downloads data files. This method is implemented for the Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.. The function is called for its side effects and does not return a value.
}
