% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_raw.R
\name{download_data.ready4_sp_import_lup}
\alias{download_data.ready4_sp_import_lup}
\title{download_data.ready4_sp_import_lup}
\usage{
\method{download_data}{ready4_sp_import_lup}(x, destination_directory,
  data_lookup_ref, lookup_variable = "name", directory_sub_divs = NULL,
  overwrite_lgl = F)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{destination_directory}{PARAM_DESCRIPTION}

\item{data_lookup_ref}{PARAM_DESCRIPTION}

\item{lookup_variable}{PARAM_DESCRIPTION, Default: 'name'}

\item{directory_sub_divs}{PARAM_DESCRIPTION, Default: NULL}

\item{overwrite_lgl}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready4s3]{ready4_sp_import_lup}}
}
