% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_profiled_area.R,
%   R/gnrc_drive_time_limit_mins.R
\name{drive_time_limit_mins-ready4_profiled_area}
\alias{drive_time_limit_mins-ready4_profiled_area}
\alias{drive_time_limit_mins}
\title{drive_time_limit_mins}
\arguments{
\item{x}{An object of class ready4_profiled_area}

\item{x}{An object}
}
\description{
Get the value of the slot drive_time_limit_mins for S4 objects of class ready4_profiled_area

S4 Generic function to get the value of the slot drive_time_limit_mins
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
