% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_drive_time_limit_mins.R
\name{drive_time_limit_mins<-}
\alias{drive_time_limit_mins<-}
\alias{drive_time_limit_mins<--ready4_profiled_area}
\alias{drive_time_limit_mins<-,ready4_profiled_area-method}
\title{drive_time_limit_mins<-}
\usage{
drive_time_limit_mins(x) <- value

\S4method{drive_time_limit_mins}{ready4_profiled_area}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_profiled_area}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot drive_time_limit_mins

Set the value of the slot drive_time_limit_mins for S4 objects of class ready4_profiled_area
}
