% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_env_sf.R
\name{env_sf}
\alias{env_sf}
\alias{env_sf-ready4_env}
\alias{env_sf,ready4_env-method}
\title{env_sf}
\usage{
env_sf(x)

\S4method{env_sf}{ready4_env}(x)
}
\arguments{
\item{x}{An object of class ready4_env}
}
\description{
S4 Generic function to get the value of the slot env_sf

Get the value of the slot env_sf for S4 objects of class ready4_env
}
