% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_env_sf.R
\name{env_sf<-}
\alias{env_sf<-}
\alias{env_sf<--ready4_env}
\title{env_sf<-}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_env}
}
\description{
S4 Generic function to set the value of the slot env_sf

Set the value of the slot env_sf for S4 objects of class ready4_env
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
