% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_env.R, R/gnrc_env_sf.R
\name{env_sf-ready4_env}
\alias{env_sf-ready4_env}
\alias{env_sf}
\title{env_sf}
\arguments{
\item{x}{An object of class ready4_env}

\item{x}{An object}
}
\description{
Get the value of the slot env_sf for S4 objects of class ready4_env

S4 Generic function to get the value of the slot env_sf
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
