% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_env_sf.R
\name{env_sf<-}
\alias{env_sf<-}
\alias{env_sf<--ready4_env}
\alias{env_sf<-,ready4_env-method}
\title{env_sf<-}
\usage{
env_sf(x) <- value

\S4method{env_sf}{ready4_env}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_env}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot env_sf

Set the value of the slot env_sf for S4 objects of class ready4_env
}
