% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prevalence.R
\name{estimate_prevalence}
\alias{estimate_prevalence}
\title{estimate_prevalence
Function to estimate the prevalence of a specified condition, for a specified disorder for a specified area / year.}
\usage{
estimate_prevalence(sp_data_sf, sp_unit_id = "pop_sp_unit_id", param_tb,
  par_name_var = "parameter_name", it_nbr)
}
\arguments{
\item{sp_data_sf}{PARAM_DESCRIPTION}

\item{prev_rates_vec}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
estimate_prevalence
Function to estimate the prevalence of a specified condition, for a specified disorder for a specified area / year.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{summarise_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{funs}}
 \code{\link[purrr]{map2}},\code{\link[purrr]{map}},\code{\link[purrr]{reduce}}
 \code{\link[stringr]{str_sub}}
 \code{\link[stats]{setNames}}
 \code{\link[rlang]{sym}}
 \code{\link[tibble]{tibble}}
}
