% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_pack.R
\name{export_attr_tb}
\alias{export_attr_tb}
\title{export_attr_tb}
\usage{
export_attr_tb(attr_tb, obj_name, processed_dir, overwrite_lgl)
}
\arguments{
\item{attr_tb}{PARAM_DESCRIPTION}

\item{obj_name}{PARAM_DESCRIPTION}

\item{processed_dir}{PARAM_DESCRIPTION}

\item{overwrite_lgl}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
