% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_data_pack.R
\name{export_starter_sf}
\alias{export_starter_sf}
\title{export_starter_sf}
\usage{
export_starter_sf(lookup_tbs_r4, path_to_starter_sf_chr)
}
\arguments{
\item{lookup_tbs_r4}{PARAM_DESCRIPTION}

\item{path_to_starter_sf_chr}{A character string.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{reduce}}
 \code{\link[sf]{st_geometry_type}},\code{\link[sf]{geos_measures}}
 \code{\link[dplyr]{mutate}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{pull}}
 \code{\link[units]{set_units}}
 \code{\link[tibble]{add_row}}
}
