% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sp_data_list.R
\name{extend_sp_data_list}
\alias{extend_sp_data_list}
\title{extend_sp_data_list}
\usage{
extend_sp_data_list(sp_data_list, input_ls, profiled_area_bands_list)
}
\arguments{
\item{sp_data_list}{PARAM_DESCRIPTION}

\item{input_ls}{PARAM_DESCRIPTION}

\item{profiled_area_bands_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready4s4]{crs_nbr}},\code{\link[ready4s4]{geom_dist_limit_km}},\code{\link[ready4s4]{drive_time_limmit_mins}},\code{\link[ready4s4]{lookup_tb}},\code{\link[ready4s4]{sp_uid_lup}},\code{\link[ready4s4]{data_year}},\code{\link[ready4s4]{sp_resolution_lup}}
 \code{\link[ready4utils]{data_get}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}}
 \code{\link[purrr]{map}},\code{\link[purrr]{map2}}
 \code{\link[nnet]{which.is.max}}
 \code{\link[sf]{geos_measures}}
}
