% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_features.R, R/s4_ready4_profiled_area.R
\name{features<-}
\alias{features<-}
\alias{features<--ready4_profiled_area}
\title{features<-}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_profiled_area}
}
\description{
S4 Generic function to set the value of the slot features

Set the value of the slot features for S4 objects of class ready4_profiled_area
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
