% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_features.R
\name{features}
\alias{features}
\alias{features-ready4_profiled_area}
\title{features}
\usage{
features(x)

\S4method{features}{ready4_profiled_area}(x)
}
\arguments{
\item{x}{An object of class ready4_profiled_area}
}
\description{
S4 Generic function to get the value of the slot features

Get the value of the slot features for S4 objects of class ready4_profiled_area
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
