% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_features.R
\name{features<-}
\alias{features<-}
\alias{features<--ready4_profiled_area}
\alias{features<-,ready4_profiled_area-method}
\title{features<-}
\usage{
features(x) <- value

\S4method{features}{ready4_profiled_area}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_profiled_area}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot features

Set the value of the slot features for S4 objects of class ready4_profiled_area
}
