% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spatial_area_within_xmin_drive_of_points.R
\name{fix_names_up_to_sfs}
\alias{fix_names_up_to_sfs}
\title{fix_names_up_to_sfs}
\usage{
fix_names_up_to_sfs(look_up_ref, one_cluster_up_to_xmin_list)
}
\arguments{
\item{look_up_ref}{PARAM_DESCRIPTION}

\item{one_cluster_up_to_xmin_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{pluck}}
 \code{\link[stringr]{str_subset}},\code{\link[stringr]{str_replace}}
 \code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}}
 \code{\link[rlang]{sym}}
}
