% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_by_yearly_age_sex.R
\name{future_pop_from_comp_g_r}
\alias{future_pop_from_comp_g_r}
\title{future_pop_from_comp_g_r
FUNCTION_DESCRIPTION}
\usage{
future_pop_from_comp_g_r(cgr, base_year_pop, nyears)
}
\arguments{
\item{cgr}{Numeric, cumulative annual growth rate.}

\item{base_year_pop}{Numeric, population count in base year.}

\item{nyears}{Number of years between baseline and follow-up.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
future_pop_from_comp_g_r
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
