% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_by_yearly_age_sex.R
\name{gen_age_sex_estimates_t0}
\alias{gen_age_sex_estimates_t0}
\title{gen_age_sex_estimates_t0
FUNCTION_DESCRIPTION}
\usage{
gen_age_sex_estimates_t0(profiled_sf, t0_date, included_age_bands_num,
  included_age_bands_num_all, age0, age1, intervals)
}
\arguments{
\item{profiled_sf}{PARAM_DESCRIPTION}

\item{t0_date}{PARAM_DESCRIPTION}

\item{included_age_bands_num}{PARAM_DESCRIPTION}

\item{included_age_bands_num_all}{PARAM_DESCRIPTION}

\item{age0}{PARAM_DESCRIPTION}

\item{age1}{PARAM_DESCRIPTION}

\item{intervals}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
gen_age_sex_estimates_t0
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{flatten}},\code{\link[purrr]{map}}
}
