% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_by_yearly_age_sex.R
\name{gen_age_sex_estimates_tx}
\alias{gen_age_sex_estimates_tx}
\title{gen_age_sex_estimates_tx
FUNCTION_DESCRIPTION}
\usage{
gen_age_sex_estimates_tx(profiled_sf, ymwd_step)
}
\arguments{
\item{profiled_sf}{PARAM_DESCRIPTION}

\item{ymwd_step}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
gen_age_sex_estimates_tx
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{funs}},\code{\link[dplyr]{select_all}},\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{pull}}
 \code{\link[stringr]{str_sub}},\code{\link[stringr]{str_replace}}
 \code{\link[purrr]{pluck}},\code{\link[purrr]{map}},\code{\link[purrr]{reduce}},\code{\link[purrr]{prepend}}
 \code{\link[lubridate]{ymd}},\code{\link[lubridate]{period}},\code{\link[lubridate]{interval}}
}
