% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_by_yearly_age_sex.R
\name{gen_demog_acgr}
\alias{gen_demog_acgr}
\title{gen_demog_acgr
FUNCTION_DESCRIPTION}
\usage{
gen_demog_acgr(profiled_sf, fn_pars, param_tb, it_nbr)
}
\arguments{
\item{profiled_sf}{PARAM_DESCRIPTION}

\item{fn_pars}{PARAM_DESCRIPTION}

\item{param_tb}{PARAM_DESCRIPTION}

\item{it_nbr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
gen_demog_acgr
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{reduce}},\code{\link[purrr]{prepend}}
 \code{\link[dplyr]{select_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{funs}},\code{\link[dplyr]{select}}
 \code{\link[stringr]{str_sub}},\code{\link[stringr]{case}}
}
