% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_by_yearly_age_sex.R
\name{gen_demog_features}
\alias{gen_demog_features}
\title{Calculate age/sex populations for each year / age of interest.}
\usage{
gen_demog_features(profiled_sf, years, age_lower, age_upper, param_tb,
  it_nbr, acgr = TRUE, age_by_year = FALSE, drop_bands = TRUE,
  intervals = 5, month = "07", day = "01", included_prefix = "")
}
\arguments{
\item{profiled_sf}{PARAM_DESCRIPTION}

\item{years}{Numeric vector of the years for which age / sex counts should be estimated.}

\item{age_lower}{Numeric, start of included age range.}

\item{age_upper}{Numeric, end of included age range.}

\item{param_tb}{PARAM_DESCRIPTION}

\item{it_nbr}{PARAM_DESCRIPTION}

\item{acgr}{PARAM_DESCRIPTION, Default: TRUE}

\item{age_by_year}{PARAM_DESCRIPTION, Default: FALSE}

\item{drop_bands}{PARAM_DESCRIPTION, Default: TRUE}

\item{intervals}{Numeric, the years included in each age group in the source sf, Default: 5}

\item{month}{PARAM_DESCRIPTION, Default: '07'}

\item{day}{PARAM_DESCRIPTION, Default: '01'}

\item{included_prefix}{PARAM_DESCRIPTION, Default: ''}

\item{profiled_sf}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}},\code{\link[dplyr]{reexports}}
}
