% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_by_yearly_age_sex.R
\name{gen_demog_fun_par_vals}
\alias{gen_demog_fun_par_vals}
\title{gen_demog_fun_par_vals
FUNCTION_DESCRIPTION}
\usage{
gen_demog_fun_par_vals(profiled_sf, years, age_lower, age_upper, intervals,
  acgr, age_by_year)
}
\arguments{
\item{profiled_sf}{PARAM_DESCRIPTION}

\item{years}{PARAM_DESCRIPTION}

\item{age_lower}{PARAM_DESCRIPTION}

\item{age_upper}{PARAM_DESCRIPTION}

\item{intervals}{PARAM_DESCRIPTION}

\item{acgr}{PARAM_DESCRIPTION}

\item{age_by_year}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
gen_demog_fun_par_vals
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}},\code{\link[purrr]{flatten}},\code{\link[purrr]{modify}}
 \code{\link[stringr]{str_subset}},\code{\link[stringr]{str_sub}}
}
