% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_gen_par_vals.R
\name{gen_par_vals}
\alias{gen_par_vals}
\alias{gen_par_vals.ready4_params_struc_mape}
\alias{gen_par_vals.ready4_par_str_envir}
\title{gen_par_vals}
\usage{
gen_par_vals(x, n_its_int, jt_dist = FALSE, ...)

\method{gen_par_vals}{ready4_params_struc_mape}(x, n_its_int, jt_dist)

\method{gen_par_vals}{ready4_par_str_envir}(x, n_its_int)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{n_its_int}{PARAM_DESCRIPTION}

\item{jt_dist}{PARAM_DESCRIPTION, Default: FALSE}

\item{...}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}

\item{n_its_int}{PARAM_DESCRIPTION}

\item{jt_dist}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}

\item{n_its_int}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
A generic function for generating new parameter value sets for each simulation iteration.
}
\details{
DETAILS

DETAILS

DETAILS
}
\section{Methods (by class)}{
\itemize{
\item \code{ready4_params_struc_mape}: Method for ready4_params_struc_mape class.

\item \code{ready4_par_str_envir}: Method for ready4_par_str_envir class.
}}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{prepend}},\code{\link[purrr]{reduce}}
 \code{\link[dplyr]{join}}

\code{\link[purrr]{prepend}},\code{\link[purrr]{reduce}}
 \code{\link[dplyr]{join}},\code{\link[dplyr]{select}}
}
