% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_gen.R
\name{gen_val_envir}
\alias{gen_val_envir}
\title{Gen value envir}
\usage{
gen_val_envir(params_struc_tb, it_nbr)
}
\arguments{
\item{params_struc_tb}{Params struc (a tibble)}

\item{it_nbr}{PARAM_DESCRIPTION}
}
\description{
gen_val_envir() is a Gen function that generates values for a specified varaible or variables. Specifically, this function implements an algorithm to gen value envir. The function is called for its side effects and does not return a value.
}
\keyword{internal}
