% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_gen.R
\name{gen_val_mape}
\alias{gen_val_mape}
\title{Gen value mape}
\usage{
gen_val_mape(params_struc_tb, it_nbr, jt_dist)
}
\arguments{
\item{params_struc_tb}{Params struc (a tibble)}

\item{it_nbr}{PARAM_DESCRIPTION}

\item{jt_dist}{PARAM_DESCRIPTION}
}
\description{
gen_val_mape() is a Gen function that generates values for a specified varaible or variables. Specifically, this function implements an algorithm to gen value mape. The function is called for its side effects and does not return a value.
}
\keyword{internal}
