% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_geom_dist_km_cuts.R
\name{geom_dist_km_cuts<-}
\alias{geom_dist_km_cuts<-}
\alias{geom_dist_km_cuts<--ready4_micro}
\title{geom_dist_km_cuts<-}
\usage{
geom_dist_km_cuts(x) <- value

\S4method{geom_dist_km_cuts}{ready4_micro}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_micro}
}
\description{
S4 Generic function to set the value of the slot geom_dist_km_cuts

Set the value of the slot geom_dist_km_cuts for S4 objects of class ready4_micro
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
