% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_geom_dist_km_cuts.R
\name{geom_dist_km_cuts}
\alias{geom_dist_km_cuts}
\alias{geom_dist_km_cuts-ready4_micro}
\title{geom_dist_km_cuts}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_micro}
}
\description{
S4 Generic function to get the value of the slot geom_dist_km_cuts

Get the value of the slot geom_dist_km_cuts for S4 objects of class ready4_micro
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
