% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_area_within_xkm_of_cluster.R
\name{geom_distance_circles_to_bands}
\alias{geom_distance_circles_to_bands}
\title{geom_distance_circles_to_bands}
\usage{
geom_distance_circles_to_bands(geom_distance_circle_sfs_list)
}
\arguments{
\item{geom_distance_circle_sfs_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}},\code{\link[purrr]{pluck}},\code{\link[purrr]{prepend}}
 \code{\link[sf]{geos_binary_ops}}
 \code{\link[dplyr]{select}}
 \code{\link[stats]{setNames}}
}
