% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_agent_areas}
\alias{get_agent_areas}
\title{Get agent areas}
\usage{
get_agent_areas(
  agent_areas_tb = NULL,
  area_var_nm_1L_chr = "Suburb",
  path_1L_chr = character(0),
  type_1L_chr = "table"
)
}
\arguments{
\item{agent_areas_tb}{Agent areas (a tibble), Default: NULL}

\item{area_var_nm_1L_chr}{Area variable name (a character vector of length one), Default: 'Suburb'}

\item{path_1L_chr}{Path (a character vector of length one), Default: character(0)}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'table'}
}
\value{
Agent areas (an output object of multiple potential types)
}
\description{
get_agent_areas() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get agent areas. Function argument agent_areas_tb specifies the where to look for the required object. The function returns Agent areas (an output object of multiple potential types).
}
\keyword{internal}
