% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_area_sqkm_sf}
\alias{get_area_sqkm_sf}
\title{Get area sqkm}
\usage{
get_area_sqkm_sf(data_sf)
}
\arguments{
\item{data_sf}{Data (a simple features object)}
}
\description{
get_area_sqkm_sf() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get area sqkm simple features object. Function argument data_sf specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
