% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suburb_based_effective_catchment.R
\name{get_client_suburb_vector}
\alias{get_client_suburb_vector}
\title{get_client_suburb_vector}
\usage{
get_client_suburb_vector(clients_by_suburb)
}
\arguments{
\item{clients_by_suburb}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}},\code{\link[dplyr]{pull}}
 \code{\link[stringr]{str_order}}
}
