% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_closest_year}
\alias{get_closest_year}
\title{Get closest year}
\usage{
get_closest_year(
  data_lookup_tb,
  incl_main_ft_vec,
  target_year,
  target_area = NULL,
  find_closest = "abs"
)
}
\arguments{
\item{data_lookup_tb}{Data lookup (a tibble)}

\item{incl_main_ft_vec}{PARAM_DESCRIPTION}

\item{target_year}{PARAM_DESCRIPTION}

\item{target_area}{PARAM_DESCRIPTION, Default: NULL}

\item{find_closest}{PARAM_DESCRIPTION, Default: 'abs'}
}
\value{
NA ()
}
\description{
get_closest_year() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get closest year. Function argument data_lookup_tb specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
