% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attr_to_sf.R
\name{get_common_vars_sf_ls}
\alias{get_common_vars_sf_ls}
\title{get_common_vars_sf_ls}
\usage{
get_common_vars_sf_ls(sf_ls)
}
\arguments{
\item{sf_ls}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
}
