% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_common_yrs_sf_ls}
\alias{get_common_yrs_sf_ls}
\title{Get common yrs}
\usage{
get_common_yrs_sf_ls(sf_ls)
}
\arguments{
\item{sf_ls}{Simple features object list (a list of simple features objects)}
}
\description{
get_common_yrs_sf_ls() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get common yrs simple features object list. Function argument sf_ls specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
