% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_get_data.R
\name{get_data.ready4_sp_abbreviations_lup}
\alias{get_data.ready4_sp_abbreviations_lup}
\alias{get_data,ready4_sp_abbreviations_lup-method}
\alias{get_data.ready4_sp_data_pack_lup}
\alias{get_data,ready4_sp_data_pack_lup-method}
\alias{get_data.ready4_sp_uid_lup}
\alias{get_data,ready4_sp_uid_lup-method}
\title{Get data method applied toeadyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations..}
\usage{
get_data.ready4_sp_abbreviations_lup(x, col_chr = "short_name", value_chr)

\S4method{get_data}{ready4_sp_abbreviations_lup}(x, col_chr = "short_name", value_chr)

get_data.ready4_sp_data_pack_lup(
  x,
  col_chr = "main_feature",
  value_chr,
  r_data_dir_chr = NA_character_
)

\S4method{get_data}{ready4_sp_data_pack_lup}(
  x,
  col_chr = "main_feature",
  value_chr,
  r_data_dir_chr = NA_character_
)

get_data.ready4_sp_uid_lup(
  x,
  col_chr = "spatial_unit",
  value_chr,
  area_bound_yr
)

\S4method{get_data}{ready4_sp_uid_lup}(x, col_chr = "spatial_unit", value_chr, area_bound_yr)
}
\arguments{
\item{x}{An instance of Readyforwhatsnext S3 class for tibble object lookup table of unique feature identifiers used for different spatial objects.}

\item{col_chr}{Column (a character vector), Default: 'spatial_unit'}

\item{value_chr}{Value (a character vector)}

\item{r_data_dir_chr}{R data directory (a character vector), Default: 'NA'}

\item{area_bound_yr}{PARAM_DESCRIPTION}
}
\description{
get_data.ready4_sp_abbreviations_lup() is a Get Data method that retrieves data from R objects loaded in memory. This method is implemented for the Readyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations.. The function is called for its side effects and does not return a value.

get_data.ready4_sp_data_pack_lup() is a Get Data method that retrieves data from R objects loaded in memory. This method is implemented for the Readyforwhatsnext S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).. The function is called for its side effects and does not return a value.

get_data.ready4_sp_uid_lup() is a Get Data method that retrieves data from R objects loaded in memory. This method is implemented for the Readyforwhatsnext S3 class for tibble object lookup table of unique feature identifiers used for different spatial objects.. The function is called for its side effects and does not return a value.
}
