% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_get_data.R
\name{get_data}
\alias{get_data}
\alias{get_data,ready4_sp_data_pack_lup-method}
\alias{get_data,ready4_sp_abbreviations_lup-method}
\alias{get_data,ready4_sp_uid_lup-method}
\title{Get data}
\usage{
\S4method{get_data}{ready4_sp_data_pack_lup}(
  x,
  col_chr = "main_feature",
  value_chr,
  r_data_dir_chr = NA_character_
)

\S4method{get_data}{ready4_sp_abbreviations_lup}(x, col_chr = "short_name", value_chr)

\S4method{get_data}{ready4_sp_uid_lup}(x, col_chr = "spatial_unit", value_chr, area_bound_yr)
}
\arguments{
\item{x}{Primary input object - see Usage section for allowable signatures}

\item{col_chr}{PARAM_DESCRIPTION, Default: 'spatial_unit'}

\item{value_chr}{PARAM_DESCRIPTION}

\item{r_data_dir_chr}{PARAM_DESCRIPTION, Default: 'NA'}

\item{area_bound_yr}{PARAM_DESCRIPTION}
}
\description{
Get data referenced in a lookup table.
}
