% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_get_data.R
\name{get_data.ready4_sp_data_pack_lup}
\alias{get_data.ready4_sp_data_pack_lup}
\title{get_data.ready4_sp_data_pack_lup}
\usage{
get_data.ready4_sp_data_pack_lup(x, col_chr = "main_feature", value_chr)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{col_chr}{PARAM_DESCRIPTION, Default: 'main_feature'}

\item{value_chr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
