% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_get_data.R
\name{get_data.ready4_sp_uid_lup}
\alias{get_data.ready4_sp_uid_lup}
\title{get_data.ready4_sp_uid_lup}
\usage{
get_data.ready4_sp_uid_lup(
  x,
  col_chr = "spatial_unit",
  value_chr,
  area_bound_yr
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{col_chr}{PARAM_DESCRIPTION, Default: 'spatial_unit'}

\item{value_chr}{PARAM_DESCRIPTION}

\item{area_bound_yr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready4utils]{data_get}}
 \code{\link[dplyr]{filter}}
}
