% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make_sp_data_list.R
\name{get_data_year_chr}
\alias{get_data_year_chr}
\title{get_data_year_chr}
\usage{
get_data_year_chr(data_ymdhms)
}
\arguments{
\item{data_ymdhms}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[lubridate]{year}}
}
