% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_dir_paths_for_data_imp}
\alias{get_dir_paths_for_data_imp}
\title{Get directory paths for data import}
\usage{
get_dir_paths_for_data_imp(
  x,
  destination_directory,
  data_lookup_ref,
  lookup_variable,
  directory_sub_divs
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{destination_directory}{PARAM_DESCRIPTION}

\item{data_lookup_ref}{PARAM_DESCRIPTION}

\item{lookup_variable}{PARAM_DESCRIPTION}

\item{directory_sub_divs}{PARAM_DESCRIPTION}
}
\description{
get_dir_paths_for_data_imp() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get directory paths for data import. Function argument x specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
