% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_group_by_var}
\alias{get_group_by_var}
\title{Get group by var}
\usage{
get_group_by_var(
  profile_unit,
  data_unit,
  group_at_profile_unit = TRUE,
  group_by_lookup_tb,
  area_bound_year
)
}
\arguments{
\item{profile_unit}{PARAM_DESCRIPTION}

\item{data_unit}{PARAM_DESCRIPTION}

\item{group_at_profile_unit}{PARAM_DESCRIPTION, Default: TRUE}

\item{group_by_lookup_tb}{Group by lookup (a tibble)}

\item{area_bound_year}{PARAM_DESCRIPTION}
}
\value{
NA ()
}
\description{
get_group_by_var() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get group by var. Function argument profile_unit specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
