% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_highest_res}
\alias{get_highest_res}
\title{Get highest res}
\usage{
get_highest_res(options_vec, year, resolution_lup_r3)
}
\arguments{
\item{options_vec}{PARAM_DESCRIPTION}

\item{year}{PARAM_DESCRIPTION}

\item{resolution_lup_r3}{Resolution (a ready4 S3 extension of lookup table)}
}
\description{
get_highest_res() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get highest res. Function argument options_vec specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
