% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_import_chr_vec}
\alias{get_import_chr_vec}
\title{Get import character vector vec}
\usage{
get_import_chr_vec(lookup_tbs_r4, data_type_chr)
}
\arguments{
\item{lookup_tbs_r4}{Lookup tibbles (a ready4 S4)}

\item{data_type_chr}{Data type (a character vector)}
}
\description{
get_import_chr_vec() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get import character vector vec. Function argument lookup_tbs_r4 specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
