% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_included_yrs_sf}
\alias{get_included_yrs_sf}
\title{Get included yrs}
\usage{
get_included_yrs_sf(sf)
}
\arguments{
\item{sf}{Simple features object (a simple features object)}
}
\description{
get_included_yrs_sf() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get included yrs simple features object. Function argument sf specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
