% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_max_or_min_yr_of_sf}
\alias{get_max_or_min_yr_of_sf}
\title{Get max or min yr of}
\usage{
get_max_or_min_yr_of_sf(sf, max = T)
}
\arguments{
\item{sf}{Simple features object (a simple features object)}

\item{...}{Additional arguments, Default: T}
}
\description{
get_max_or_min_yr_of_sf() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get max or min yr of simple features object. Function argument sf specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
