% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_menu_of_type_detail_for_imp}
\alias{get_menu_of_type_detail_for_imp}
\title{Get menu of type detail for import}
\usage{
get_menu_of_type_detail_for_imp(x, lookup_ref)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{lookup_ref}{PARAM_DESCRIPTION}
}
\description{
get_menu_of_type_detail_for_imp() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get menu of type detail for import. Function argument x specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
