% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_merge_sf_str}
\alias{get_merge_sf_str}
\title{Get merge simple features object setter}
\usage{
get_merge_sf_str(lookup_r4, sp_import_r3_slice, processed_dir = NULL)
}
\arguments{
\item{lookup_r4}{Lookup (a ready4 S4)}

\item{sp_import_r3_slice}{PARAM_DESCRIPTION}

\item{processed_dir}{PARAM_DESCRIPTION, Default: NULL}
}
\description{
get_merge_sf_str() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get merge simple features object setter. Function argument lookup_r4 specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
