% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_model_end_ymdhs}
\alias{get_model_end_ymdhs}
\title{Get model end ymdhs}
\usage{
get_model_end_ymdhs(input_ls)
}
\arguments{
\item{input_ls}{Input (a list)}
}
\description{
get_model_end_ymdhs() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get model end ymdhs. Function argument input_ls specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
