% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_name_from_path_chr}
\alias{get_name_from_path_chr}
\title{Get name from path character vector}
\usage{
get_name_from_path_chr(path_1L_chr, with_ext_1L_lgl = TRUE)
}
\arguments{
\item{path_1L_chr}{Path (a character vector of length one)}

\item{with_ext_1L_lgl}{With extension (a logical vector of length one), Default: TRUE}
}
\value{
Name (a character vector of length one)
}
\description{
get_name_from_path_chr() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get name from path character vector. The function returns Name (a character vector of length one).
}
\keyword{internal}
