% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_non_shape_items_for_imp}
\alias{get_non_shape_items_for_imp}
\title{Get non shape items for import}
\usage{
get_non_shape_items_for_imp(path_str, x)
}
\arguments{
\item{path_str}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}
}
\description{
get_non_shape_items_for_imp() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get non shape items for import. Function argument path_str specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
