% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update_counts_by_area.R
\name{get_popl_var_prefix}
\alias{get_popl_var_prefix}
\title{get_popl_var_prefix}
\usage{
get_popl_var_prefix(age_sex_pop_resolution, tot_pop_resolution = NULL,
  data_year)
}
\arguments{
\item{age_sex_pop_resolution}{PARAM_DESCRIPTION}

\item{tot_pop_resolution}{PARAM_DESCRIPTION, Default: NULL}

\item{data_year}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
