% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_r_import_path_chr}
\alias{get_r_import_path_chr}
\title{Get r import path}
\usage{
get_r_import_path_chr(r_data_dir_chr, name_chr, data_type_chr)
}
\arguments{
\item{r_data_dir_chr}{R data directory (a character vector)}

\item{name_chr}{Name (a character vector)}

\item{data_type_chr}{Data type (a character vector)}
}
\description{
get_r_import_path_chr() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get r import path character vector. Function argument r_data_dir_chr specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
