% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_res_specific_vars}
\alias{get_res_specific_vars}
\title{Get res specific vars}
\usage{
get_res_specific_vars(var_names, data_type, data_year, popl_var_prefix)
}
\arguments{
\item{var_names}{PARAM_DESCRIPTION}

\item{data_type}{PARAM_DESCRIPTION}

\item{data_year}{PARAM_DESCRIPTION}

\item{popl_var_prefix}{PARAM_DESCRIPTION}
}
\value{
NA ()
}
\description{
get_res_specific_vars() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get res specific vars. Function argument var_names specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
