% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sp_data.R
\name{get_resolution_hierarchy}
\alias{get_resolution_hierarchy}
\title{get_resolution_hierarchy}
\usage{
get_resolution_hierarchy(data_year, resolution_lup_r3, whole_area = TRUE)
}
\arguments{
\item{data_year}{PARAM_DESCRIPTION}

\item{resolution_lup_r3}{PARAM_DESCRIPTION}

\item{whole_area}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{desc}},\code{\link[dplyr]{pull}}
}
