% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_set_diff_lng_lat_sf}
\alias{get_set_diff_lng_lat_sf}
\title{Get set difference longitude lat simple features object}
\usage{
get_set_diff_lng_lat_sf(
  geometry_sf,
  cut_sf,
  crs_nbr_dbl,
  min_polygon_area_dbl = units::set_units(0.05, km^2),
  validate_1L_lgl = T
)
}
\arguments{
\item{geometry_sf}{Geometry (a simple features object)}

\item{cut_sf}{Cut (a simple features object)}

\item{crs_nbr_dbl}{Coordinates reference system number (a double vector)}

\item{min_polygon_area_dbl}{Minimum polygon area (a double vector), Default: units::set_units(0.05, km^2)}

\item{validate_1L_lgl}{Validate (a logical vector of length one), Default: T}
}
\value{
Difference (a simple features object)
}
\description{
get_set_diff_lng_lat_sf() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get set difference longitude lat simple features object. The function returns Difference (a simple features object).
}
\keyword{internal}
