% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_set_diff_lon_lat_sf}
\alias{get_set_diff_lon_lat_sf}
\title{Get set diff lon lat}
\usage{
get_set_diff_lon_lat_sf(
  profile_sf,
  cut_sf,
  crs_nbr_vec,
  validate_lgl = T,
  min_poly_area_dbl = units::set_units(0.05, km^2)
)
}
\arguments{
\item{profile_sf}{Profile (a simple features object)}

\item{cut_sf}{Cut (a simple features object)}

\item{crs_nbr_vec}{PARAM_DESCRIPTION}

\item{validate_lgl}{Validate (a logical vector), Default: T}

\item{min_poly_area_dbl}{Min poly area (a double vector), Default: units::set_units(0.05, km^2)}
}
\description{
get_set_diff_lon_lat_sf() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get set diff lon lat simple features object. Function argument profile_sf specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
