% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_sngl_path_for_imp}
\alias{get_sngl_path_for_imp}
\title{Get sngl path for import}
\usage{
get_sngl_path_for_imp(downloaded_data_tb, lookup_reference, data_directory)
}
\arguments{
\item{downloaded_data_tb}{Downloaded data (a tibble)}

\item{lookup_reference}{PARAM_DESCRIPTION}

\item{data_directory}{PARAM_DESCRIPTION}
}
\description{
get_sngl_path_for_imp() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get sngl path for import. Function argument downloaded_data_tb specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
