% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_spatial_data_names}
\alias{get_spatial_data_names}
\title{Get spatial data names}
\usage{
get_spatial_data_names(
  input_ls,
  sub_div_unit = NULL,
  require_year_match = TRUE,
  excl_diff_bound_yr = TRUE
)
}
\arguments{
\item{input_ls}{Input (a list)}

\item{sub_div_unit}{PARAM_DESCRIPTION, Default: NULL}

\item{require_year_match}{PARAM_DESCRIPTION, Default: TRUE}

\item{excl_diff_bound_yr}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
get_spatial_data_names() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get spatial data names. Function argument input_ls specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
