% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_starter_sf_for_profiled_area}
\alias{get_starter_sf_for_profiled_area}
\title{Get starter simple features object for profiled area}
\usage{
get_starter_sf_for_profiled_area(pa_r4, group_by_var)
}
\arguments{
\item{pa_r4}{Pa (a ready4 S4)}

\item{group_by_var}{PARAM_DESCRIPTION}
}
\value{
Starter (a simple features object)
}
\description{
get_starter_sf_for_profiled_area() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get starter simple features object for profiled area. Function argument pa_r4 specifies the where to look for the required object. The function returns Starter (a simple features object).
}
\keyword{internal}
