% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sp_data.R
\name{get_sys_data_tbs_ls}
\alias{get_sys_data_tbs_ls}
\title{get_sys_data_tbs_ls}
\usage{
get_sys_data_tbs_ls()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
