% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_global_region.R
\name{global_region<-}
\alias{global_region<-}
\alias{global_region<--ready4_macro}
\title{global_region<-}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_macro}
}
\description{
S4 Generic function to set the value of the slot global_region

Set the value of the slot global_region for S4 objects of class ready4_macro
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
