% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_global_region.R
\name{global_region<-}
\alias{global_region<-}
\alias{global_region<--ready4_macro}
\alias{global_region<-,ready4_macro-method}
\title{global_region<-}
\usage{
global_region(x) <- value

\S4method{global_region}{ready4_macro}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_macro}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot global_region

Set the value of the slot global_region for S4 objects of class ready4_macro
}
