% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_pack.R
\name{import_boundary_ls}
\alias{import_boundary_ls}
\title{import_boundary_ls}
\usage{
import_boundary_ls(lookup_tbs_r4, raw_data_dir)
}
\arguments{
\item{lookup_tbs_r4}{PARAM_DESCRIPTION}

\item{raw_data_dir}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready4s4]{sp_import_lup}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{pull}}
 \code{\link[stats]{setNames}}
}
