% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_import_data.R
\name{import_data.ready4_sp_import_lup}
\alias{import_data.ready4_sp_import_lup}
\alias{import_data,ready4_sp_import_lup-method}
\title{Import data method applied toeadyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import..}
\usage{
import_data.ready4_sp_import_lup(
  x,
  included_items_names,
  item_data_type,
  data_directory,
  r_data_dir_chr,
  save_lgl = T
)

\S4method{import_data}{ready4_sp_import_lup}(
  x,
  included_items_names,
  item_data_type,
  data_directory,
  r_data_dir_chr,
  save_lgl = T
)
}
\arguments{
\item{x}{An instance of Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.}

\item{included_items_names}{PARAM_DESCRIPTION}

\item{item_data_type}{PARAM_DESCRIPTION}

\item{data_directory}{PARAM_DESCRIPTION}

\item{r_data_dir_chr}{R data directory (a character vector)}

\item{save_lgl}{Save (a logical vector), Default: T}
}
\value{
NA ()
}
\description{
import_data.ready4_sp_import_lup() is an Import Data method that imports data from saved files and loads them into memory as R objects. This method is implemented for the Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.. The function is called for its side effects and does not return a value.
}
