% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data.ready4_sp_import_lup}
\alias{import_data.ready4_sp_import_lup}
\title{import_data.ready4_sp_import_lup}
\usage{
\method{import_data}{ready4_sp_import_lup}(x, included_items_names,
  item_data_type, data_directory, r_data_dir_chr, save_lgl = T)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{included_items_names}{PARAM_DESCRIPTION}

\item{item_data_type}{PARAM_DESCRIPTION}

\item{data_directory}{PARAM_DESCRIPTION}

\item{r_data_dir_chr}{PARAM_DESCRIPTION}

\item{save_lgl}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}}
 \code{\link[purrr]{map}}
 \code{\link[sf]{st_read}}
 \code{\link[stats]{setNames}}
}
