% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_pack.R
\name{import_ls}
\alias{import_ls}
\title{import_ls}
\usage{
import_ls(lookup_tbs_r4, raw_data_dir, r_data_dir_chr,
  data_type_chr = "Geometry", overwrite_lgl = F)
}
\arguments{
\item{lookup_tbs_r4}{PARAM_DESCRIPTION}

\item{raw_data_dir}{PARAM_DESCRIPTION}

\item{r_data_dir_chr}{PARAM_DESCRIPTION}

\item{data_type_chr}{PARAM_DESCRIPTION, Default: 'Geometry'}

\item{overwrite_lgl}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready4s4]{sp_import_lup}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{pull}}
 \code{\link[stats]{setNames}}
}
