% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_sp_local_proc.R
\name{import_this_ls<--ready4_sp_local_proc}
\alias{import_this_ls<--ready4_sp_local_proc}
\title{import_this_ls<-}
\arguments{
\item{x}{An object of class ready4_sp_local_proc}
}
\description{
Set the value of the slot import_this_ls for S4 objects of class ready4_sp_local_proc
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
