% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_ingest.R
\name{ingest.vicinity_processed}
\alias{ingest.vicinity_processed}
\alias{ingest,vicinity_processed-method}
\alias{ingest.vicinity_raw}
\alias{ingest,vicinity_raw-method}
\alias{ingest-VicinityProfile}
\alias{ingest,VicinityProfile-method}
\title{Ingest data}
\usage{
ingest.vicinity_processed(
  x,
  col_nm_1L_chr = "main_feature_chr",
  match_value_xx,
  processed_fls_dir_1L_chr = NA_character_
)

\S4method{ingest}{vicinity_processed}(
  x,
  col_nm_1L_chr = "main_feature_chr",
  match_value_xx,
  processed_fls_dir_1L_chr = NA_character_
)

ingest.vicinity_raw(
  x,
  imports_chr = character(0),
  data_type_1L_chr = character(0),
  path_1L_chr = character(0),
  raw_fls_dir_1L_chr = character(0),
  processed_fls_dir_1L_chr = character(0),
  what_1L_chr = "list",
  write_1L_lgl = T
)

\S4method{ingest}{vicinity_raw}(
  x,
  imports_chr = character(0),
  data_type_1L_chr = character(0),
  path_1L_chr = character(0),
  raw_fls_dir_1L_chr = character(0),
  processed_fls_dir_1L_chr = character(0),
  what_1L_chr = "list",
  write_1L_lgl = T
)

\S4method{ingest}{VicinityProfile}(x, key_var_1L_chr = character(0), what_1L_chr = "starter")
}
\arguments{
\item{x}{An object of class VicinityProfile}

\item{col_nm_1L_chr}{Column name (a character vector of length one), Default: 'main_feature_chr'}

\item{match_value_xx}{Match value (an output object of multiple potential types)}

\item{processed_fls_dir_1L_chr}{Processed files directory (a character vector of length one), Default: character(0)}

\item{imports_chr}{Imports (a character vector), Default: character(0)}

\item{data_type_1L_chr}{Data type (a character vector of length one), Default: character(0)}

\item{path_1L_chr}{Path (a character vector of length one), Default: character(0)}

\item{raw_fls_dir_1L_chr}{Raw files directory (a character vector of length one), Default: character(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: 'starter'}

\item{write_1L_lgl}{Write (a logical vector of length one), Default: T}

\item{key_var_1L_chr}{Key variable (a character vector of length one), Default: character(0)}
}
\value{
Object (an output object of multiple potential types)

Ingest (an output object of multiple potential types)

Object (an output object of multiple potential types)
}
\description{
ingest.vicinity_processed() is an ingest method that ingests data saved in external files into R objects stored in working memory. This method is implemented for the ready4 S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data). The function returns Object (an output object of multiple potential types).

ingest.vicinity_raw() is an ingest method that ingests data saved in external files into R objects stored in working memory. This method is implemented for the ready4 S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import. The function returns Ingest (an output object of multiple potential types).

ingest method applied to VicinityProfile
}
