% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_intersect.R
\name{intersect_lon_lat_sfs}
\alias{intersect_lon_lat_sfs}
\title{Intersect lon lat sfs}
\usage{
intersect_lon_lat_sfs(sf_1, sf_2, crs_nbr_vec, validate_lgl = T)
}
\arguments{
\item{sf_1}{PARAM_DESCRIPTION}

\item{sf_2}{PARAM_DESCRIPTION}

\item{crs_nbr_vec}{PARAM_DESCRIPTION}

\item{validate_lgl}{Validate (a logical vector), Default: T}
}
\description{
intersect_lon_lat_sfs() is an Intersect function that gets the intersection between two or more data objects. Specifically, this function implements an algorithm to intersect lon lat sfs. The function is called for its side effects and does not return a value.
}
\keyword{internal}
