% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_sfs_keep_counts.R
\name{intersect_sf_drop_cols}
\alias{intersect_sf_drop_cols}
\title{intersect_sf_drop_cols}
\usage{
intersect_sf_drop_cols(main_sf, adjunct_sf)
}
\arguments{
\item{main_sf}{PARAM_DESCRIPTION}

\item{adjunct_sf}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_subset}}
 \code{\link[sf]{geos_binary_ops}},\code{\link[sf]{st_geometry_type}}
 \code{\link[dplyr]{select}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{filter}}
}
