% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_sfs_keep_counts.R
\name{intersect_sfs_keep_counts}
\alias{intersect_sfs_keep_counts}
\alias{intersect_sf_drop_cols}
\title{intersect_sfs_keep_counts
This function:
  -}
\usage{
intersect_sfs_keep_counts(profiled_sf, profiled_colref = NA,
  profiled_rowref = NA, attribute_sf, attribute_unit, data_type,
  data_year)

intersect_sf_drop_cols(main_sf, adjunct_sf)
}
\arguments{
\item{profiled_sf}{A SF object corresponding to the geographic unit that is to be profiled.}

\item{profiled_colref}{A variable name from the profiled_sf object that is the basis for
filtering the object if not all areas from the object are to be used.}

\item{profiled_rowref}{A value from the profiled_colref variable which, if specified, will
be used to identify a subset of the profiled_sf object.}

\item{attribute_sf}{A SF object corresponding to the geographic unit that is resolution
at which the profiled_sf object will be profiled.}
}
\value{
A simple features object.
}
\description{
intersect_sfs_keep_counts
This function:
  -
}
\details{

}
\section{Functions}{
\itemize{
\item \code{intersect_sf_drop_cols}: Function to intesect two sfs and drop extra columns.
}}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}},\code{\link[dplyr]{pull}},\code{\link[dplyr]{slice}}
 \code{\link[stringr]{str_subset}}
 \code{\link[sf]{geos_binary_ops}}
}
