% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_intersect.R
\name{intersect_sfs_update_counts}
\alias{intersect_sfs_update_counts}
\title{Intersect sfs update counts}
\usage{
intersect_sfs_update_counts(
  profiled_sf,
  profiled_colref = NA,
  profiled_rowref = NA,
  sp_data_list,
  tot_pop_resolution,
  age_sex_pop_resolution,
  group_by_var,
  age_sex_counts_grouped_by,
  data_year,
  crs_nbr_vec
)
}
\arguments{
\item{profiled_sf}{Profiled (a simple features object)}

\item{profiled_colref}{PARAM_DESCRIPTION, Default: NA}

\item{profiled_rowref}{PARAM_DESCRIPTION, Default: NA}

\item{sp_data_list}{PARAM_DESCRIPTION}

\item{tot_pop_resolution}{PARAM_DESCRIPTION}

\item{age_sex_pop_resolution}{PARAM_DESCRIPTION}

\item{group_by_var}{PARAM_DESCRIPTION}

\item{age_sex_counts_grouped_by}{PARAM_DESCRIPTION}

\item{data_year}{PARAM_DESCRIPTION}

\item{crs_nbr_vec}{PARAM_DESCRIPTION}
}
\value{
Profiled (a simple features object)
}
\description{
intersect_sfs_update_counts() is an Intersect function that gets the intersection between two or more data objects. Specifically, this function implements an algorithm to intersect sfs update counts. The function returns Profiled (a simple features object).
}
\keyword{internal}
