% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_par_str_envir.R
\name{is_ready4_par_str_envir}
\alias{is_ready4_par_str_envir}
\title{Is Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to the spatial environment.}
\usage{
is_ready4_par_str_envir(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to the spatial environment.
}
\description{
Check whether an object is a valid instance of the Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to the spatial environment.
}
\details{
Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to the spatial environment.
}
