% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_par_val_envir.R
\name{is_ready4_param_val_envir}
\alias{is_ready4_param_val_envir}
\title{Is Readyforwhatsnext S3 class for tibble object that stores simulation parameter values for each iteration.}
\usage{
is_ready4_param_val_envir(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the Readyforwhatsnext S3 class for tibble object that stores simulation parameter values for each iteration.
}
\description{
Check whether an object is a valid instance of the Readyforwhatsnext S3 class for tibble object that stores simulation parameter values for each iteration.
}
\details{
Readyforwhatsnext S3 class for tibble object that stores simulation parameter values for each iteration.
}
