% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_starter_sf_lup.R
\name{is_ready4_sp_starter_sf_lup}
\alias{is_ready4_sp_starter_sf_lup}
\title{Is Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.}
\usage{
is_ready4_sp_starter_sf_lup(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
\description{
Check whether an object is a valid instance of the Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
