% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_uid_lup.R
\name{is_ready4_sp_uid_lup}
\alias{is_ready4_sp_uid_lup}
\title{Is Readyforwhatsnext S3 class for tibble object lookup table of unique feature identifiers used for different spatial objects.}
\usage{
is_ready4_sp_uid_lup(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the Readyforwhatsnext S3 class for tibble object lookup table of unique feature identifiers used for different spatial objects.
}
\description{
Check whether an object is a valid instance of the Readyforwhatsnext S3 class for tibble object lookup table of unique feature identifiers used for different spatial objects.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of unique feature identifiers used for different spatial objects.
}
