% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_intersect_sfs_keep_counts.R
\name{join_lon_lat_sfs}
\alias{join_lon_lat_sfs}
\title{join_lon_lat_sfs}
\usage{
join_lon_lat_sfs(polys_sf, points_sf, crs_nbr_vec, validate_lgl = T)
}
\arguments{
\item{polys_sf}{PARAM_DESCRIPTION}

\item{points_sf}{PARAM_DESCRIPTION}

\item{crs_nbr_vec}{PARAM_DESCRIPTION}

\item{validate_lgl}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{st_join}},\code{\link[sf]{st_transform}}
}
