% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_profiled_area.R, R/gnrc_lookup_tb.R
\name{lookup_tb-ready4_profiled_area}
\alias{lookup_tb-ready4_profiled_area}
\alias{lookup_tb}
\title{lookup_tb}
\arguments{
\item{x}{An object of class ready4_profiled_area}

\item{x}{An object}
}
\description{
Get the value of the slot lookup_tb for S4 objects of class ready4_profiled_area

S4 Generic function to get the value of the slot lookup_tb
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
