% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_lookup_tb.R
\name{lookup_tb<-}
\alias{lookup_tb<-}
\alias{lookup_tb<--ready4_macro}
\alias{lookup_tb<-,ready4_macro-method}
\title{lookup_tb<-}
\usage{
lookup_tb(x) <- value

\S4method{lookup_tb}{ready4_macro}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_macro}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot lookup_tb

Set the value of the slot lookup_tb for S4 objects of class ready4_macro
}
