% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_lup_tbs_r4.R
\name{lup_tbs_r4}
\alias{lup_tbs_r4}
\alias{lup_tbs_r4-ready4_sp_local}
\alias{lup_tbs_r4,ready4_sp_local-method}
\alias{lup_tbs_r4-ready4_sp_local_raw}
\alias{lup_tbs_r4,ready4_sp_local_raw-method}
\alias{lup_tbs_r4-ready4_sp_local_proc}
\alias{lup_tbs_r4,ready4_sp_local_proc-method}
\title{lup_tbs_r4}
\usage{
lup_tbs_r4(x)

\S4method{lup_tbs_r4}{ready4_sp_local}(x)

\S4method{lup_tbs_r4}{ready4_sp_local_raw}(x)

\S4method{lup_tbs_r4}{ready4_sp_local_proc}(x)
}
\arguments{
\item{x}{An object of class ready4_sp_local_proc}
}
\description{
S4 Generic function to get the value of the slot lup_tbs_r4

Get the value of the slot lup_tbs_r4 for S4 objects of class ready4_sp_local

Get the value of the slot lup_tbs_r4 for S4 objects of class ready4_sp_local_raw

Get the value of the slot lup_tbs_r4 for S4 objects of class ready4_sp_local_proc
}
