% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_lup_tbs_r4.R, R/s4_ready4_sp_local.R,
%   R/s4_ready4_sp_local_proc.R, R/s4_ready4_sp_local_raw.R
\name{lup_tbs_r4<-}
\alias{lup_tbs_r4<-}
\alias{lup_tbs_r4<--ready4_sp_local}
\alias{lup_tbs_r4<--ready4_sp_local_proc}
\alias{lup_tbs_r4<--ready4_sp_local_raw}
\title{lup_tbs_r4<-}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_sp_local}

\item{x}{An object of class ready4_sp_local_proc}

\item{x}{An object of class ready4_sp_local_raw}
}
\description{
S4 Generic function to set the value of the slot lup_tbs_r4

Set the value of the slot lup_tbs_r4 for S4 objects of class ready4_sp_local

Set the value of the slot lup_tbs_r4 for S4 objects of class ready4_sp_local_proc

Set the value of the slot lup_tbs_r4 for S4 objects of class ready4_sp_local_raw
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
