% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_lup_tbs_r4.R
\name{lup_tbs_r4<-}
\alias{lup_tbs_r4<-}
\alias{lup_tbs_r4<--ready4_sp_local}
\alias{lup_tbs_r4<-,ready4_sp_local-method}
\alias{lup_tbs_r4<--ready4_sp_local_raw}
\alias{lup_tbs_r4<-,ready4_sp_local_raw-method}
\alias{lup_tbs_r4<--ready4_sp_local_proc}
\alias{lup_tbs_r4<-,ready4_sp_local_proc-method}
\title{lup_tbs_r4<-}
\usage{
lup_tbs_r4(x) <- value

\S4method{lup_tbs_r4}{ready4_sp_local}(x) <- value

\S4method{lup_tbs_r4}{ready4_sp_local_raw}(x) <- value

\S4method{lup_tbs_r4}{ready4_sp_local_proc}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_sp_local_proc}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot lup_tbs_r4

Set the value of the slot lup_tbs_r4 for S4 objects of class ready4_sp_local

Set the value of the slot lup_tbs_r4 for S4 objects of class ready4_sp_local_raw

Set the value of the slot lup_tbs_r4 for S4 objects of class ready4_sp_local_proc
}
