% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_1_clstr_1_srvc_trvl_tm}
\alias{make_1_clstr_1_srvc_trvl_tm}
\title{Make 1 clstr 1 srvc trvl tm}
\usage{
make_1_clstr_1_srvc_trvl_tm(
  cluster_tb,
  service,
  time_min,
  time_max,
  nbr_time_steps
)
}
\arguments{
\item{cluster_tb}{Cluster (a tibble)}

\item{service}{PARAM_DESCRIPTION}

\item{time_min}{PARAM_DESCRIPTION}

\item{time_max}{PARAM_DESCRIPTION}

\item{nbr_time_steps}{PARAM_DESCRIPTION}
}
\value{
One service (a simple features object)
}
\description{
make_1_clstr_1_srvc_trvl_tm() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make 1 clstr 1 srvc trvl tm. The function returns One service (a simple features object).
}
\keyword{internal}
