% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_agent_locations_tb}
\alias{make_agent_locations_tb}
\title{Make agent locations tibble}
\usage{
make_agent_locations_tb(
  profiled_area_sf,
  crs_dbl,
  person_ctg_1L_chr = character(0),
  person_type_1L_chr = "p",
  resolution_unit_1L_chr,
  year_1L_chr
)
}
\arguments{
\item{profiled_area_sf}{Profiled area (a simple features object)}

\item{crs_dbl}{Coordinates reference system (a double vector)}

\item{person_ctg_1L_chr}{Person category (a character vector of length one), Default: character(0)}

\item{person_type_1L_chr}{Person type (a character vector of length one), Default: 'p'}

\item{resolution_unit_1L_chr}{Resolution unit (a character vector of length one)}

\item{year_1L_chr}{Year (a character vector of length one)}
}
\value{
Agent coordinates (a tibble)
}
\description{
make_agent_locations_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make agent locations tibble. The function returns Agent coordinates (a tibble).
}
\keyword{internal}
