% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_agt_coords_tb}
\alias{make_agt_coords_tb}
\title{Make agt coords}
\usage{
make_agt_coords_tb(
  profiled_area_sf,
  disorder,
  year,
  case_type = "expected.incidence",
  person_type = "p",
  resolution_unit
)
}
\arguments{
\item{profiled_area_sf}{Profiled area (a simple features object)}

\item{disorder}{PARAM_DESCRIPTION}

\item{year}{PARAM_DESCRIPTION}

\item{case_type}{PARAM_DESCRIPTION, Default: 'expected.incidence'}

\item{person_type}{PARAM_DESCRIPTION, Default: 'p'}

\item{resolution_unit}{PARAM_DESCRIPTION}
}
\value{
Agent coordinates (a tibble)
}
\description{
make_agt_coords_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make agt coords tibble. The function returns Agent coordinates (a tibble).
}
\keyword{internal}
