% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_attr_data_xx}
\alias{make_attr_data_xx}
\title{Make attr data}
\usage{
make_attr_data_xx(lookup_tb_r4, lookup_ref, starter_sf)
}
\arguments{
\item{lookup_tb_r4}{Lookup (a ready4 S4 collection of tibbles)}

\item{lookup_ref}{PARAM_DESCRIPTION}

\item{starter_sf}{Starter (a simple features object)}
}
\description{
make_attr_data_xx() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make attr data output object of multiple potential types. The function is called for its side effects and does not return a value.
}
\keyword{internal}
