% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_closest_yrs_ls}
\alias{make_closest_yrs_ls}
\title{Make closest years list}
\usage{
make_closest_yrs_ls(
  data_lookup_tb,
  main_incld_feature_chr,
  target_year_1L_chr,
  target_area_1L_chr = character(0),
  approximation_1L_chr = "abs"
)
}
\arguments{
\item{data_lookup_tb}{Data lookup (a tibble)}

\item{main_incld_feature_chr}{Main included feature (a character vector)}

\item{target_year_1L_chr}{Target year (a character vector of length one)}

\item{target_area_1L_chr}{Target area (a character vector of length one), Default: character(0)}

\item{approximation_1L_chr}{Approximation (a character vector of length one), Default: 'abs'}
}
\value{
Closest years (a list)
}
\description{
make_closest_yrs_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make closest years list. The function returns Closest years (a list).
}
\keyword{internal}
