% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_cluster_bndys}
\alias{make_cluster_bndys}
\title{Make cluster boundaries}
\usage{
make_cluster_bndys(
  clusters_chr,
  crs_nbr_dbl,
  distance_in_km_1L_dbl,
  land_boundary_sf,
  vicinity_points_ls
)
}
\arguments{
\item{clusters_chr}{Clusters (a character vector)}

\item{crs_nbr_dbl}{Coordinates reference system number (a double vector)}

\item{distance_in_km_1L_dbl}{Distance in kilometre (a double vector of length one)}

\item{land_boundary_sf}{Land boundary (a simple features object)}

\item{vicinity_points_ls}{Vicinity points (a list)}
}
\value{
Cluster boundaries (a list)
}
\description{
make_cluster_bndys() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make cluster boundaries. The function returns Cluster boundaries (a list).
}
\keyword{internal}
