% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_cluster_isochrones}
\alias{make_cluster_isochrones}
\title{Make cluster isochrones}
\usage{
make_cluster_isochrones(
  vicinity_points_ls,
  index_val_1L_int,
  time_min_1L_dbl = 0,
  time_max_1L_dbl = 30,
  time_steps_1L_dbl = 5,
  travel_mode_1L_chr = "car"
)
}
\arguments{
\item{vicinity_points_ls}{Vicinity points (a list)}

\item{index_val_1L_int}{Index value (an integer vector of length one)}

\item{time_min_1L_dbl}{Time minimum (a double vector of length one), Default: 0}

\item{time_max_1L_dbl}{Time maximum (a double vector of length one), Default: 30}

\item{time_steps_1L_dbl}{Time steps (a double vector of length one), Default: 5}

\item{travel_mode_1L_chr}{Travel mode (a character vector of length one), Default: 'car'}
}
\value{
Discrete temporal bands (a list)
}
\description{
make_cluster_isochrones() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make cluster isochrones. The function returns Discrete temporal bands (a list).
}
\keyword{internal}
