% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_common_sf_yrs_ls}
\alias{make_common_sf_yrs_ls}
\title{Make common simple features object years list}
\usage{
make_common_sf_yrs_ls(sf_ls)
}
\arguments{
\item{sf_ls}{Simple features object list (a list of simple features objects)}
}
\value{
Common simple features object years (a list)
}
\description{
make_common_sf_yrs_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make common simple features object years list. The function returns Common simple features object years (a list).
}
\keyword{internal}
