% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_data_pack.R
\name{write_fls_and_mk_sngl_row_data_lup}
\alias{write_fls_and_mk_sngl_row_data_lup}
\title{write_fls_and_mk_sngl_row_data_lup}
\usage{
write_fls_and_mk_sngl_row_data_lup(
  x,
  merge_with,
  pckg_name,
  raw_data_dir,
  processed_dir,
  crs_nbr_vec = NA_real_,
  overwrite_lgl = F
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{merge_with}{PARAM_DESCRIPTION}

\item{pckg_name}{PARAM_DESCRIPTION}

\item{raw_data_dir}{PARAM_DESCRIPTION}

\item{processed_dir}{PARAM_DESCRIPTION}

\item{crs_nbr_vec}{PARAM_DESCRIPTION, Default: NA}

\item{overwrite_lgl}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{pull}}
 \code{\link[purrr]{map2}}
}
