% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_pack.R
\name{make_data_packs.ready4_sp_import_lup}
\alias{make_data_packs.ready4_sp_import_lup}
\title{make_data_packs.ready4_sp_import_lup}
\usage{
\method{make_data_packs}{ready4_sp_import_lup}(x, init_lookup_r4 = NULL,
  pckg_name, raw_data_dir, lup_dir = "data", processed_dir = "data",
  lup_r4_name)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{init_lookup_r4}{PARAM_DESCRIPTION, Default: NULL}

\item{pckg_name}{PARAM_DESCRIPTION}

\item{raw_data_dir}{PARAM_DESCRIPTION}

\item{lup_dir}{PARAM_DESCRIPTION, Default: 'data'}

\item{processed_dir}{PARAM_DESCRIPTION, Default: 'data'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready4s4]{ready4_lookup}}
 \code{\link[purrr]{reduce}}
 \code{\link[ready4utils]{add_all_tbs_in_r4}}
 \code{\link[dplyr]{slice}}
}
