% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_data_yrs_chr}
\alias{make_data_yrs_chr}
\title{Make data years character vector}
\usage{
make_data_yrs_chr(data_dtm)
}
\arguments{
\item{data_dtm}{Data (a date vector)}
}
\value{
Data years (a character vector)
}
\description{
make_data_yrs_chr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make data years character vector. The function returns Data years (a character vector).
}
\keyword{internal}
