% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_distance_based_bands}
\alias{make_distance_based_bands}
\title{Make distance based bands}
\usage{
make_distance_based_bands(
  distance_km_outer,
  nbr_distance_bands,
  service_cluster_tb,
  profiled_sf,
  crs_nbr
)
}
\arguments{
\item{distance_km_outer}{PARAM_DESCRIPTION}

\item{nbr_distance_bands}{PARAM_DESCRIPTION}

\item{service_cluster_tb}{Service cluster (a tibble)}

\item{profiled_sf}{Profiled (a simple features object)}

\item{crs_nbr}{PARAM_DESCRIPTION}
}
\value{
NA ()
}
\description{
make_distance_based_bands() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make distance based bands. The function is called for its side effects and does not return a value.
}
\keyword{internal}
