% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_each_uid_a_poly_sf}
\alias{make_each_uid_a_poly_sf}
\title{Make each uid a poly}
\usage{
make_each_uid_a_poly_sf(sf, uid_chr)
}
\arguments{
\item{sf}{Simple features object (a simple features object)}

\item{uid_chr}{Uid (a character vector)}
}
\description{
make_each_uid_a_poly_sf() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make each uid a poly simple features object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
