% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_env_param_tb}
\alias{make_env_param_tb}
\title{Make env param}
\usage{
make_env_param_tb(n_its_int, env_str_par_tb, mape_str_par_tb, jt_dist)
}
\arguments{
\item{n_its_int}{N its (an integer vector)}

\item{env_str_par_tb}{Env setter parameter (a tibble)}

\item{mape_str_par_tb}{Mape setter parameter (a tibble)}

\item{jt_dist}{PARAM_DESCRIPTION}
}
\description{
make_env_param_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make env param tibble. The function is called for its side effects and does not return a value.
}
\keyword{internal}
