% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_featured_var_pfx}
\alias{make_featured_var_pfx}
\title{Make featured variable prefix}
\usage{
make_featured_var_pfx(
  dynamic_var_rsl_1L_chr,
  reference_vals_chr,
  reference_var_rsl_1L_chr = NULL,
  data_year_1L_dbl
)
}
\arguments{
\item{dynamic_var_rsl_1L_chr}{Dynamic variable resolution (a character vector of length one)}

\item{reference_vals_chr}{Reference values (a character vector)}

\item{reference_var_rsl_1L_chr}{Reference variable resolution (a character vector of length one), Default: NULL}

\item{data_year_1L_dbl}{Data year (a double vector of length one)}
}
\value{
Prefix (a character vector of length one)
}
\description{
make_featured_var_pfx() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make featured variable prefix. The function returns Prefix (a character vector of length one).
}
\keyword{internal}
