% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_filter_by_year_logic}
\alias{make_filter_by_year_logic}
\title{Make filter by year logic}
\usage{
make_filter_by_year_logic(data_tb, years_chr)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{years_chr}{Years (a character vector)}
}
\value{
Filter by year (a logical vector)
}
\description{
make_filter_by_year_logic() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make filter by year logic. The function returns Filter by year (a logical vector).
}
\keyword{internal}
