% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_geomc_dist_boundrs}
\alias{make_geomc_dist_boundrs}
\title{Make geomc dist boundrs}
\usage{
make_geomc_dist_boundrs(point_locations, land_sf, distance, crs_nbr)
}
\arguments{
\item{point_locations}{PARAM_DESCRIPTION}

\item{land_sf}{Land (a simple features object)}

\item{distance}{PARAM_DESCRIPTION}

\item{crs_nbr}{PARAM_DESCRIPTION}
}
\value{
Distance from prototypes on land (a simple features object)
}
\description{
make_geomc_dist_boundrs() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make geomc dist boundrs. The function returns Distance from prototypes on land (a simple features object).
}
\keyword{internal}
