% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_import_object.R
\name{make_import_object.ready4_sp_import_lup}
\alias{make_import_object.ready4_sp_import_lup}
\title{make_import_object.ready4_sp_import_lup}
\usage{
make_import_object.ready4_sp_import_lup(x, var_val_vec, path_str)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{var_val_vec}{PARAM_DESCRIPTION}

\item{path_str}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
