% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_intersecting_geometries}
\alias{make_intersecting_geometries}
\title{Make intersecting geometries}
\usage{
make_intersecting_geometries(
  geometry_one_sf,
  geometry_two_sf,
  crs_nbr_dbl,
  validate_1L_lgl = T
)
}
\arguments{
\item{geometry_one_sf}{Geometry one (a simple features object)}

\item{geometry_two_sf}{Geometry two (a simple features object)}

\item{crs_nbr_dbl}{Coordinates reference system number (a double vector)}

\item{validate_1L_lgl}{Validate (a logical vector of length one), Default: T}
}
\value{
Intersection (a simple features object)
}
\description{
make_intersecting_geometries() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make intersecting geometries. The function returns Intersection (a simple features object).
}
\keyword{internal}
