% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_intersecting_profiled_area}
\alias{make_intersecting_profiled_area}
\title{Make intersecting profiled area}
\usage{
make_intersecting_profiled_area(
  attribute_rsl_1L_chr,
  attribute_sf,
  crs_nbr_dbl,
  data_type_chr,
  data_year_1L_chr,
  profiled_sf,
  featured_var_pfx_1L_chr = character(0),
  profiled_sf_col_1L_chr = NA_character_,
  profiled_sf_row_1L_chr = NA_character_
)
}
\arguments{
\item{attribute_rsl_1L_chr}{Attribute resolution (a character vector of length one)}

\item{attribute_sf}{Attribute (a simple features object)}

\item{crs_nbr_dbl}{Coordinates reference system number (a double vector)}

\item{data_type_chr}{Data type (a character vector)}

\item{data_year_1L_chr}{Data year (a character vector of length one)}

\item{profiled_sf}{Profiled (a simple features object)}

\item{featured_var_pfx_1L_chr}{Featured variable prefix (a character vector of length one), Default: character(0)}

\item{profiled_sf_col_1L_chr}{Profiled simple features object column (a character vector of length one), Default: 'NA'}

\item{profiled_sf_row_1L_chr}{Profiled simple features object row (a character vector of length one), Default: 'NA'}
}
\value{
Profiled (a simple features object)
}
\description{
make_intersecting_profiled_area() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make intersecting profiled area. The function returns Profiled (a simple features object).
}
\keyword{internal}
