% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_isochrone_bands}
\alias{make_isochrone_bands}
\title{Make isochrone bands}
\usage{
make_isochrone_bands(
  index_val_1L_int,
  cluster_isochrone_ls,
  travel_mode_1L_chr
)
}
\arguments{
\item{index_val_1L_int}{Index value (an integer vector of length one)}

\item{cluster_isochrone_ls}{Cluster isochrone (a list)}

\item{travel_mode_1L_chr}{Travel mode (a character vector of length one)}
}
\value{
Time band (a list of simple features objects)
}
\description{
make_isochrone_bands() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make isochrone bands. The function returns Time band (a list of simple features objects).
}
\keyword{internal}
