% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_isochrones}
\alias{make_isochrones}
\title{Make isochrones}
\usage{
make_isochrones(
  lat_1L_dbl,
  lng_1L_dbl,
  server_1L_chr = character(0),
  time_min_1L_dbl,
  time_max_1L_dbl,
  time_steps_1L_dbl,
  travel_mode_1L_chr = "car"
)
}
\arguments{
\item{lat_1L_dbl}{Lat (a double vector of length one)}

\item{lng_1L_dbl}{Longitude (a double vector of length one)}

\item{server_1L_chr}{Server (a character vector of length one), Default: character(0)}

\item{time_min_1L_dbl}{Time minimum (a double vector of length one)}

\item{time_max_1L_dbl}{Time maximum (a double vector of length one)}

\item{time_steps_1L_dbl}{Time steps (a double vector of length one)}

\item{travel_mode_1L_chr}{Travel mode (a character vector of length one), Default: 'car'}
}
\value{
Iso (a simple features object)
}
\description{
make_isochrones() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make isochrones. The function returns Iso (a simple features object).
}
\keyword{internal}
