% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_isochrs}
\alias{make_isochrs}
\title{Make isochrs}
\usage{
make_isochrs(long, lat, time_min, time_max, nbr_time_steps)
}
\arguments{
\item{long}{PARAM_DESCRIPTION}

\item{time_min}{PARAM_DESCRIPTION}

\item{time_max}{PARAM_DESCRIPTION}

\item{nbr_time_steps}{PARAM_DESCRIPTION}

\item{...}{Additional arguments}
}
\value{
Iso (a simple features object)
}
\description{
make_isochrs() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make isochrs. The function returns Iso (a simple features object).
}
\keyword{internal}
