% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_km_sqd_dbl}
\alias{make_km_sqd_dbl}
\title{Make kilometre squared double vector}
\usage{
make_km_sqd_dbl(data_sf)
}
\arguments{
\item{data_sf}{Data (a simple features object)}
}
\value{
Kilometre squared (a double vector)
}
\description{
make_km_sqd_dbl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make kilometre squared double vector. The function returns Kilometre squared (a double vector).
}
\keyword{internal}
