% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_ready4_local_proc_r4.R
\name{make_local_proc_r4}
\alias{make_local_proc_r4}
\alias{make_local_proc_r4,ready4_script_data-method}
\alias{make_local_proc_r4,ready4_sp_local-method}
\title{Make a ready4_local_proc_r4 object}
\usage{
\S4method{make_local_proc_r4}{ready4_script_data}(x, raw_data_dir_chr, save_lgl)

\S4method{make_local_proc_r4}{ready4_sp_local}(x, import_chr_vec, raw_data_dir_chr, save_lgl)
}
\arguments{
\item{x}{Primary input object - see Usage section for allowable types}

\item{raw_data_dir_chr}{A character}

\item{save_lgl}{A logical}

\item{import_chr_vec}{A character}
}
\description{
Make a ready4_local_proc_r4 object from a range of potential inputs
}
