% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_main_intersect_tb.R
\name{make_main_intersect_tb}
\alias{make_main_intersect_tb}
\alias{make_main_intersect_tb,ready4_lookup-method}
\title{make_main_intersect_tb}
\usage{
make_main_intersect_tb(x, ...)

\S4method{make_main_intersect_tb}{ready4_lookup}(
  x,
  r_data_dir_chr,
  template_tb,
  tb_ref_var_chr,
  ref_bound_yr,
  ref_unit_chr,
  tgt_bound_yr,
  tgt_unit_chr
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{r_data_dir_chr}{PARAM_DESCRIPTION}

\item{template_tb}{PARAM_DESCRIPTION}

\item{tb_ref_var_chr}{PARAM_DESCRIPTION}

\item{ref_bound_yr}{PARAM_DESCRIPTION}

\item{ref_unit_chr}{PARAM_DESCRIPTION}

\item{tgt_bound_yr}{PARAM_DESCRIPTION}

\item{tgt_unit_chr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
