% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_main_intersect_tb.R
\name{make_main_intersect_tb.ready4_lookup}
\alias{make_main_intersect_tb.ready4_lookup}
\title{make_main_intersect_tb.ready4_lookup}
\usage{
make_main_intersect_tb.ready4_lookup(
  x,
  r_data_dir_chr,
  template_tb,
  tb_ref_var_chr,
  ref_bound_yr,
  ref_unit_chr,
  tgt_bound_yr,
  tgt_unit_chr
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{r_data_dir_chr}{PARAM_DESCRIPTION}

\item{template_tb}{PARAM_DESCRIPTION}

\item{tb_ref_var_chr}{PARAM_DESCRIPTION}

\item{ref_bound_yr}{PARAM_DESCRIPTION}

\item{ref_unit_chr}{PARAM_DESCRIPTION}

\item{tgt_bound_yr}{PARAM_DESCRIPTION}

\item{tgt_unit_chr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{pull}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}}
 \code{\link[rlang]{sym}}
 \code{\link[sf]{geos_binary_ops}},\code{\link[sf]{geos_measures}},\code{\link[sf]{st_geometry}}
 \code{\link[units]{set_units}}
 \code{\link[purrr]{pluck}},\code{\link[purrr]{map}}
 \code{\link[ready4utils]{data_get}}
}
