% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_par_str_envir.R
\name{make_new_ready4_param_str_envir}
\alias{make_new_ready4_param_str_envir}
\title{Make new Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to the spatial environment.}
\usage{
make_new_ready4_param_str_envir(x)
}
\arguments{
\item{x}{A prototype for the Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to the spatial environment.}
}
\value{
An unvalidated instance of the Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to the spatial environment.
}
\description{
Create a new unvalidated instance of the Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to the spatial environment.
}
\details{
Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to the spatial environment.
}
