% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_par_val_envir.R
\name{make_new_ready4_param_val_envir}
\alias{make_new_ready4_param_val_envir}
\title{Make new Readyforwhatsnext S3 class for tibble object that stores simulation parameter values for each iteration.}
\usage{
make_new_ready4_param_val_envir(x)
}
\arguments{
\item{x}{A prototype for the Readyforwhatsnext S3 class for tibble object that stores simulation parameter values for each iteration.}
}
\value{
An unvalidated instance of the Readyforwhatsnext S3 class for tibble object that stores simulation parameter values for each iteration.
}
\description{
Create a new unvalidated instance of the Readyforwhatsnext S3 class for tibble object that stores simulation parameter values for each iteration.
}
\details{
Readyforwhatsnext S3 class for tibble object that stores simulation parameter values for each iteration.
}
