% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_abbreviations_lup.R
\name{make_new_ready4_sp_abbreviations_lup}
\alias{make_new_ready4_sp_abbreviations_lup}
\title{Make new Readyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations.}
\usage{
make_new_ready4_sp_abbreviations_lup(x)
}
\arguments{
\item{x}{A prototype for the Readyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations.}
}
\value{
An unvalidated instance of the Readyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations.
}
\description{
Create a new unvalidated instance of the Readyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations.
}
