% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_data_pack_lup.R
\name{make_new_ready4_sp_data_pack_lup}
\alias{make_new_ready4_sp_data_pack_lup}
\title{Make new Readyforwhatsnext S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).}
\usage{
make_new_ready4_sp_data_pack_lup(x)
}
\arguments{
\item{x}{A prototype for the Readyforwhatsnext S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).}
}
\value{
An unvalidated instance of the Readyforwhatsnext S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).
}
\description{
Create a new unvalidated instance of the Readyforwhatsnext S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).
}
