% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_import_lup.R
\name{make_new_ready4_sp_import_lup}
\alias{make_new_ready4_sp_import_lup}
\title{Make new Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.}
\usage{
make_new_ready4_sp_import_lup(x)
}
\arguments{
\item{x}{A prototype for the Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.}
}
\value{
An unvalidated instance of the Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.
}
\description{
Create a new unvalidated instance of the Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.
}
