% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_site_coord_lup.R
\name{make_new_ready4_sp_site_coord_lup}
\alias{make_new_ready4_sp_site_coord_lup}
\title{Make new Readyforwhatsnext S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.}
\usage{
make_new_ready4_sp_site_coord_lup(x)
}
\arguments{
\item{x}{A prototype for the Readyforwhatsnext S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.}
}
\value{
An unvalidated instance of the Readyforwhatsnext S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.
}
\description{
Create a new unvalidated instance of the Readyforwhatsnext S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.
}
