% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_starter_sf_lup.R
\name{make_new_ready4_sp_starter_sf_lup}
\alias{make_new_ready4_sp_starter_sf_lup}
\title{Make new Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.}
\usage{
make_new_ready4_sp_starter_sf_lup(x)
}
\arguments{
\item{x}{A prototype for the Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.}
}
\value{
An unvalidated instance of the Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
\description{
Create a new unvalidated instance of the Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
