% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_uid_lup.R
\name{make_new_ready4_sp_uid_lup}
\alias{make_new_ready4_sp_uid_lup}
\title{Make new Readyforwhatsnext S3 class for tibble object lookup table of unique feature identifiers used for different spatial objects.}
\usage{
make_new_ready4_sp_uid_lup(x)
}
\arguments{
\item{x}{A prototype for the Readyforwhatsnext S3 class for tibble object lookup table of unique feature identifiers used for different spatial objects.}
}
\value{
An unvalidated instance of the Readyforwhatsnext S3 class for tibble object lookup table of unique feature identifiers used for different spatial objects.
}
\description{
Create a new unvalidated instance of the Readyforwhatsnext S3 class for tibble object lookup table of unique feature identifiers used for different spatial objects.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of unique feature identifiers used for different spatial objects.
}
