% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_vicinity_abbreviations.R
\name{make_new_vicinity_abbreviations}
\alias{make_new_vicinity_abbreviations}
\title{make new vicinity abbreviations ready4 submodule class for tibble object lookup table for spatial data abbreviations.}
\usage{
make_new_vicinity_abbreviations(x)
}
\arguments{
\item{x}{A prototype for the ready4 submodule class for tibble object lookup table for spatial data abbreviations.}
}
\value{
An unvalidated instance of the ready4 submodule class for tibble object lookup table for spatial data abbreviations.
}
\description{
Create a new unvalidated instance of the ready4 submodule class for tibble object lookup table for spatial data abbreviations.
}
\details{
ready4 submodule for tibble object lookup table for spatial data abbreviations.
}
\keyword{internal}
