% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_vicinity_mapes.R
\name{make_new_vicinity_mapes}
\alias{make_new_vicinity_mapes}
\title{make new vicinity mean absolute prediction errors ready4 S3 class for tibble object that stores spatial simulation parameters relating to Mean Absolute Prediction Errors.}
\usage{
make_new_vicinity_mapes(x)
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for tibble object that stores spatial simulation parameters relating to Mean Absolute Prediction Errors.}
}
\value{
An unvalidated instance of the ready4 S3 class for tibble object that stores spatial simulation parameters relating to Mean Absolute Prediction Errors.
}
\description{
Create a new unvalidated instance of the ready4 S3 class for tibble object that stores spatial simulation parameters relating to Mean Absolute Prediction Errors.
}
\details{
ready4 S3 class for tibble object that stores spatial simulation parameters relating to Mean Absolute Prediction Errors.
}
\keyword{internal}
