% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_nse_objs_ls}
\alias{make_nse_objs_ls}
\title{Make nse objects}
\usage{
make_nse_objs_ls(
  sp_unit,
  concept,
  tot_pop_col = NULL,
  grouping_1 = NULL,
  data_year,
  popl_var_prefix
)
}
\arguments{
\item{sp_unit}{PARAM_DESCRIPTION}

\item{concept}{PARAM_DESCRIPTION}

\item{tot_pop_col}{PARAM_DESCRIPTION, Default: NULL}

\item{grouping_1}{PARAM_DESCRIPTION, Default: NULL}

\item{data_year}{PARAM_DESCRIPTION}

\item{popl_var_prefix}{PARAM_DESCRIPTION}
}
\description{
make_nse_objs_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make nse objects list. The function is called for its side effects and does not return a value.
}
\keyword{internal}
