% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_paths_to_fls_for_ingest}
\alias{make_paths_to_fls_for_ingest}
\title{Make paths to files for ingest}
\usage{
make_paths_to_fls_for_ingest(data_type_chr, name_chr, processed_fls_dir_1L_chr)
}
\arguments{
\item{data_type_chr}{Data type (a character vector)}

\item{name_chr}{Name (a character vector)}

\item{processed_fls_dir_1L_chr}{Processed files directory (a character vector of length one)}
}
\value{
Paths (a character vector)
}
\description{
make_paths_to_fls_for_ingest() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make paths to files for ingest. The function returns Paths (a character vector).
}
\keyword{internal}
