% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_polygons_from_duplicates}
\alias{make_polygons_from_duplicates}
\title{Make polygons from duplicates}
\usage{
make_polygons_from_duplicates(sf, uid_1L_chr)
}
\arguments{
\item{sf}{Simple features object (a simple features object)}

\item{uid_1L_chr}{Unique identifier (a character vector of length one)}
}
\value{
Polygons (a simple features object)
}
\description{
make_polygons_from_duplicates() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make polygons from duplicates. The function returns Polygons (a simple features object).
}
\keyword{internal}
