% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prev_rates_sum_tb.R
\name{make_prev_rates_sum_tb}
\alias{make_prev_rates_sum_tb}
\title{make_prev_rates_sum_tb
Create a summary table of prevalence rate data inputs to simulation model.}
\usage{
make_prev_rates_sum_tb(prev_rates_vec)
}
\arguments{
\item{prev_rates_vec}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
make_prev_rates_sum_tb
Create a summary table of prevalence rate data inputs to simulation model.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_detect}},\code{\link[stringr]{str_sub}}
 \code{\link[tibble]{tibble}}
 \code{\link[dplyr]{join}}
}
