% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prevalence.R
\name{make_prev_struc_par_tb}
\alias{make_prev_struc_par_tb}
\title{make_prev_struc_par_tb
FUNCTION_DESCRIPTION}
\usage{
make_prev_struc_par_tb(disorder, period, ages, sexes, pref_source,
  prev_rates)
}
\arguments{
\item{disorder}{PARAM_DESCRIPTION}

\item{period}{PARAM_DESCRIPTION}

\item{ages}{PARAM_DESCRIPTION}

\item{sexes}{PARAM_DESCRIPTION}

\item{pref_source}{PARAM_DESCRIPTION}

\item{prev_rates}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
make_prev_struc_par_tb
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}},\code{\link[dplyr]{reexports}}
 \code{\link[purrr]{map2}}
 \code{\link[stringr]{str_sub}}
 \code{\link[stats]{setNames}}
}
