% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prevalence.R
\name{make_prev_summ_tb}
\alias{make_prev_summ_tb}
\title{make_prev_summ_tb
FUNCTION_DESCRIPTION}
\usage{
make_prev_summ_tb(prev_summary)
}
\arguments{
\item{prev_summary}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
make_prev_summ_tb
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{summarise_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{funs}},\code{\link[dplyr]{mutate}}
 \code{\link[stringr]{str_sub}}
 \code{\link[tibble]{tibble}}
}
