% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_profiled_area_objs}
\alias{make_profiled_area_objs}
\title{Make profiled area objects}
\usage{
make_profiled_area_objs(pa_r4)
}
\arguments{
\item{pa_r4}{Pa (a ready4 S4)}
}
\description{
make_profiled_area_objs() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make profiled area objects. The function is called for its side effects and does not return a value.
}
\keyword{internal}
