% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_par_val_envir.R
\name{make_pt_ready4_param_val_envir}
\alias{make_pt_ready4_param_val_envir}
\title{Make prototype Readyforwhatsnext S3 class for tibble object that stores simulation parameter values for each iteration.}
\usage{
make_pt_ready4_param_val_envir(
  param_name_chr = character(0),
  v_it_1 = numeric(0)
)
}
\arguments{
\item{param_name_chr}{PARAM_DESCRIPTION, Default: character(0)}

\item{v_it_1}{PARAM_DESCRIPTION, Default: numeric(0)}
}
\value{
A prototype for Readyforwhatsnext S3 class for tibble object that stores simulation parameter values for each iteration.
}
\description{
Create a new prototype for the Readyforwhatsnext S3 class for tibble object that stores simulation parameter values for each iteration.
}
\details{
Readyforwhatsnext S3 class for tibble object that stores simulation parameter values for each iteration.
}
