% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_abbreviations_lup.R
\name{make_pt_ready4_sp_abbreviations_lup}
\alias{make_pt_ready4_sp_abbreviations_lup}
\title{Make prototype Readyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations.}
\usage{
make_pt_ready4_sp_abbreviations_lup(
  long_name = character(0),
  short_name = character(0)
)
}
\arguments{
\item{long_name}{PARAM_DESCRIPTION, Default: character(0)}

\item{short_name}{PARAM_DESCRIPTION, Default: character(0)}
}
\value{
A prototype for Readyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations.
}
\description{
Create a new prototype for the Readyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations.
}
