% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_data_pack_lup.R
\name{make_pt_ready4_sp_data_pack_lup}
\alias{make_pt_ready4_sp_data_pack_lup}
\title{Make prototype Readyforwhatsnext S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).}
\usage{
make_pt_ready4_sp_data_pack_lup(
  name = character(0),
  country = character(0),
  area_type = character(0),
  area_bound_yr = character(0),
  region = character(0),
  data_type = character(0),
  main_feature = character(0),
  year = character(0),
  year_start = character(0),
  year_end = character(0),
  source_reference = character(0),
  additional_detail = character(0)
)
}
\arguments{
\item{name}{PARAM_DESCRIPTION, Default: character(0)}

\item{country}{PARAM_DESCRIPTION, Default: character(0)}

\item{area_type}{PARAM_DESCRIPTION, Default: character(0)}

\item{area_bound_yr}{PARAM_DESCRIPTION, Default: character(0)}

\item{region}{PARAM_DESCRIPTION, Default: character(0)}

\item{data_type}{PARAM_DESCRIPTION, Default: character(0)}

\item{main_feature}{PARAM_DESCRIPTION, Default: character(0)}

\item{year}{PARAM_DESCRIPTION, Default: character(0)}

\item{year_start}{PARAM_DESCRIPTION, Default: character(0)}

\item{year_end}{PARAM_DESCRIPTION, Default: character(0)}

\item{source_reference}{PARAM_DESCRIPTION, Default: character(0)}

\item{additional_detail}{PARAM_DESCRIPTION, Default: character(0)}
}
\value{
A prototype for Readyforwhatsnext S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).
}
\description{
Create a new prototype for the Readyforwhatsnext S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).
}
