% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_import_lup.R
\name{make_pt_ready4_sp_import_lup}
\alias{make_pt_ready4_sp_import_lup}
\title{Make prototype Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.}
\usage{
make_pt_ready4_sp_import_lup(
  file_type = character(0),
  file_name = character(0),
  data_repo = character(0),
  data_repo_ui = character(0),
  data_repo_db_ui = character(0),
  data_repo_file_ext = character(0),
  data_repo_write_type_1L_chr = character(0),
  local_file_src_chr = character(0),
  path_to_make_script_chr = character(0),
  download_url_chr = character(0),
  inc_file_main_chr = character(0),
  inc_fls_to_rename_ls = list(),
  new_nms_for_inc_fls_ls = list(),
  name = character(0),
  country = character(0),
  area_type = character(0),
  area_bound_yr = character(0),
  region = character(0),
  data_type = character(0),
  main_feature = character(0),
  year = character(0),
  year_start = character(0),
  year_end = character(0),
  uid = character(0),
  add_boundaries = list()
)
}
\arguments{
\item{file_type}{PARAM_DESCRIPTION, Default: character(0)}

\item{file_name}{PARAM_DESCRIPTION, Default: character(0)}

\item{data_repo}{PARAM_DESCRIPTION, Default: character(0)}

\item{data_repo_ui}{PARAM_DESCRIPTION, Default: character(0)}

\item{data_repo_db_ui}{PARAM_DESCRIPTION, Default: character(0)}

\item{data_repo_file_ext}{PARAM_DESCRIPTION, Default: character(0)}

\item{data_repo_write_type_1L_chr}{PARAM_DESCRIPTION, Default: character(0)}

\item{local_file_src_chr}{PARAM_DESCRIPTION, Default: character(0)}

\item{path_to_make_script_chr}{PARAM_DESCRIPTION, Default: character(0)}

\item{download_url_chr}{PARAM_DESCRIPTION, Default: character(0)}

\item{inc_file_main_chr}{PARAM_DESCRIPTION, Default: character(0)}

\item{inc_fls_to_rename_ls}{PARAM_DESCRIPTION, Default: list()}

\item{new_nms_for_inc_fls_ls}{PARAM_DESCRIPTION, Default: list()}

\item{name}{PARAM_DESCRIPTION, Default: character(0)}

\item{country}{PARAM_DESCRIPTION, Default: character(0)}

\item{area_type}{PARAM_DESCRIPTION, Default: character(0)}

\item{area_bound_yr}{PARAM_DESCRIPTION, Default: character(0)}

\item{region}{PARAM_DESCRIPTION, Default: character(0)}

\item{data_type}{PARAM_DESCRIPTION, Default: character(0)}

\item{main_feature}{PARAM_DESCRIPTION, Default: character(0)}

\item{year}{PARAM_DESCRIPTION, Default: character(0)}

\item{year_start}{PARAM_DESCRIPTION, Default: character(0)}

\item{year_end}{PARAM_DESCRIPTION, Default: character(0)}

\item{add_boundaries}{PARAM_DESCRIPTION, Default: list()}

\item{...}{Additional arguments, Default: character(0)}
}
\value{
A prototype for Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.
}
\description{
Create a new prototype for the Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.
}
