% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_resolution_lup.R
\name{make_pt_ready4_sp_resolution_lup}
\alias{make_pt_ready4_sp_resolution_lup}
\title{Make prototype Readyforwhatsnext S3 class for tibble object lookup table of the relative resolutions of different spatial objects.}
\usage{
make_pt_ready4_sp_resolution_lup(
  parent_area = character(0),
  boundary_year = numeric(0),
  area_type = character(0),
  area_count = numeric(0),
  complete = logical(0),
  summed_area = numeric(0),
  mean_size = numeric(0)
)
}
\arguments{
\item{parent_area}{PARAM_DESCRIPTION, Default: character(0)}

\item{boundary_year}{PARAM_DESCRIPTION, Default: numeric(0)}

\item{area_type}{PARAM_DESCRIPTION, Default: character(0)}

\item{area_count}{PARAM_DESCRIPTION, Default: numeric(0)}

\item{complete}{PARAM_DESCRIPTION, Default: logical(0)}

\item{summed_area}{PARAM_DESCRIPTION, Default: numeric(0)}

\item{mean_size}{PARAM_DESCRIPTION, Default: numeric(0)}
}
\value{
A prototype for Readyforwhatsnext S3 class for tibble object lookup table of the relative resolutions of different spatial objects.
}
\description{
Create a new prototype for the Readyforwhatsnext S3 class for tibble object lookup table of the relative resolutions of different spatial objects.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of the relative resolutions of different spatial objects.
}
