% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_site_coord_lup.R
\name{make_pt_ready4_sp_site_coord_lup}
\alias{make_pt_ready4_sp_site_coord_lup}
\title{Make prototype Readyforwhatsnext S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.}
\usage{
make_pt_ready4_sp_site_coord_lup(
  service_type = character(0),
  cluster_name = character(0),
  service_name = character(0),
  lat = numeric(0),
  long = numeric(0)
)
}
\arguments{
\item{service_type}{PARAM_DESCRIPTION, Default: character(0)}

\item{cluster_name}{PARAM_DESCRIPTION, Default: character(0)}

\item{service_name}{PARAM_DESCRIPTION, Default: character(0)}

\item{long}{PARAM_DESCRIPTION, Default: numeric(0)}

\item{...}{Additional arguments, Default: numeric(0)}
}
\value{
A prototype for Readyforwhatsnext S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.
}
\description{
Create a new prototype for the Readyforwhatsnext S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.
}
