% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_starter_sf_lup.R
\name{make_pt_ready4_sp_starter_sf_lup}
\alias{make_pt_ready4_sp_starter_sf_lup}
\title{Make prototype Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.}
\usage{
make_pt_ready4_sp_starter_sf_lup(
  country = character(0),
  area_type = character(0),
  area_bound_yr = character(0),
  starter_sf = character(0),
  sf_main_sub_div = character(0)
)
}
\arguments{
\item{country}{PARAM_DESCRIPTION, Default: character(0)}

\item{area_type}{PARAM_DESCRIPTION, Default: character(0)}

\item{area_bound_yr}{PARAM_DESCRIPTION, Default: character(0)}

\item{starter_sf}{Starter (a simple features object), Default: character(0)}

\item{sf_main_sub_div}{PARAM_DESCRIPTION, Default: character(0)}
}
\value{
A prototype for Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
\description{
Create a new prototype for the Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
