% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_uid_lup.R
\name{make_pt_ready4_sp_uid_lup}
\alias{make_pt_ready4_sp_uid_lup}
\title{Make prototype Readyforwhatsnext S3 class for tibble object lookup table of unique feature identifiers used for different spatial objects.}
\usage{
make_pt_ready4_sp_uid_lup(
  spatial_unit = character(0),
  year = character(0),
  var_name = character(0)
)
}
\arguments{
\item{spatial_unit}{PARAM_DESCRIPTION, Default: character(0)}

\item{year}{PARAM_DESCRIPTION, Default: character(0)}

\item{var_name}{PARAM_DESCRIPTION, Default: character(0)}
}
\value{
A prototype for Readyforwhatsnext S3 class for tibble object lookup table of unique feature identifiers used for different spatial objects.
}
\description{
Create a new prototype for the Readyforwhatsnext S3 class for tibble object lookup table of unique feature identifiers used for different spatial objects.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of unique feature identifiers used for different spatial objects.
}
