% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_data_pack.R
\name{write_raw_format_dir}
\alias{write_raw_format_dir}
\title{write_raw_format_dir}
\usage{
write_raw_format_dir(data_type_chr, raw_data_dir)
}
\arguments{
\item{data_type_chr}{PARAM_DESCRIPTION}

\item{raw_data_dir}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
