% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_raw_format_dir_chr}
\alias{make_raw_format_dir_chr}
\title{Make raw format directory}
\usage{
make_raw_format_dir_chr(raw_data_dir, category)
}
\arguments{
\item{raw_data_dir}{PARAM_DESCRIPTION}

\item{category}{PARAM_DESCRIPTION}
}
\description{
make_raw_format_dir_chr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make raw format directory character vector. The function is called for its side effects and does not return a value.
}
\keyword{internal}
