% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_reconciled_intersecting_area}
\alias{make_reconciled_intersecting_area}
\title{Make reconciled intersecting area}
\usage{
make_reconciled_intersecting_area(
  profiled_sf,
  crs_nbr_dbl,
  data_year_1L_chr,
  dynamic_var_rsl_1L_chr,
  group_by_var_1L_chr,
  reference_grouping_1L_chr,
  reference_vals_chr,
  reference_var_rsl_1L_chr,
  spatial_atts_ls,
  profiled_sf_col_1L_chr = NA,
  profiled_sf_row_1L_chr = NA
)
}
\arguments{
\item{profiled_sf}{Profiled (a simple features object)}

\item{crs_nbr_dbl}{Coordinates reference system number (a double vector)}

\item{data_year_1L_chr}{Data year (a character vector of length one)}

\item{dynamic_var_rsl_1L_chr}{Dynamic variable resolution (a character vector of length one)}

\item{group_by_var_1L_chr}{Group by variable (a character vector of length one)}

\item{reference_grouping_1L_chr}{Reference grouping (a character vector of length one)}

\item{reference_vals_chr}{Reference values (a character vector)}

\item{reference_var_rsl_1L_chr}{Reference variable resolution (a character vector of length one)}

\item{spatial_atts_ls}{Spatial attributes (a list)}

\item{profiled_sf_col_1L_chr}{Profiled simple features object column (a character vector of length one), Default: NA}

\item{profiled_sf_row_1L_chr}{Profiled simple features object row (a character vector of length one), Default: NA}
}
\value{
Profiled (a simple features object)
}
\description{
make_reconciled_intersecting_area() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make reconciled intersecting area. The function returns Profiled (a simple features object).
}
\keyword{internal}
