% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_servc_clstr_isochrs_ls}
\alias{make_servc_clstr_isochrs_ls}
\title{Make servc clstr isochrs}
\usage{
make_servc_clstr_isochrs_ls(
  cluster_tbs_list,
  look_up_ref,
  time_min = 0,
  time_max = 60,
  nbr_time_steps = 5
)
}
\arguments{
\item{cluster_tbs_list}{PARAM_DESCRIPTION}

\item{look_up_ref}{PARAM_DESCRIPTION}

\item{time_min}{PARAM_DESCRIPTION, Default: 0}

\item{time_max}{PARAM_DESCRIPTION, Default: 60}

\item{nbr_time_steps}{PARAM_DESCRIPTION, Default: 5}
}
\value{
NA ()
}
\description{
make_servc_clstr_isochrs_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make servc clstr isochrs list. The function is called for its side effects and does not return a value.
}
\keyword{internal}
