% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_sf_ls}
\alias{make_sf_ls}
\title{Make simple features object list}
\usage{
make_sf_ls(profiled_sf, group_by_var_1L_chr)
}
\arguments{
\item{profiled_sf}{Profiled (a simple features object)}

\item{group_by_var_1L_chr}{Group by variable (a character vector of length one)}
}
\value{
Simple features object list (a list of simple features objects)
}
\description{
make_sf_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make simple features object list. The function returns Simple features object list (a list of simple features objects).
}
\keyword{internal}
