% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_sf_rows_fn}
\alias{make_sf_rows_fn}
\title{Make simple features object rows function}
\usage{
make_sf_rows_fn(...)
}
\arguments{
\item{...}{Additional arguments}
}
\value{
Simple features object rows (a function)
}
\description{
make_sf_rows_fn() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make simple features object rows function. The function returns Simple features object rows (a function).
}
\keyword{internal}
