% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_sp_data_list}
\alias{make_sp_data_list}
\title{Make sp data list}
\usage{
make_sp_data_list(input_ls, sub_div_units_vec)
}
\arguments{
\item{input_ls}{Input (a list)}

\item{sub_div_units_vec}{PARAM_DESCRIPTION}
}
\value{
NA ()
}
\description{
make_sp_data_list() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make sp data list. The function is called for its side effects and does not return a value.
}
\keyword{internal}
