% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_srvc_clstr_geomc_dist_boundrs}
\alias{make_srvc_clstr_geomc_dist_boundrs}
\title{Make srvc clstr geomc dist boundrs}
\usage{
make_srvc_clstr_geomc_dist_boundrs(
  distance_km,
  clusters_vec,
  clusters_tbs_list,
  land_boundary_sf,
  crs_nbr
)
}
\arguments{
\item{distance_km}{PARAM_DESCRIPTION}

\item{clusters_vec}{PARAM_DESCRIPTION}

\item{clusters_tbs_list}{PARAM_DESCRIPTION}

\item{land_boundary_sf}{Land boundary (a simple features object)}

\item{crs_nbr}{PARAM_DESCRIPTION}
}
\description{
make_srvc_clstr_geomc_dist_boundrs() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make srvc clstr geomc dist boundrs. The function is called for its side effects and does not return a value.
}
\keyword{internal}
