% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_time_band_sf_ls}
\alias{make_time_band_sf_ls}
\title{Make time band}
\usage{
make_time_band_sf_ls(look_up_ref, one_cluster_travel_time_sf_list)
}
\arguments{
\item{look_up_ref}{PARAM_DESCRIPTION}

\item{one_cluster_travel_time_sf_list}{PARAM_DESCRIPTION}
}
\value{
Time band (a list of simple features objects)
}
\description{
make_time_band_sf_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make time band simple features object list. The function returns Time band (a list of simple features objects).
}
\keyword{internal}
