% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_trvl_tm_isochrs}
\alias{make_trvl_tm_isochrs}
\title{Make trvl tm isochrs}
\usage{
make_trvl_tm_isochrs(
  appID,
  apiKey,
  origin,
  mode_of_transport = "driving",
  travel_time_hours,
  crs
)
}
\arguments{
\item{appID}{PARAM_DESCRIPTION}

\item{apiKey}{PARAM_DESCRIPTION}

\item{origin}{PARAM_DESCRIPTION}

\item{mode_of_transport}{PARAM_DESCRIPTION, Default: 'driving'}

\item{travel_time_hours}{PARAM_DESCRIPTION}

\item{...}{Additional arguments}
}
\value{
New (a simple features object)
}
\description{
make_trvl_tm_isochrs() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make trvl tm isochrs. The function returns New (a simple features object).
}
\keyword{internal}
