% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_valid_new_sf}
\alias{make_valid_new_sf}
\title{Make valid new simple features object}
\usage{
make_valid_new_sf(sf)
}
\arguments{
\item{sf}{Simple features object (a simple features object)}
}
\value{
Valid (a simple features object)
}
\description{
make_valid_new_sf() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make valid new simple features object. The function returns Valid (a simple features object).
}
\keyword{internal}
