% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_year_filter_logic_vec}
\alias{make_year_filter_logic_vec}
\title{Make year filter logic vec}
\usage{
make_year_filter_logic_vec(data_tb, included_years_vec)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{included_years_vec}{PARAM_DESCRIPTION}
}
\description{
make_year_filter_logic_vec() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make year filter logic vec. The function is called for its side effects and does not return a value.
}
\keyword{internal}
