% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get_sp_data.R
\name{make_year_vec}
\alias{make_year_vec}
\title{make_year_vec}
\usage{
make_year_vec(input_ls)
}
\arguments{
\item{input_ls}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready4s4]{lookup_tb}},\code{\link[ready4s4]{sp_data_pack_lup}}
 \code{\link[lubridate]{year}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{pull}}
 \code{\link[stringr]{str_length}}
 \code{\link[purrr]{pluck}}
}
