% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_nbr_bands.R
\name{nbr_bands<-}
\alias{nbr_bands<-}
\alias{nbr_bands<--ready4_profiled_area}
\title{nbr_bands<-}
\usage{
nbr_bands(x) <- value

\S4method{nbr_bands}{ready4_profiled_area}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_profiled_area}
}
\description{
S4 Generic function to set the value of the slot nbr_bands

Set the value of the slot nbr_bands for S4 objects of class ready4_profiled_area
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
