% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_profiled_area.R, R/gnrc_nbr_bands.R
\name{nbr_bands-ready4_profiled_area}
\alias{nbr_bands-ready4_profiled_area}
\alias{nbr_bands}
\title{nbr_bands}
\arguments{
\item{x}{An object of class ready4_profiled_area}

\item{x}{An object}
}
\description{
Get the value of the slot nbr_bands for S4 objects of class ready4_profiled_area

S4 Generic function to get the value of the slot nbr_bands
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
