% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make_isochrs.R
\name{make_time_band_sf_ls}
\alias{make_time_band_sf_ls}
\title{make_time_band_sf_ls}
\usage{
make_time_band_sf_ls(look_up_ref, one_cluster_travel_time_sf_list)
}
\arguments{
\item{look_up_ref}{PARAM_DESCRIPTION}

\item{one_cluster_travel_time_sf_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{pluck}},\code{\link[purrr]{map}}
 \code{\link[dplyr]{pull}},\code{\link[dplyr]{filter}}
 \code{\link[stats]{setNames}}
 \code{\link[stringr]{str_replace}}
}
