% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_order_tb.R
\name{order_tb.ready4_sp_import_lup}
\alias{order_tb.ready4_sp_import_lup}
\alias{order_tb,ready4_sp_import_lup-method}
\title{Order tibble method applied toeadyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import..}
\usage{
order_tb.ready4_sp_import_lup(x)

\S4method{order_tb}{ready4_sp_import_lup}(x)
}
\arguments{
\item{x}{An instance of Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.}
}
\description{
order_tb.ready4_sp_import_lup() is an Order Tibble method that orders a tibble. This method is implemented for the Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.. The function is called for its side effects and does not return a value.
}
