% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_pack.R
\name{order_tb_ready4_sp_import_lup}
\alias{order_tb_ready4_sp_import_lup}
\title{order_tb_ready4_sp_import_lup}
\usage{
order_tb_ready4_sp_import_lup(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{pull}}
 \code{\link[purrr]{map}},\code{\link[purrr]{map2}},\code{\link[purrr]{reduce}}
}
