% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_ready4_sp_local.R,
%   R/s4_ready4_sp_local_proc.R, R/s4_ready4_sp_local_raw.R
\name{overwrite_lgl<--ready4_sp_local}
\alias{overwrite_lgl<--ready4_sp_local}
\alias{overwrite_lgl<--ready4_sp_local_proc}
\alias{overwrite_lgl<--ready4_sp_local_raw}
\title{overwrite_lgl<-}
\arguments{
\item{x}{An object of class ready4_sp_local}

\item{x}{An object of class ready4_sp_local_proc}

\item{x}{An object of class ready4_sp_local_raw}
}
\description{
Set the value of the slot overwrite_lgl for S4 objects of class ready4_sp_local

Set the value of the slot overwrite_lgl for S4 objects of class ready4_sp_local_proc

Set the value of the slot overwrite_lgl for S4 objects of class ready4_sp_local_raw
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
