% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_env.R, R/gnrc_par_vals.R
\name{par_vals<--ready4_env}
\alias{par_vals<--ready4_env}
\alias{par_vals<-}
\title{par_vals<-}
\arguments{
\item{x}{An object of class ready4_env}

\item{x}{An object}
}
\description{
Set the value of the slot par_vals for S4 objects of class ready4_env

S4 Generic function to set the value of the slot par_vals
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
