% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_parameter_name.R, R/s4_ready4_struc.R
\name{parameter_name}
\alias{parameter_name}
\alias{parameter_name-ready4_struc}
\title{parameter_name}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_struc}
}
\description{
S4 Generic function to get the value of the slot parameter_name

Get the value of the slot parameter_name for S4 objects of class ready4_struc
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
