% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_ready4_sp_local_proc.R
\name{path_to_starter_sf_chr<--ready4_sp_local_proc}
\alias{path_to_starter_sf_chr<--ready4_sp_local_proc}
\title{path_to_starter_sf_chr<-}
\arguments{
\item{x}{An object of class ready4_sp_local_proc}
}
\description{
Set the value of the slot path_to_starter_sf_chr for S4 objects of class ready4_sp_local_proc
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
