% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prevalence.R
\name{pick_rate_from_source}
\alias{pick_rate_from_source}
\title{pick_rate_from_source
FUNCTION_DESCRIPTION}
\usage{
pick_rate_from_source(disorder, period, source, age, sex,
  prev_rates = safety_prev_rates)
}
\arguments{
\item{disorder}{PARAM_DESCRIPTION}

\item{period}{PARAM_DESCRIPTION}

\item{source}{PARAM_DESCRIPTION}

\item{age}{PARAM_DESCRIPTION}

\item{sex}{PARAM_DESCRIPTION}

\item{prev_rates}{PARAM_DESCRIPTION, Default: safety_prev_rates}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
pick_rate_from_source
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
