% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prevalence.R
\name{pick_source}
\alias{pick_source}
\title{pick_source
FUNCTION_DESCRIPTION}
\usage{
pick_source(disorder, period, age, sex, pref_source = safety_pref_source)
}
\arguments{
\item{disorder}{PARAM_DESCRIPTION}

\item{period}{PARAM_DESCRIPTION}

\item{age}{PARAM_DESCRIPTION}

\item{sex}{PARAM_DESCRIPTION}

\item{pref_source}{PARAM_DESCRIPTION, Default: safety_pref_source}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
pick_source
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
