% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_pop_sum_15.24.R
\name{pop_sum_15.24}
\alias{pop_sum_15.24}
\title{Summarise the 15-24 populaion by region for a given year.}
\usage{
pop_sum_15.24(data, year)
}
\arguments{
\item{data}{A tibble with population data summarised by age, sex and region.}

\item{first_colref}{A positive integer.}
}
\value{
A tibble.
}
\description{
This function creates a summary tibble of the female and male 15-24 year old
populations for each inputted region.
}
\details{

}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{sym}}
 \code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{reexports}}
 \code{\link[sf]{st_geometry}}

Other spatial functions.: \code{\link{spatial_adjust_partial_in_custom_area}},
  \code{\link{spatial_adjust_population_by_included_fraction}},
  \code{\link{spatial_cases_coordinates}},
  \code{\link{spatial_create_agent_coordinates_tb}},
  \code{\link{spatial_merge_areas_attributes}},
  \code{\link{spatial_one_year_projections_15_24}},
  \code{\link{spatial_population_growth}},
  \code{\link{spatial_profile_by_sa2}},
  \code{\link{spatial_select_rename_age_sex}},
  \code{\link{spatial_summarise_population_counts_0to29}},
  \code{\link{spatial_vic_pop_growth_lga}}
}
\concept{spatial functions.}
