% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prevalence.R
\name{pref_sources_for_age_range}
\alias{pref_sources_for_age_range}
\title{pref_sources_for_age_rang
FUNCTION_DESCRIPTION}
\usage{
pref_sources_for_age_range(disorder, period, ages, sexes, pref_source)
}
\arguments{
\item{disorder}{PARAM_DESCRIPTION}

\item{period}{PARAM_DESCRIPTION}

\item{ages}{PARAM_DESCRIPTION}

\item{sexes}{PARAM_DESCRIPTION}

\item{pref_source}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
pref_sources_for_age_rang
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
