% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_sp_local_proc.R
\name{proc_data_dir_chr<--ready4_sp_local_proc}
\alias{proc_data_dir_chr<--ready4_sp_local_proc}
\title{proc_data_dir_chr<-}
\arguments{
\item{x}{An object of class ready4_sp_local_proc}
}
\description{
Set the value of the slot proc_data_dir_chr for S4 objects of class ready4_sp_local_proc
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
