% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_data_pack.R
\name{process_import_xx}
\alias{process_import_xx}
\title{process_import_xx}
\usage{
process_import_xx(
  x,
  import_this_ls,
  path_to_starter_sf_chr,
  merge_with,
  processed_dir,
  crs_nbr_vec = NA_real_,
  overwrite_lgl = F
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{import_this_ls}{PARAM_DESCRIPTION}

\item{path_to_starter_sf_chr}{PARAM_DESCRIPTION}

\item{merge_with}{PARAM_DESCRIPTION}

\item{processed_dir}{PARAM_DESCRIPTION}

\item{crs_nbr_vec}{PARAM_DESCRIPTION, Default: NA}

\item{overwrite_lgl}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{pull}}
 \code{\link[purrr]{map2}}
}
