% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_procure.R
\name{procure.vicinity_abbreviations}
\alias{procure.vicinity_abbreviations}
\alias{procure,vicinity_abbreviations-method}
\alias{procure.vicinity_identifiers}
\alias{procure,vicinity_identifiers-method}
\alias{procure.vicinity_raw}
\alias{procure,vicinity_raw-method}
\alias{procure.vicinity_resolutions}
\alias{procure,vicinity_resolutions-method}
\alias{procure-VicinityProfile}
\alias{procure,VicinityProfile-method}
\title{Procure items from a dataset}
\usage{
procure.vicinity_abbreviations(
  x,
  col_nm_1L_chr = "short_name_chr",
  match_value_xx
)

\S4method{procure}{vicinity_abbreviations}(x, col_nm_1L_chr = "short_name_chr", match_value_xx)

procure.vicinity_identifiers(
  x,
  col_nm_1L_chr = "spatial_unit_chr",
  geometry_rsl_1L_chr = character(0),
  group_at_geom_unit_1L_lgl = TRUE,
  data_rsl_1L_chr = character(0),
  match_value_xx = NULL,
  area_bndy_yr_chr = character(0),
  what_1L_chr = "match"
)

\S4method{procure}{vicinity_identifiers}(
  x,
  col_nm_1L_chr = "spatial_unit_chr",
  geometry_rsl_1L_chr = character(0),
  group_at_geom_unit_1L_lgl = TRUE,
  data_rsl_1L_chr = character(0),
  match_value_xx = NULL,
  area_bndy_yr_chr = character(0),
  what_1L_chr = "match"
)

procure.vicinity_raw(
  x,
  inc_script_1L_lgl = T,
  forced_choice_chr = NA_character_,
  match_value_xx = NULL,
  what_1L_chr = "source"
)

\S4method{procure}{vicinity_raw}(
  x,
  inc_script_1L_lgl = T,
  forced_choice_chr = NA_character_,
  match_value_xx = NULL,
  what_1L_chr = "source"
)

procure.vicinity_resolutions(
  x,
  options_chr = character(0),
  what_1L_chr = "resolution",
  whole_area_1L_lgl = TRUE,
  year_1L_dbl = numeric(0)
)

\S4method{procure}{vicinity_resolutions}(
  x,
  options_chr = character(0),
  what_1L_chr = "resolution",
  whole_area_1L_lgl = TRUE,
  year_1L_dbl = numeric(0)
)

\S4method{procure}{VicinityProfile}(
  x,
  exclude_dif_bndy_yr_1L_lgl = TRUE,
  highest_rsl_chr = character(0),
  key_var_1L_chr = character(0),
  match_year_1L_lgl = TRUE,
  travel_mode_1L_chr = character(0),
  years_chr = character(0),
  what_1L_chr = "grouping"
)
}
\arguments{
\item{x}{An object of class VicinityProfile}

\item{col_nm_1L_chr}{Column name (a character vector of length one), Default: 'spatial_unit_chr'}

\item{match_value_xx}{Match value (an output object of multiple potential types), Default: NULL}

\item{geometry_rsl_1L_chr}{Geometry resolution (a character vector of length one), Default: character(0)}

\item{group_at_geom_unit_1L_lgl}{Group at geometry unit (a logical vector of length one), Default: TRUE}

\item{data_rsl_1L_chr}{Data resolution (a character vector of length one), Default: character(0)}

\item{area_bndy_yr_chr}{Area boundary year (a character vector), Default: character(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: 'grouping'}

\item{inc_script_1L_lgl}{Include script (a logical vector of length one), Default: T}

\item{forced_choice_chr}{Forced choice (a character vector), Default: 'NA'}

\item{options_chr}{Options (a character vector), Default: character(0)}

\item{whole_area_1L_lgl}{Whole area (a logical vector of length one), Default: TRUE}

\item{year_1L_dbl}{Year (a double vector of length one), Default: numeric(0)}

\item{exclude_dif_bndy_yr_1L_lgl}{Exclude different boundary year (a logical vector of length one), Default: TRUE}

\item{highest_rsl_chr}{Highest resolution (a character vector), Default: character(0)}

\item{key_var_1L_chr}{Key variable (a character vector of length one), Default: character(0)}

\item{match_year_1L_lgl}{Match year (a logical vector of length one), Default: TRUE}

\item{travel_mode_1L_chr}{Travel mode (a character vector of length one), Default: character(0)}

\item{years_chr}{Years (a character vector), Default: character(0)}
}
\value{
Object (an output object of multiple potential types)

Object (an output object of multiple potential types)

Object (an output object of multiple potential types)

Resolution (an output object of multiple potential types)

Object (an output object of multiple potential types)
}
\description{
procure.vicinity_abbreviations() is a procure method that procures data by executing a search and retrieval algorithm using data contained in an instance of a class. This method is implemented for the ready4 submodule class for tibble object lookup table for spatial data abbreviations. The function returns Object (an output object of multiple potential types).

procure.vicinity_identifiers() is a procure method that procures data by executing a search and retrieval algorithm using data contained in an instance of a class. This method is implemented for the ready4 submodule class for tibble object lookup table of unique feature identifiers used for different spatial objects. The function returns Object (an output object of multiple potential types).

procure.vicinity_raw() is a procure method that procures data by executing a search and retrieval algorithm using data contained in an instance of a class. This method is implemented for the ready4 submodule class for tibble object lookup table of metadata about raw (un-processed) spatial data to import. The function returns Object (an output object of multiple potential types).

procure.vicinity_resolutions() is a procure method that procures data by executing a search and retrieval algorithm using data contained in an instance of a class. This method is implemented for the ready4 submodule class for tibble object lookup table of the relative resolutions of different spatial objects. The function returns Resolution (an output object of multiple potential types).

procure method applied to VicinityProfile
}
