% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_epi_for_area.R
\name{project_epi_for_area}
\alias{project_epi_for_area}
\title{project_epi_for_area}
\usage{
project_epi_for_area(state_territory = "Victoria",
  profiled_area = "Service cluster - Orygen headspaces",
  age_lower = 12, age_upper = 18, project_for_year = 2023)
}
\arguments{
\item{state_territory}{PARAM_DESCRIPTION, Default: 'Victoria'}

\item{profiled_area}{PARAM_DESCRIPTION, Default: 'Service cluster - Orygen headspaces'}

\item{age_lower}{PARAM_DESCRIPTION, Default: 12}

\item{age_upper}{PARAM_DESCRIPTION, Default: 18}

\item{project_for_year}{PARAM_DESCRIPTION, Default: 2023}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Project epidemiology for area at specified point in time
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready.agents]{character(0)}}
 \code{\link[ready.aus.data]{character(0)}},\code{\link[ready.aus.data]{aus_boundary_phns_sf}}
 \code{\link[dplyr]{summarise_all}},\code{\link[dplyr]{funs}},\code{\link[dplyr]{filter}}
 \code{\link[sf]{geos_combine}}
 \code{\link[tibble]{tibble}}
 \code{\link[ready.sim]{ready_env}},\code{\link[ready.sim]{ready_sim_data}},\code{\link[ready.sim]{runSimulation}},\code{\link[ready.sim]{env_sf}},\code{\link[ready.sim]{st_envir}}
}
