% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suburb_based_effective_catchment.R
\name{proper}
\alias{proper}
\title{proper}
\usage{
proper(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
