% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_randomise.R
\name{randomise_locations}
\alias{randomise_locations}
\title{Randomise locations}
\usage{
randomise_locations(profiled_sf, cases_int, crs_dbl = 4326)
}
\arguments{
\item{profiled_sf}{Profiled (a simple features object)}

\item{cases_int}{Cases (an integer vector)}

\item{crs_dbl}{Coordinates reference system (a double vector), Default: 4326}
}
\value{
Sample patient locations (an output object of multiple potential types)
}
\description{
randomise_locations() is a Randomise function that randomly samples from data. Specifically, this function implements an algorithm to randomise locations. The function returns Sample patient locations (an output object of multiple potential types).
}
\keyword{internal}
