% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_ratify.R
\name{ratify.vicinity_processed}
\alias{ratify.vicinity_processed}
\alias{ratify,vicinity_processed-method}
\title{Ratify that input or output data meet validity criteria}
\usage{
ratify.vicinity_processed(
  x,
  data_items_chr = character(0),
  key_var_1L_chr = character(0),
  what_1L_chr = "population"
)

\S4method{ratify}{vicinity_processed}(
  x,
  data_items_chr = character(0),
  key_var_1L_chr = character(0),
  what_1L_chr = "population"
)
}
\arguments{
\item{x}{An instance of \code{vicinity_processed}, a ready4 submodule class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).}

\item{data_items_chr}{Data items (a character vector), Default: character(0)}

\item{key_var_1L_chr}{Key variable (a character vector of length one), Default: character(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: 'population'}
}
\value{
Ratified (a logical vector of length one)
}
\description{
ratify.vicinity_processed() is a ratify method that ratifies that an instance of a class conforms to specified criteria. This method is implemented for the ready4 submodule class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data). The function returns Ratified (a logical vector of length one).
}
