% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rready_space.R
\docType{package}
\name{ready.space-package}
\alias{ready.space}
\alias{ready.space-package}
\title{ready.space: Package of functions for modelling in space and time with the readfyforwhatsnext suite of resilience packages.}
\description{
ready.space is a package of functions that creates the spatial and temporal environment for the readyforwhatsnext simulation toolkit.
}
\details{
It has three main goals:
\itemize{
\item Goal 1.
\item Goal 2
\item Goal 3.
}
To learn more about ready.space, start with the vignettes:
`browseVignettes(package = "ready.space")`
}
\section{First section}{

\describe{
\item{firstitem}
}
}

\section{Second section}{

Some text.
\describe{
\item{moretext}
}
}

\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au}

}
