% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_ready4_env.R
\name{ready4_env}
\alias{ready4_env}
\title{ready4_env}
\usage{
ready4_env(st_data = list(list()), env_sf = sf::st_sf(sf::st_sfc()),
  par_vals = tibble::tibble())
}
\arguments{
\item{st_data}{list, Default: list(list())}

\item{env_sf}{sf, Default: sf::st_sf(sf::st_sfc())}

\item{par_vals}{tbl_df, Default: tibble::tibble()}
}
\value{
An S4 object of the ready4_env class
}
\description{
An S4 class to represent Spatiotemporal environment

Create a new S4 object of the class:ready4_env
}
\section{Slots}{

\describe{
\item{\code{st_data}}{list}

\item{\code{env_sf}}{sf}

\item{\code{par_vals}}{tbl_df}
}}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{sf}},\code{\link[sf]{sfc}}
 \code{\link[tibble]{tibble}}
 \code{\link[methods]{new}}
}
