% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_macro.R
\docType{class}
\name{ready4_macro}
\alias{ready4_macro}
\title{ready4_macro}
\description{
An S4 class to represent Macro level context
}
\section{Slots}{

\describe{
\item{\code{global_region}}{character}

\item{\code{country}}{character}

\item{\code{country_bound_year}}{numeric}

\item{\code{lookup_tb}}{ready4_lookup}

\item{\code{crs_nbr}}{numeric}

\item{\code{temporal_min}}{POSIXt}

\item{\code{temporal_max}}{POSIXt}
}}

