% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_meso_area.R
\docType{class}
\name{ready4_meso_area}
\alias{ready4_meso_area}
\title{ready4_meso_area}
\description{
An S4 class to represent Meso level context - area
}
\section{Slots}{

\describe{
\item{\code{area_type}}{character}

\item{\code{area}}{character}

\item{\code{area_bound_year}}{numeric}

\item{\code{region_type}}{character}

\item{\code{region}}{character}

\item{\code{region_bound_year}}{numeric}

\item{\code{global_region}}{character}

\item{\code{country}}{character}

\item{\code{country_bound_year}}{numeric}

\item{\code{lookup_tb}}{ready4_lookup}

\item{\code{crs_nbr}}{numeric}

\item{\code{temporal_min}}{POSIXt}

\item{\code{temporal_max}}{POSIXt}
}}

