% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_meso_region.R
\docType{class}
\name{ready4_meso_region}
\alias{ready4_meso_region}
\title{ready4_meso_region}
\usage{
ready4_meso_region(
  global_region = NA_character_,
  country = NA_character_,
  country_bound_year = NA_real_,
  lookup_tb = ready4_lookup(),
  crs_nbr = NA_real_,
  temporal_min = .POSIXct(NA_character_),
  temporal_max = .POSIXct(NA_character_),
  region_type = NA_character_,
  region = NA_character_,
  region_bound_year = NA_real_
)
}
\arguments{
\item{global_region}{character, Default: 'NA'}

\item{country}{character, Default: 'NA'}

\item{country_bound_year}{numeric, Default: NA}

\item{lookup_tb}{ready4_lookup, Default: ready4_lookup()}

\item{crs_nbr}{numeric, Default: NA}

\item{temporal_min}{POSIXt, Default: .POSIXct(NA_character_)}

\item{temporal_max}{POSIXt, Default: .POSIXct(NA_character_)}

\item{region_type}{character, Default: 'NA'}

\item{region}{PARAM_DESCRIPTION, Default: 'NA'}

\item{region_bound_year}{numeric, Default: NA}
}
\value{
An S4 object of the ready4_meso_region class
}
\description{
An S4 class to represent Meso level context - region

Create a new S4 object of the class:ready4_meso_region
}
\section{Slots}{

\describe{
\item{\code{region_type}}{character}

\item{\code{region}}{character}

\item{\code{region_bound_year}}{numeric}

\item{\code{global_region}}{character}

\item{\code{country}}{character}

\item{\code{country_bound_year}}{numeric}

\item{\code{lookup_tb}}{ready4_lookup}

\item{\code{crs_nbr}}{numeric}

\item{\code{temporal_min}}{POSIXt}

\item{\code{temporal_max}}{POSIXt}
}}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[methods]{new}}
}
