% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_micro.R
\docType{class}
\name{ready4_micro}
\alias{ready4_micro}
\title{ready4_micro}
\description{
An S4 class to represent Micro level context
}
\section{Slots}{

\describe{
\item{\code{geom_dist_km_cuts}}{numeric}

\item{\code{area_type}}{character}

\item{\code{region_type}}{character}

\item{\code{global_region}}{character}
}}

