% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_ready4_par_struc_mape.R
\name{ready4_par_struc_mape}
\alias{ready4_par_struc_mape}
\alias{new_ready4_par_struc_mape}
\alias{make_prototype_ready4_par_struc_mape}
\alias{validate_ready4_par_struc_mape}
\alias{is_ready4_par_struc_mape}
\title{ready4_par_struc_mape}
\usage{
ready4_par_struc_mape(x = make_prototype_ready4_par_struc_mape())

new_ready4_par_struc_mape(x)

make_prototype_ready4_par_struc_mape()

validate_ready4_par_struc_mape(x)

is_ready4_par_struc_mape(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION, Default: make_prototype_ready4_par_struc_mape()}

\item{x}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}
}
\value{
A validated instance of the ready4_par_struc_mape class

An unvalidated instance of the ready4_par_struc_mape class

A prototpe for ready4_par_struc_mape class

A prototpe for ready4_par_struc_mape class

A logical value, TRUE if a valid instance of the ready4_par_struc_mape class
}
\description{
Create a new valid instance of the S3 class: ready4_par_struc_mape

Create a new unvalidated instance of the S3 class: new_ready4_par_struc_mape

Create a new prototype for S3 class: make_prototype_ready4_par_struc_mape

Validate an instance of the S3 class: validate_ready4_par_struc_mape

Check whether an object is a valid instance of the S3 class: is_ready4_par_struc_mape
}
\details{
Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{is_tibble}}

\code{\link[tibble]{tibble}}

\code{\link[stringr]{str_detect}},\code{\link[stringr]{str_c}}
 \code{\link[dplyr]{summarise_all}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{pull}}
 \code{\link[tidyr]{gather}}
 \code{\link[purrr]{map2}}
}
