% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_par_str_envir.R
\name{ready4_param_str_envir}
\alias{ready4_param_str_envir}
\title{Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to the spatial environment.}
\usage{
ready4_param_str_envir(x = make_pt_ready4_param_str_envir())
}
\arguments{
\item{x}{A prototype for the Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to the spatial environment., Default: make_pt_ready4_param_str_envir()}
}
\value{
A validated instance of the Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to the spatial environment.
}
\description{
Create a new valid instance of the Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to the spatial environment.
}
\details{
Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to the spatial environment.
}
