% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_ready4_profiled_area.R
\name{ready4_profiled_area}
\alias{ready4_profiled_area}
\title{ready4_profiled_area}
\usage{
ready4_profiled_area(country = NA_character_,
  area_type = NA_character_, area_bound_year = NA_real_,
  features = NA_character_, use_coord_lup = NA,
  lookup_tb = ready4_lookup(), crs_nbr = NA_real_,
  geom_dist_limit_km = NA_real_, drive_time_limit_mins = NA_real_,
  nbr_bands = NA_real_, data_year = NA_character_,
  data_ymds = .POSIXct(NA_character_))
}
\arguments{
\item{country}{character, Default: 'NA'}

\item{area_type}{character, Default: 'NA'}

\item{area_bound_year}{numeric, Default: NA}

\item{features}{character, Default: 'NA'}

\item{use_coord_lup}{logical, Default: NA}

\item{lookup_tb}{ready4_lookup, Default: ready4_lookup()}

\item{crs_nbr}{numeric, Default: NA}

\item{geom_dist_limit_km}{numeric, Default: NA}

\item{drive_time_limit_mins}{numeric, Default: NA}

\item{nbr_bands}{numeric, Default: NA}

\item{data_year}{character, Default: 'NA'}

\item{data_ymds}{POSIXt, Default: .POSIXct(NA_character_)}
}
\value{
An S4 object of the ready4_profiled_area class
}
\description{
An S4 class to represent Information to create a profiled area object

Create a new S4 object of the class:ready4_profiled_area
}
\section{Slots}{

\describe{
\item{\code{country}}{character}

\item{\code{area_type}}{character}

\item{\code{area_bound_year}}{numeric}

\item{\code{features}}{character}

\item{\code{use_coord_lup}}{logical}

\item{\code{lookup_tb}}{ready4_lookup}

\item{\code{crs_nbr}}{numeric}

\item{\code{geom_dist_limit_km}}{numeric}

\item{\code{drive_time_limit_mins}}{numeric}

\item{\code{nbr_bands}}{numeric}

\item{\code{data_year}}{character}

\item{\code{data_ymds}}{POSIXt}
}}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[methods]{new}}
}
