% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_abbreviations_lup.R
\name{ready4_sp_abbreviations_lup}
\alias{ready4_sp_abbreviations_lup}
\title{Readyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations.}
\usage{
ready4_sp_abbreviations_lup(x = make_pt_ready4_sp_abbreviations_lup())
}
\arguments{
\item{x}{A prototype for the Readyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations., Default: make_pt_ready4_sp_abbreviations_lup()}
}
\value{
A validated instance of the Readyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations.
}
\description{
Create a new valid instance of the Readyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations.
}
