% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_data_pack_lup.R
\name{ready4_sp_data_pack_lup}
\alias{ready4_sp_data_pack_lup}
\title{Readyforwhatsnext S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).}
\usage{
ready4_sp_data_pack_lup(x = make_pt_ready4_sp_data_pack_lup())
}
\arguments{
\item{x}{A prototype for the Readyforwhatsnext S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data)., Default: make_pt_ready4_sp_data_pack_lup()}
}
\value{
A validated instance of the Readyforwhatsnext S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).
}
\description{
Create a new valid instance of the Readyforwhatsnext S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).
}
