% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_import_lup.R
\name{ready4_sp_import_lup}
\alias{ready4_sp_import_lup}
\title{Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.}
\usage{
ready4_sp_import_lup(x = make_pt_ready4_sp_import_lup())
}
\arguments{
\item{x}{A prototype for the Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import., Default: make_pt_ready4_sp_import_lup()}
}
\value{
A validated instance of the Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.
}
\description{
Create a new valid instance of the Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.
}
