% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_sp_local.R
\docType{class}
\name{ready4_sp_local}
\alias{ready4_sp_local}
\title{ready4_sp_local}
\description{
An S4 class to represent Object defining data to be saved in local directory.
}
\section{Slots}{

\describe{
\item{\code{lup_tbs_r4}}{ready4_lookup}

\item{\code{merge_itms_chr}}{character}
}}

