% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_sp_local_raw.R
\docType{class}
\name{ready4_sp_local_raw}
\alias{ready4_sp_local_raw}
\title{ready4_sp_local_raw}
\description{
An S4 class to represent Object defining data to be saved in local directory in a raw (unprocessed) format.
}
\section{Slots}{

\describe{
\item{\code{lup_tbs_r4}}{ready4_lookup}

\item{\code{save_type}}{character}

\item{\code{merge_with_chr_vec}}{character}
}}

