% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_resolution_lup.R
\name{ready4_sp_resolution_lup}
\alias{ready4_sp_resolution_lup}
\title{Readyforwhatsnext S3 class for tibble object lookup table of the relative resolutions of different spatial objects.}
\usage{
ready4_sp_resolution_lup(x = make_pt_ready4_sp_resolution_lup())
}
\arguments{
\item{x}{A prototype for the Readyforwhatsnext S3 class for tibble object lookup table of the relative resolutions of different spatial objects., Default: make_pt_ready4_sp_resolution_lup()}
}
\value{
A validated instance of the Readyforwhatsnext S3 class for tibble object lookup table of the relative resolutions of different spatial objects.
}
\description{
Create a new valid instance of the Readyforwhatsnext S3 class for tibble object lookup table of the relative resolutions of different spatial objects.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of the relative resolutions of different spatial objects.
}
