% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_site_coord_lup.R
\name{ready4_sp_site_coord_lup}
\alias{ready4_sp_site_coord_lup}
\title{Readyforwhatsnext S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.}
\usage{
ready4_sp_site_coord_lup(x = make_pt_ready4_sp_site_coord_lup())
}
\arguments{
\item{x}{A prototype for the Readyforwhatsnext S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes., Default: make_pt_ready4_sp_site_coord_lup()}
}
\value{
A validated instance of the Readyforwhatsnext S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.
}
\description{
Create a new valid instance of the Readyforwhatsnext S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.
}
