% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_starter_sf_lup.R
\name{ready4_sp_starter_sf_lup}
\alias{ready4_sp_starter_sf_lup}
\title{Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.}
\usage{
ready4_sp_starter_sf_lup(x = make_pt_ready4_sp_starter_sf_lup())
}
\arguments{
\item{x}{A prototype for the Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects., Default: make_pt_ready4_sp_starter_sf_lup()}
}
\value{
A validated instance of the Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
\description{
Create a new valid instance of the Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
