% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_sp_local_proc.R
\docType{class}
\name{ready4_spProcessed}
\alias{ready4_spProcessed}
\title{ready4_spProcessed}
\description{
An S4 class to represent Object defining data to be saved in local directory in a processed (R) format.
}
\section{Slots}{

\describe{
\item{\code{lup_tbs_r4}}{ready4_lookup}

\item{\code{write_type_1L_chr}}{character}

\item{\code{merge_itms_chr}}{character}
}}

