% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_ready4_struc.R
\name{ready4_struc}
\alias{ready4_struc}
\title{ready4_struc}
\usage{
ready4_struc(parameter_name = NA_character_)
}
\arguments{
\item{parameter_name}{character, Default: 'NA'}
}
\value{
An S4 object of the ready4_struc class
}
\description{
An S4 class to represent Structural parameter probability distributions and transformations

Create a new S4 object of the class:ready4_struc
}
\section{Slots}{

\describe{
\item{\code{parameter_name}}{character}
}}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[methods]{new}}
}
