% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pckg_ready4space.R
\docType{package}
\name{ready4space-package}
\alias{ready4space}
\alias{ready4space-package}
\title{ready4space: Readyforwhatsnext Spatio-temporal Data Processing.}
\description{
ready4space is a package of functions that creates the spatial and temporal environment for the readyforwhatsnext simulation toolkit.
}
\details{
To learn more about ready4space, start with the vignettes:
`browseVignettes(package = "ready4space")`
}
\section{First section}{

Some text.
}

\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Other contributors:
\itemize{
  \item Glen Wiesner \email{Glen.Wiesner@vu.edu.au} (\href{https://orcid.org/0000-0002-0071-130X}{ORCID}) [reviewer]
  \item Alexandra Parker \email{Alex.Parker@vu.edu.au} (\href{https://orcid.org/0000-0002-2398-6306}{ORCID}) [reviewer]
  \item Cathrine Mihalopoulos \email{cathy.mihalopoulos@deakin.edu.au} (\href{https://orcid.org/0000-0002-7127-9462}{ORCID}) [reviewer]
  \item Jonathan Karnon \email{Jonathan.Karnon@flinders.edu.au} (\href{https://orcid.org/0000-0003-3220-2099}{ORCID}) [reviewer]
  \item Petra Plencnerova \email{Petra.Plencnerova@vu.edu.au} (\href{https://orcid.org/0000-0001-9698-9084}{ORCID}) [reviewer]
  \item Orygen [copyright holder, funder]
  \item VicHealth [funder]
  \item Victoria University [funder]
}

}
