% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4space_values.R
\name{ready4space_values}
\alias{ready4space_values}
\alias{make_pt_ready4space_values}
\alias{is_ready4space_values}
\title{ready4 S3 class for tibble object that stores simulation parameter values for each iteration.}
\usage{
ready4space_values(x = make_pt_ready4space_values())

make_pt_ready4space_values(
  param_name_chr = character(0),
  iteration_1_dbl = numeric(0)
)

is_ready4space_values(x)
}
\arguments{
\item{x}{An object of any type}

\item{param_name_chr}{Parameter name (a character vector), Default: character(0)}

\item{iteration_1_dbl}{Iteration 1 (a double vector), Default: numeric(0)}
}
\value{
A validated instance of the ready4 S3 class for tibble object that stores simulation parameter values for each iteration.

A prototype for ready4 S3 class for tibble object that stores simulation parameter values for each iteration.

A logical value, TRUE if a valid instance of the ready4 S3 class for tibble object that stores simulation parameter values for each iteration.
}
\description{
Create a new valid instance of the ready4 S3 class for tibble object that stores simulation parameter values for each iteration.
}
\details{
ready4 S3 class for tibble object that stores simulation parameter values for each iteration.
}
