% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4spaceenv.R
\docType{class}
\name{Ready4spaceenv-class}
\alias{Ready4spaceenv-class}
\alias{Ready4spaceenv}
\title{Ready4spaceenv}
\description{
Spatiotemporal environment
}
\section{Slots}{

\describe{
\item{\code{data_ls}}{Data (a list)}

\item{\code{env_sf}}{Environment (a simple features object)}

\item{\code{param_vals_tb}}{Parameter values (a tibble)}
}}

