% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_reckon.R
\name{reckon.vicinity_mapes}
\alias{reckon.vicinity_mapes}
\alias{reckon,vicinity_mapes-method}
\alias{reckon.vicinity_parameters}
\alias{reckon,vicinity_parameters-method}
\title{Reckon (calculate) a value}
\usage{
reckon.vicinity_mapes(
  x,
  iter_1L_int = 1,
  n_its_int = 1,
  joint_dstr_1L_lgl = T,
  type_1L_chr = "outer"
)

\S4method{reckon}{vicinity_mapes}(
  x,
  iter_1L_int = 1,
  n_its_int = 1,
  joint_dstr_1L_lgl = T,
  type_1L_chr = "outer"
)

reckon.vicinity_parameters(
  x,
  iter_1L_int = integer(0),
  n_its_int = integer(0),
  type_1L_chr = "outer"
)

\S4method{reckon}{vicinity_parameters}(
  x,
  iter_1L_int = integer(0),
  n_its_int = integer(0),
  type_1L_chr = "outer"
)
}
\arguments{
\item{x}{An instance of \code{vicinity_parameters}, a ready4 submodule class for tibble object that stores simulation structural parameters relating to the spatial environment.}

\item{iter_1L_int}{Iteration (an integer vector of length one), Default: integer(0)}

\item{n_its_int}{N its (an integer vector), Default: integer(0)}

\item{joint_dstr_1L_lgl}{Joint distribution (a logical vector of length one), Default: T}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'outer'}
}
\value{
Mean absolute prediction error (an output object of multiple potential types)

Value (an output object of multiple potential types)
}
\description{
reckon.vicinity_mapes() is a reckon method that reckons a value by performing a calculation using data contained in an instance of a class. This method is implemented for the ready4 submodule class for tibble object that stores spatial simulation parameters relating to Mean Absolute Prediction Errors. The function returns Mean absolute prediction error (an output object of multiple potential types).

reckon.vicinity_parameters() is a reckon method that reckons a value by performing a calculation using data contained in an instance of a class. This method is implemented for the ready4 submodule class for tibble object that stores simulation structural parameters relating to the spatial environment. The function returns Value (an output object of multiple potential types).
}
