% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suburb_based_effective_catchment.R
\name{reconcile_suburb_names}
\alias{reconcile_suburb_names}
\title{reconcile_suburb_names}
\usage{
reconcile_suburb_names(raw_name)
}
\arguments{
\item{raw_name}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_replace}}
}
