% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_region.R
\name{region}
\alias{region}
\alias{region-ready4_meso_region}
\title{region}
\usage{
region(x)

\S4method{region}{ready4_meso_region}(x)
}
\arguments{
\item{x}{An object of class ready4_meso_region}
}
\description{
S4 Generic function to get the value of the slot region

Get the value of the slot region for S4 objects of class ready4_meso_region
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
