% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_region_bound_year.R
\name{region_bound_year}
\alias{region_bound_year}
\alias{region_bound_year-ready4_meso_region}
\alias{region_bound_year,ready4_meso_region-method}
\title{region_bound_year}
\usage{
region_bound_year(x)

\S4method{region_bound_year}{ready4_meso_region}(x)
}
\arguments{
\item{x}{An object of class ready4_meso_region}
}
\description{
S4 Generic function to get the value of the slot region_bound_year

Get the value of the slot region_bound_year for S4 objects of class ready4_meso_region
}
