% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_region.R
\name{region<-}
\alias{region<-}
\alias{region<--ready4_meso_region}
\alias{region<-,ready4_meso_region-method}
\title{region<-}
\usage{
region(x) <- value

\S4method{region}{ready4_meso_region}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_meso_region}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot region

Set the value of the slot region for S4 objects of class ready4_meso_region
}
