% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_region_type.R
\name{region_type<-}
\alias{region_type<-}
\alias{region_type<--ready4_meso_region}
\alias{region_type<-,ready4_meso_region-method}
\title{region_type<-}
\usage{
region_type(x) <- value

\S4method{region_type}{ready4_meso_region}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_meso_region}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot region_type

Set the value of the slot region_type for S4 objects of class ready4_meso_region
}
