% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_remove.R
\name{remove_grouped_popl_vars}
\alias{remove_grouped_popl_vars}
\title{Remove grouped population variables}
\usage{
remove_grouped_popl_vars(profiled_sf, featured_var_pfx_1L_chr)
}
\arguments{
\item{profiled_sf}{Profiled (a simple features object)}

\item{featured_var_pfx_1L_chr}{Featured variable prefix (a character vector of length one)}
}
\value{
Profiled (a simple features object)
}
\description{
remove_grouped_popl_vars() is a Remove function that edits an object, removing a specified element or elements. Specifically, this function implements an algorithm to remove grouped population variables. Function argument profiled_sf specifies the object to be updated. Argument featured_var_pfx_1L_chr provides the object to be updated. The function returns Profiled (a simple features object).
}
\keyword{internal}
