% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_remove.R
\name{remove_outlier_areas}
\alias{remove_outlier_areas}
\title{Remove outlier areas}
\usage{
remove_outlier_areas(areas_sf, outliers_chr, area_var_nm_1L_chr = character(0))
}
\arguments{
\item{areas_sf}{Areas (a simple features object)}

\item{outliers_chr}{Outliers (a character vector)}

\item{area_var_nm_1L_chr}{Area variable name (a character vector of length one), Default: character(0)}
}
\value{
Areas (a simple features object)
}
\description{
remove_outlier_areas() is a Remove function that edits an object, removing a specified element or elements. Specifically, this function implements an algorithm to remove outlier areas. Function argument areas_sf specifies the object to be updated. Argument outliers_chr provides the object to be updated. The function returns Areas (a simple features object).
}
\keyword{internal}
