% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suburb_based_effective_catchment.R
\name{remove_st_accronym}
\alias{remove_st_accronym}
\title{remove_st_accronym}
\usage{
remove_st_accronym(x, s_t,
  included_sts = c("Australian Capital Territory", "New South Wales",
  "Northern Territory", "Queensland", "South Australia", "Tasmania",
  "Victoria", "Western Australia"))
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{s_t}{PARAM_DESCRIPTION}

\item{included_sts}{PARAM_DESCRIPTION, Default: c("Australian Capital Territory", "New South Wales", "Northern Territory",
"Queensland", "South Australia", "Tasmania", "Victoria",
"Western Australia")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{pull}}
 \code{\link[stringr]{str_replace}}
}
