% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_rename.R
\name{rename_areas}
\alias{rename_areas}
\title{Rename areas}
\usage{
rename_areas(area_names_chr, correspondences_lup = NULL)
}
\arguments{
\item{area_names_chr}{Area names (a character vector)}

\item{correspondences_lup}{Correspondences (a lookup table), Default: NULL}
}
\value{
Area names (a character vector)
}
\description{
rename_areas() is a Rename function that renames elements of an object based on a pre-speccified schema. Specifically, this function implements an algorithm to rename areas. The function returns Area names (a character vector).
}
\keyword{internal}
