% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_intersect_sfs_keep_counts.R
\name{rename_vars_based_on_res}
\alias{rename_vars_based_on_res}
\title{rename_vars_based_on_res}
\usage{
rename_vars_based_on_res(sf, feature_nm, data_type, data_year, popl_var_prefix)
}
\arguments{
\item{sf}{PARAM_DESCRIPTION}

\item{feature_nm}{PARAM_DESCRIPTION}

\item{data_type}{PARAM_DESCRIPTION}

\item{data_year}{PARAM_DESCRIPTION}

\item{popl_var_prefix}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{funs}}
}
