% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_renew.R
\name{renew.vicinity_points}
\alias{renew.vicinity_points}
\alias{renew,vicinity_points-method}
\alias{renew.vicinity_processed}
\alias{renew,vicinity_processed-method}
\alias{renew.vicinity_raw}
\alias{renew,vicinity_raw-method}
\alias{renew-VicinityLookup}
\alias{renew,VicinityLookup-method}
\alias{renew-VicinityArguments}
\alias{renew,VicinityArguments-method}
\title{Renew (update) values}
\usage{
renew.vicinity_points(
  x,
  service_type_chr = character(0),
  cluster_name_chr = character(0),
  service_name_chr = character(0),
  lat_dbl = numeric(0),
  lng_dbl = numeric(0),
  what_1L_chr = "table",
  ...
)

\S4method{renew}{vicinity_points}(
  x,
  service_type_chr = character(0),
  cluster_name_chr = character(0),
  service_name_chr = character(0),
  lat_dbl = numeric(0),
  lng_dbl = numeric(0),
  what_1L_chr = "table",
  ...
)

renew.vicinity_processed(
  x,
  args_ls = NULL,
  additional_detail_chr = character(0),
  area_type_chr = character(0),
  area_bndy_yr_chr = character(0),
  country_chr = character(0),
  data_type_chr = character(0),
  main_feature_chr = character(0),
  name_chr = character(0),
  region_chr = character(0),
  source_reference_chr = character(0),
  year_chr = character(0),
  year_end_chr = character(0),
  year_start_chr = character(0),
  what_1L_chr = "table"
)

\S4method{renew}{vicinity_processed}(
  x,
  args_ls = NULL,
  additional_detail_chr = character(0),
  area_type_chr = character(0),
  area_bndy_yr_chr = character(0),
  country_chr = character(0),
  data_type_chr = character(0),
  main_feature_chr = character(0),
  name_chr = character(0),
  region_chr = character(0),
  source_reference_chr = character(0),
  year_chr = character(0),
  year_end_chr = character(0),
  year_start_chr = character(0),
  what_1L_chr = "table"
)

renew.vicinity_raw(
  x,
  area_type_chr = character(0),
  area_bndy_yr_chr = character(0),
  country_chr = character(0),
  data_type_chr = character(0),
  main_feature_chr = character(0),
  name_chr = character(0),
  processed_fls_dir_1L_chr = character(0),
  region_chr = character(0),
  source_reference_chr = character(0),
  year_chr = character(0),
  year_end_chr = ycharacter(0),
  year_start_chr = character(0),
  what_1L_chr = "table"
)

\S4method{renew}{vicinity_raw}(
  x,
  area_type_chr = character(0),
  area_bndy_yr_chr = character(0),
  country_chr = character(0),
  data_type_chr = character(0),
  main_feature_chr = character(0),
  name_chr = character(0),
  processed_fls_dir_1L_chr = character(0),
  region_chr = character(0),
  source_reference_chr = character(0),
  year_chr = character(0),
  year_end_chr = ycharacter(0),
  year_start_chr = character(0),
  what_1L_chr = "table"
)

\S4method{renew}{VicinityLookup}(
  x,
  package_1L_chr = character(0),
  path_1L_chr = character(0),
  tbl_data_type_1L_chr = "Geometry",
  template_ls = NULL,
  what_1L_chr = "processed"
)

\S4method{renew}{VicinityArguments}(x, raw_fls_dir_1L_chr, write_1L_lgl)
}
\arguments{
\item{x}{An object of class VicinityArguments}

\item{service_type_chr}{Service type (a character vector), Default: character(0)}

\item{cluster_name_chr}{Cluster name (a character vector), Default: character(0)}

\item{service_name_chr}{Service name (a character vector), Default: character(0)}

\item{lat_dbl}{Lat (a double vector), Default: numeric(0)}

\item{lng_dbl}{Longitude (a double vector), Default: numeric(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: 'processed'}

\item{...}{Additional arguments}

\item{args_ls}{Arguments (a list), Default: NULL}

\item{additional_detail_chr}{Additional detail (a character vector), Default: character(0)}

\item{area_type_chr}{Area type (a character vector), Default: character(0)}

\item{area_bndy_yr_chr}{Area boundary year (a character vector), Default: character(0)}

\item{country_chr}{Country (a character vector), Default: character(0)}

\item{data_type_chr}{Data type (a character vector), Default: character(0)}

\item{main_feature_chr}{Main feature (a character vector), Default: character(0)}

\item{name_chr}{Name (a character vector), Default: character(0)}

\item{region_chr}{Region (a character vector), Default: character(0)}

\item{source_reference_chr}{Source reference (a character vector), Default: character(0)}

\item{year_chr}{Year (a character vector), Default: character(0)}

\item{year_end_chr}{Year end (a character vector), Default: ycharacter(0)}

\item{year_start_chr}{Year start (a character vector), Default: character(0)}

\item{processed_fls_dir_1L_chr}{Processed files directory (a character vector of length one), Default: character(0)}

\item{package_1L_chr}{Package (a character vector of length one), Default: character(0)}

\item{path_1L_chr}{Path (a character vector of length one), Default: character(0)}

\item{tbl_data_type_1L_chr}{Table data type (a character vector of length one), Default: 'Geometry'}

\item{template_ls}{Template (a list), Default: NULL}

\item{raw_fls_dir_1L_chr}{Raw files directory (a character vector of length one)}

\item{write_1L_lgl}{Write (a logical vector of length one)}
}
\value{
x (An object)

x (An object)

x (An object)

x (An object of class VicinityLookup)

x (An object of class VicinityArguments)
}
\description{
renew.vicinity_points() is a renew method that renews an instance of a class by updating it with new data. This method is implemented for the ready4 submodule class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes. The function is called for its side effects and does not return a value.

renew.vicinity_processed() is a renew method that renews an instance of a class by updating it with new data. This method is implemented for the ready4 submodule class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data). The function is called for its side effects and does not return a value.

renew.vicinity_raw() is a renew method that renews an instance of a class by updating it with new data. This method is implemented for the ready4 submodule class for tibble object lookup table of metadata about raw (un-processed) spatial data to import. The function is called for its side effects and does not return a value.

renew method applied to VicinityLookup

renew method applied to VicinityArguments
}
