% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_reorder.R
\name{reorder_clusters_by_distances}
\alias{reorder_clusters_by_distances}
\title{Reorder clusters by distances}
\usage{
reorder_clusters_by_distances(
  clusters_by_distance_ls,
  distances_dbl,
  index_val_1L_int
)
}
\arguments{
\item{clusters_by_distance_ls}{Clusters by distance (a list)}

\item{distances_dbl}{Distances (a double vector)}

\item{index_val_1L_int}{Index value (an integer vector of length one)}
}
\value{
Clusters by distance (a list)
}
\description{
reorder_clusters_by_distances() is a Reorder function that reorders an object to conform to a pre-specified schema. Specifically, this function implements an algorithm to reorder clusters by distances. The function returns Clusters by distance (a list).
}
\keyword{internal}
