% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_reorder.R
\name{reorder_distance_list_by_cluster}
\alias{reorder_distance_list_by_cluster}
\title{Reorder distance list by cluster}
\usage{
reorder_distance_list_by_cluster(
  look_up_ref,
  clusters_by_distance_list,
  distances_vec
)
}
\arguments{
\item{look_up_ref}{PARAM_DESCRIPTION}

\item{clusters_by_distance_list}{PARAM_DESCRIPTION}

\item{distances_vec}{PARAM_DESCRIPTION}
}
\description{
reorder_distance_list_by_cluster() is a Reorder function that changes ordering of an object. Specifically, this function implements an algorithm to reorder distance list by cluster. The function is called for its side effects and does not return a value.
}
\keyword{internal}
