% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_reshape.R
\name{reshape_pop_projs}
\alias{reshape_pop_projs}
\title{Reshape pop projs}
\usage{
reshape_pop_projs(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\description{
reshape_pop_projs() is a Reshape function that modifies the shape of a data object. Specifically, this function implements an algorithm to reshape pop projs. The function is called for its side effects and does not return a value.
}
\keyword{internal}
