% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_sample.R
\name{sample_agent_coords}
\alias{sample_agent_coords}
\title{Sample agent coords}
\usage{
sample_agent_coords(profiled_sf, incident_cases)
}
\arguments{
\item{profiled_sf}{Profiled (a simple features object)}

\item{incident_cases}{PARAM_DESCRIPTION}
}
\value{
NA ()
}
\description{
sample_agent_coords() is a Sample function that samples values from specified distributions. Specifically, this function implements an algorithm to sample agent coords. The function is called for its side effects and does not return a value.
}
\keyword{internal}
