% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_data_pack.R
\name{write_fls_from_imp_and_upd_r4}
\alias{write_fls_from_imp_and_upd_r4}
\title{write_fls_from_imp_and_upd_r4}
\usage{
write_fls_from_imp_and_upd_r4(x, processed_dir_chr, crs_nbr_vec)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{processed_dir_chr}{PARAM_DESCRIPTION}

\item{crs_nbr_vec}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
