% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_save_raw.R
\name{save_raw.ready4_sp_import_lup}
\alias{save_raw.ready4_sp_import_lup}
\alias{save_raw,ready4_sp_import_lup-method}
\title{Save raw method applied toeadyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import..}
\usage{
save_raw.ready4_sp_import_lup(
  x,
  required_data,
  destination_directory,
  overwrite_lgl = F
)

\S4method{save_raw}{ready4_sp_import_lup}(x, required_data, destination_directory, overwrite_lgl = F)
}
\arguments{
\item{x}{An instance of Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.}

\item{required_data}{PARAM_DESCRIPTION}

\item{destination_directory}{PARAM_DESCRIPTION}

\item{overwrite_lgl}{Overwrite (a logical vector), Default: F}
}
\description{
save_raw.ready4_sp_import_lup() is a Save Raw method that saves the native version of a file format. This method is implemented for the Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.. The function is called for its side effects and does not return a value.
}
