% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_save_raw.R
\name{save_raw}
\alias{save_raw}
\alias{save_raw,ready4_sp_import_lup-method}
\alias{save_raw,ready4_sp_local-method}
\title{Save raw data}
\usage{
\S4method{save_raw}{ready4_sp_import_lup}(x, required_data, destination_directory, overwrite_lgl = F)

\S4method{save_raw}{ready4_sp_local}(x, return_r4_lgl)
}
\arguments{
\item{x}{Primary input object - see Usage section for allowable types}

\item{required_data}{PARAM_DESCRIPTION}

\item{destination_directory}{PARAM_DESCRIPTION}

\item{overwrite_lgl}{PARAM_DESCRIPTION, Default: F}

\item{return_r4_lgl}{PARAM_DESCRIPTION}
}
\description{
Save a local copy of the raw (unprocessed) input data.
}
