% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_ready4_sp_local_proc.R,
%   R/s4_ready4_sp_local_raw.R
\name{save_type<--ready4_sp_local_proc}
\alias{save_type<--ready4_sp_local_proc}
\alias{save_type<--ready4_sp_local_raw}
\title{save_type<-}
\arguments{
\item{x}{An object of class ready4_sp_local_proc}

\item{x}{An object of class ready4_sp_local_raw}
}
\description{
Set the value of the slot save_type for S4 objects of class ready4_sp_local_proc

Set the value of the slot save_type for S4 objects of class ready4_sp_local_raw
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
