% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_simplify.R
\name{simplify_geoms_in_lup}
\alias{simplify_geoms_in_lup}
\title{Simplify geoms in}
\usage{
simplify_geoms_in_lup(lup_r4, r_data_dir, crs_nbr)
}
\arguments{
\item{lup_r4}{Lookup table ready4 S4 (a ready4 S4 collection of lookup tables)}

\item{r_data_dir}{PARAM_DESCRIPTION}

\item{crs_nbr}{PARAM_DESCRIPTION}
}
\description{
simplify_geoms_in_lup() is a Simplify function that simplifies an object. Specifically, this function implements an algorithm to simplify geoms in lookup table. The function is called for its side effects and does not return a value.
}
\keyword{internal}
