% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sp_data_list.R
\name{simplify_sf}
\alias{simplify_sf}
\title{simplify_sf}
\usage{
simplify_sf(sf, crs = NULL)
}
\arguments{
\item{sf}{PARAM_DESCRIPTION}

\item{crs}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{st_crs}},\code{\link[sf]{st_geometry_type}},\code{\link[sf]{st_collection_extract}},\code{\link[sf]{st_transform}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{select}}
 \code{\link[geojsonio]{geojson_json}},\code{\link[geojsonio]{geojson_sf}}
 \code{\link[rmapshaper]{ms_simplify}}
}
