% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_sp_abbreviations_lup.R
\name{sp_abbreviations_lup}
\alias{sp_abbreviations_lup}
\alias{sp_abbreviations_lup-ready4_lookup}
\title{sp_abbreviations_lup}
\usage{
sp_abbreviations_lup(x)

\S4method{sp_abbreviations_lup}{ready4_lookup}(x)
}
\arguments{
\item{x}{An object of class ready4_lookup}
}
\description{
S4 Generic function to get the value of the slot sp_abbreviations_lup

Get the value of the slot sp_abbreviations_lup for S4 objects of class ready4_lookup
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
