% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_sp_abbreviations_lup.R
\name{sp_abbreviations_lup<--ready4_lookup}
\alias{sp_abbreviations_lup<--ready4_lookup}
\alias{sp_abbreviations_lup<-,ready4_lookup-method}
\title{sp_abbreviations_lup<-}
\usage{
\S4method{sp_abbreviations_lup}{ready4_lookup}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_lookup}
}
\description{
Set the value of the slot sp_abbreviations_lup for S4 objects of class ready4_lookup
}
