% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_sp_data_pack_lup.R
\name{sp_data_pack_lup}
\alias{sp_data_pack_lup}
\alias{sp_data_pack_lup-ready4_lookup}
\alias{sp_data_pack_lup,ready4_lookup-method}
\title{sp_data_pack_lup}
\usage{
sp_data_pack_lup(x)

\S4method{sp_data_pack_lup}{ready4_lookup}(x)
}
\arguments{
\item{x}{An object of class ready4_lookup}
}
\description{
S4 Generic function to get the value of the slot sp_data_pack_lup

Get the value of the slot sp_data_pack_lup for S4 objects of class ready4_lookup
}
