% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_sp_resolution_lup.R
\name{sp_resolution_lup}
\alias{sp_resolution_lup}
\title{sp_resolution_lup}
\arguments{
\item{x}{An object}
}
\description{
S4 Generic function to get the value of the slot sp_resolution_lup
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
