% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_sp_starter_sf_lup.R
\name{sp_starter_sf_lup<-}
\alias{sp_starter_sf_lup<-}
\title{sp_starter_sf_lup<-}
\arguments{
\item{x}{An object}
}
\description{
S4 Generic function to set the value of the slot sp_starter_sf_lup
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
