% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_sp_starter_sf_lup.R
\name{sp_starter_sf_lup<-}
\alias{sp_starter_sf_lup<-}
\alias{sp_starter_sf_lup<--ready4_lookup}
\alias{sp_starter_sf_lup<-,ready4_lookup-method}
\title{sp_starter_sf_lup<-}
\usage{
sp_starter_sf_lup(x) <- value

\S4method{sp_starter_sf_lup}{ready4_lookup}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_lookup}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot sp_starter_sf_lup

Set the value of the slot sp_starter_sf_lup for S4 objects of class ready4_lookup
}
