% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_sp_uid_lup.R
\name{sp_uid_lup}
\alias{sp_uid_lup}
\alias{sp_uid_lup-ready4_lookup}
\alias{sp_uid_lup,ready4_lookup-method}
\title{sp_uid_lup}
\usage{
sp_uid_lup(x)

\S4method{sp_uid_lup}{ready4_lookup}(x)
}
\arguments{
\item{x}{An object of class ready4_lookup}
}
\description{
S4 Generic function to get the value of the slot sp_uid_lup

Get the value of the slot sp_uid_lup for S4 objects of class ready4_lookup
}
