% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_sp_uid_lup.R
\name{sp_uid_lup<-}
\alias{sp_uid_lup<-}
\title{sp_uid_lup<-}
\arguments{
\item{x}{An object}
}
\description{
S4 Generic function to set the value of the slot sp_uid_lup
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
