% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_adjust_partial_in_custom_area.R
\name{spatial_adjust_partial_in_custom_area}
\alias{spatial_adjust_partial_in_custom_area}
\title{Adjust partially included sub-unit of custom area of interest.}
\usage{
spatial_adjust_partial_in_custom_area(main_unit_sf, excluded_unit_sf,
  excluded_refs, filter_one, filter_two)
}
\arguments{
\item{main_unit_sf}{A simple features object.}

\item{excluded_unit_sf}{A simple features object comprised of SA1s.}

\item{excluded_refs}{A .....}

\item{filter_one}{A .....}

\item{filter_two}{A .....}
}
\value{
A simple features object.
}
\description{
This function:
 - retrieves spatial information about the LGA which is partially contained within the
   custom area of interest;
 - identifies SA1s included in that LGA that are not part of custom area of interest and
   appends population attribute data to those SA1s;
 - identifies how much area and how much population is represented by the portion of the
   partially contained LGA that is outside that LGA;
 - creates an object representing that part of the partially contained LGA that lies within
   the custom area of interest; and
 - updates the area and population counts for this new object to represent only the proportion
   included in the custom area of interest.
}
\details{
Need to generalise. Currently assumes LGA as main unit and SA1 as minor unit.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{summarise}},\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate}}
 \code{\link[sf]{st_geometry}},\code{\link[sf]{geos_binary_ops}},\code{\link[sf]{geos_combine}}

Other spatial functions.: \code{\link{pop_sum_15.24}},
  \code{\link{spatial_cases_coordinates}},
  \code{\link{spatial_create_agent_coordinates_tb}},
  \code{\link{spatial_merge_areas_attributes}},
  \code{\link{spatial_one_year_projections_15_24}},
  \code{\link{spatial_population_growth}},
  \code{\link{spatial_profile_by_sa2}},
  \code{\link{spatial_select_rename_age_sex}},
  \code{\link{spatial_summarise_population_counts_0to29}},
  \code{\link{spatial_vic_pop_growth_lga}}
}
\concept{spatial functions.}
