% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/spatial_adjust_population_by_included_fraction.R
\name{spatial_adjust_population_by_included_fraction}
\alias{spatial_adjust_population_by_included_fraction}
\title{Adjust population counts by fraction of spatial unit included in a profiled area.}
\usage{
spatial_adjust_population_by_included_fraction(resolution_sa1s_sf,
  profiled_by_sa2_sf, return_resolution)
}
\arguments{
\item{resolution_sa1s_sf}{A simple features object comprised of SA1s}

\item{profiled_by_sa2_sf}{A simple features object comprised of SA2s.}

\item{return_resolution}{A string specifying "SA1" or "SA2" as the resolution of the
returned object,}
}
\value{
A simple features object.
}
\description{
This function:
 -
 -
}
\details{
Need to review if removal of linestrings is appropriate.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{join}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}},\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{select_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{funs}},\code{\link[dplyr]{summarise_all}}
 \code{\link[sf]{st_geometry}},\code{\link[sf]{geos_measures}}

Other spatial functions.: \code{\link{pop_sum_15.24}},
  \code{\link{spatial_adjust_partial_in_custom_area}},
  \code{\link{spatial_cases_coordinates}},
  \code{\link{spatial_create_agent_coordinates_tb}},
  \code{\link{spatial_intersect_profiled_resolution_units}},
  \code{\link{spatial_merge_areas_attributes}},
  \code{\link{spatial_one_year_projections_15_24}},
  \code{\link{spatial_population_growth}},
  \code{\link{spatial_profile_by_resolution_and_update_counts}},
  \code{\link{spatial_profile_by_sa2}},
  \code{\link{spatial_select_rename_age_sex}},
  \code{\link{spatial_summarise_population_counts_0to29}},
  \code{\link{spatial_vic_pop_growth_lga}}
}
\concept{spatial functions.}
