% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_area_within_xkm_of_cluster.R
\name{spatial_area_within_xkm_of_cluster}
\alias{spatial_area_within_xkm_of_cluster}
\alias{reorder_distance_list_by_cluster}
\alias{geom_distance_circles_to_bands}
\title{Calculate area within a specified distance of a cluster of points.}
\usage{
spatial_area_within_xkm_of_cluster(distance_km, clusters_vec,
  clusters_tbs_list, land_boundary_sf)

reorder_distance_list_by_cluster(look_up_ref, clusters_by_distance_list,
  distances_vec)

geom_distance_circles_to_bands(geom_distance_circle_sfs_list)
}
\arguments{
\item{distance_km}{PARAM_DESCRIPTION}

\item{clusters_vec}{PARAM_DESCRIPTION}

\item{clusters_tbs_list}{PARAM_DESCRIPTION}

\item{land_boundary_sf}{PARAM_DESCRIPTION}

\item{look_up_ref}{PARAM_DESCRIPTION}

\item{clusters_by_distance_list}{PARAM_DESCRIPTION}

\item{geom_distance_circle_sfs_list}{PARAM_DESCRIPTION}

\item{distances_vecclusters_tbs_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\section{Functions}{
\itemize{
\item \code{reorder_distance_list_by_cluster}: Calculates the .....

\item \code{geom_distance_circles_to_bands}: Calculates the .....
}}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}},\code{\link[purrr]{pluck}}
 \code{\link[stats]{setNames}}
}
