% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make_isochrs.R
\name{make_isochrs}
\alias{make_isochrs}
\title{make_isochrs}
\usage{
make_isochrs(
  long,
  lat,
  time_min,
  time_max,
  nbr_time_steps
)
}
\arguments{
\item{long}{PARAM_DESCRIPTION}

\item{lat}{PARAM_DESCRIPTION}

\item{time_min}{PARAM_DESCRIPTION}

\item{time_max}{PARAM_DESCRIPTION}

\item{nbr_time_steps}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Calculate area within specified drives timeof a cluster of points.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[osrm]{osrmIsochrone}}
 \code{\link[sf]{st_as_sf}}
 \code{\link[dplyr]{mutate}},\code{\link[dplyr]{arrange}}
}
