% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_sample_agent_coords.R
\name{sample_agent_coords}
\alias{sample_agent_coords}
\title{sample_agent_coords}
\usage{
sample_agent_coords(profiled_sf, incident_cases)
}
\arguments{
\item{profiled_sf}{PARAM_DESCRIPTION}

\item{incident_cases}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Generates coordinates for agents.

FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{st_sample}},\code{\link[sf]{st_transform}},\code{\link[sf]{st_geometry}}
 \code{\link[tibble]{as_tibble}},\code{\link[tibble]{deprecated}}
 \code{\link[stats]{setNames}}
}
