% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make_agt_coords_tb.R
\name{make_agt_coords_tb}
\alias{make_agt_coords_tb}
\title{make_agt_coords_tb}
\usage{
make_agt_coords_tb(
  profiled_area_sf,
  disorder,
  year,
  case_type = "expected.incidence",
  person_type = "p",
  resolution_unit
)
}
\arguments{
\item{profiled_area_sf}{PARAM_DESCRIPTION}

\item{disorder}{PARAM_DESCRIPTION}

\item{year}{PARAM_DESCRIPTION}

\item{case_type}{PARAM_DESCRIPTION, Default: 'expected.incidence'}

\item{person_type}{PARAM_DESCRIPTION, Default: 'p'}

\item{resolution_unit}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Creates a tibble of coordinates for agents.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_sub}}
 \code{\link[sf]{st_geometry}}
 \code{\link[dplyr]{select}},\code{\link[dplyr]{pull}},\code{\link[dplyr]{filter}}
 \code{\link[rlang]{sym}}
 \code{\link[purrr]{map2}}
}
