% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_intersect_profiled_resolution_units.R
\name{spatial_intersect_profiled_resolution_units}
\alias{spatial_intersect_profiled_resolution_units}
\title{Create a simple features object of two intersecting areas.}
\usage{
spatial_intersect_profiled_resolution_units(profiled_sf,
  profiled_colref = NA, profiled_rowref = NA, resolution_sf)
}
\arguments{
\item{profiled_sf}{A SF object corresponding to the geographic unit that is to be profiled.}

\item{profiled_colref}{A variable name from the profiled_sf object that is the basis for
subsetting the object if not all areas from the object are to be used.}

\item{resolution_sf}{A SF object corresponding to the geographic unit that is resolution
at which the profiled_sf object will be profiled.}

\item{profiled_rworef}{A value from the profiled_colref variable which, if specified, will
be used to identify a subset of the profiled_sf object.}
}
\value{
A simple features object.
}
\description{
This function:
  -
  -
}
\details{

}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}},\code{\link[dplyr]{pull}},\code{\link[dplyr]{slice}}
 \code{\link[stringr]{str_subset}}
 \code{\link[sf]{geos_binary_ops}}

Other spatial functions.: \code{\link{pop_sum_15.24}},
  \code{\link{spatial_adjust_partial_in_custom_area}},
  \code{\link{spatial_adjust_population_by_included_fraction}},
  \code{\link{spatial_cases_coordinates}},
  \code{\link{spatial_create_agent_coordinates_tb}},
  \code{\link{spatial_merge_areas_attributes}},
  \code{\link{spatial_one_year_projections_15_24}},
  \code{\link{spatial_population_growth}},
  \code{\link{spatial_profile_by_resolution_and_update_counts}},
  \code{\link{spatial_profile_by_sa2}},
  \code{\link{spatial_select_rename_age_sex}},
  \code{\link{spatial_summarise_population_counts_0to29}},
  \code{\link{spatial_vic_pop_growth_lga}}
}
\concept{spatial functions.}
