% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_merge_areas_attributes.R
\name{spatial_merge_areas_attributes}
\alias{spatial_merge_areas_attributes}
\title{Create tibble of population data for all spatial units of a specified type (currently Australian LGA and SA2).}
\usage{
spatial_merge_areas_attributes(area_unit, area_sf, child_youth_pop_t0_tb,
  child_youth_pop_t1_tb, seifa_deciles_by_unit)
}
\arguments{
\item{area_unit}{A String specifying that the area type is either "LGA" or "SA2".}

\item{area_sf}{A SF opject, with the boundary geometeries of the specified spatial unit type.}

\item{child_youth_pop_t0_tb}{A tibble of age and sex population data for each area unit specified above
for timepoint 0.}

\item{child_youth_pop_t1_tb}{A tibble of age and sex population data for each area unit specified above
for timepoint 1.}

\item{seifa_deciles_by_unit}{A tibble of SEIFA data for each area unit specified above.}
}
\value{
Returns a tibble.
}
\description{
Sources and merges population and SEIFA data for all spatial units of a specified type.
}
\details{
Makes data transformations (variable name, changes from strings to factors) necessary
to ensure that merged objects are compatible.
}
\examples{

}
\seealso{
Other spatial functions.: \code{\link{pop_sum_15.24}},
  \code{\link{spatial_adjust_partial_in_custom_area}},
  \code{\link{spatial_adjust_population_by_included_fraction}},
  \code{\link{spatial_cases_coordinates}},
  \code{\link{spatial_create_agent_coordinates_tb}},
  \code{\link{spatial_intersect_profiled_resolution_units}},
  \code{\link{spatial_one_year_projections_15_24}},
  \code{\link{spatial_population_growth}},
  \code{\link{spatial_profile_by_resolution_and_update_counts}},
  \code{\link{spatial_profile_by_sa2}},
  \code{\link{spatial_select_rename_age_sex}},
  \code{\link{spatial_summarise_population_counts_0to29}},
  \code{\link{spatial_vic_pop_growth_lga}}
}
\concept{spatial functions.}
