% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_one_year_projections_15.24.R
\name{spatial_one_year_projections_15_24}
\alias{spatial_one_year_projections_15_24}
\title{Summarise 15-24 female and male population by region for one year.}
\usage{
spatial_one_year_projections_15_24(population_forecast_range_one_year,
  namestub)
}
\arguments{
\item{population_forecast_range_one_year}{A COMPLETETEXT}

\item{namestub}{A String.}
}
\value{
A list.
}
\description{
This function creates a list comprised of female population projections
and male population projections for specified regions in a specified
year.
}
\details{

}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[magrittr]{extract}}
 \code{\link[stringr]{str_subset}},\code{\link[stringr]{str_replace}}
 \code{\link[purrr]{reduce}}

Other spatial functions.: \code{\link{pop_sum_15.24}},
  \code{\link{spatial_adjust_partial_in_custom_area}},
  \code{\link{spatial_cases_coordinates}},
  \code{\link{spatial_create_agent_coordinates_tb}},
  \code{\link{spatial_merge_areas_attributes}},
  \code{\link{spatial_population_growth}},
  \code{\link{spatial_profile_by_sa2}},
  \code{\link{spatial_summarise_population_counts_0to29}}
}
\concept{spatial functions.}
