% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_population_growth.R
\name{spatial_population_growth}
\alias{spatial_population_growth}
\title{Calculate population growth numbers and rate.}
\usage{
spatial_population_growth(population_tib, t0, t1)
}
\arguments{
\item{population_tib}{A tibble of population projection data for a profiled region
broken down by unit of resolution.}

\item{t0}{A String specifying the baseline year. This must match up with the years
specified in the datasource specified in population_tib.}

\item{t1}{A String specifying the followup year. This must match up with the years
specified in the datasource specified in population_tib.}
}
\value{
A tibble.
}
\description{
This function updates an inputted tibble with the actual or projected growth in a profiled
region by each unit of resolution.
}
\details{

}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{sym}}
 \code{\link[dplyr]{mutate}}

Other spatial functions.: \code{\link{pop_sum_15.24}},
  \code{\link{spatial_adjust_partial_in_custom_area}},
  \code{\link{spatial_cases_coordinates}},
  \code{\link{spatial_create_agent_coordinates_tb}},
  \code{\link{spatial_merge_areas_attributes}},
  \code{\link{spatial_one_year_projections_15_24}},
  \code{\link{spatial_profile_by_sa2}},
  \code{\link{spatial_summarise_population_counts_0to29}}
}
\concept{spatial functions.}
