% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/spatial_profile_by_resoultion_and_update_counts.R
\name{spatial_profile_by_resolution_and_update_counts}
\alias{spatial_profile_by_resolution_and_update_counts}
\title{Create a simple features object of two intersecting areas and adjust population counts by
fraction of spatial unit included in a profiled area.}
\usage{
spatial_profile_by_resolution_and_update_counts(profiled_sf,
  profiled_colref = NA, profiled_rowref = NA, resolution_sf,
  resolution_sa1s_sf, resolution_sa2s_sf, return_resolution)
}
\arguments{
\item{profiled_sf}{A SF object corresponding to the geographic unit that is to be profiled.}

\item{profiled_colref}{A variable name from the profiled_sf object that is the basis for
subsetting the object if not all areas from the object are to be used.}

\item{resolution_sf}{A SF object corresponding to the geographic unit that is resolution
at which the profiled_sf object will be profiled.}

\item{resolution_sa1s_sf}{A simple features object comprised of SA1s}

\item{resolution_sa2s_sf}{A simple features object comprised of SA2s.}

\item{return_resolution}{A string specifying "SA1", "SA2" or "LGA" as the resolution of the
returned object,}

\item{profiled_rworef}{A value from the profiled_colref variable which, if specified, will
be used to identify a subset of the profiled_sf object.}
}
\value{
A simple features object.
}
\description{
This function:
 -
 -
}
\details{

}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
Other spatial functions.: \code{\link{pop_sum_15.24}},
  \code{\link{spatial_adjust_partial_in_custom_area}},
  \code{\link{spatial_adjust_population_by_included_fraction}},
  \code{\link{spatial_cases_coordinates}},
  \code{\link{spatial_create_agent_coordinates_tb}},
  \code{\link{spatial_intersect_profiled_resolution_units}},
  \code{\link{spatial_merge_areas_attributes}},
  \code{\link{spatial_one_year_projections_15_24}},
  \code{\link{spatial_population_growth}},
  \code{\link{spatial_profile_by_sa2}},
  \code{\link{spatial_select_rename_age_sex}},
  \code{\link{spatial_summarise_population_counts_0to29}},
  \code{\link{spatial_vic_pop_growth_lga}}
}
\concept{spatial functions.}
