% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_profile_by_sa2.R
\name{spatial_profile_by_sa2}
\alias{spatial_profile_by_sa2}
\title{Create a simple features object with population counts based on proportion of SA2
included in the profiled area.}
\usage{
spatial_profile_by_sa2(profiled_sf, col_ref, res_sa1s_sf, res_sa2s_sf)
}
\arguments{
\item{profiled_sf}{A SF object corresponding to the geographic unit that is to be profiled.}

\item{col_ref}{A variable name from the profiled_sf object from which a vector of values will
be used to create merge objects.}

\item{res_sa1s_sf}{A simple features object comprised of SA1s}

\item{res_sa2s_sf}{A simple features object comprised of SA2s.}
}
\value{
A simple features object.
}
\description{
This function:
 -
 -
}
\details{

}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[seplyr]{filter_se}}
 \code{\link[purrr]{map}},\code{\link[purrr]{reduce}}
 \code{\link[dplyr]{pull}}

Other spatial functions.: \code{\link{pop_sum_15.24}},
  \code{\link{spatial_adjust_partial_in_custom_area}},
  \code{\link{spatial_cases_coordinates}},
  \code{\link{spatial_create_agent_coordinates_tb}},
  \code{\link{spatial_merge_areas_attributes}},
  \code{\link{spatial_one_year_projections_15_24}},
  \code{\link{spatial_population_growth}},
  \code{\link{spatial_select_rename_age_sex}},
  \code{\link{spatial_summarise_population_counts_0to29}},
  \code{\link{spatial_vic_pop_growth_lga}}
}
\concept{spatial functions.}
