% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_select_rename_age_sex.R
\name{spatial_select_rename_age_sex}
\alias{spatial_select_rename_age_sex}
\title{Select and rename variables for 0-29 year old males and females in a population tibble.}
\usage{
spatial_select_rename_age_sex(population_tib, year, also_include)
}
\arguments{
\item{population_tib}{A tibble of population projection data for a profiled region
broken down by unit of resolution.}

\item{year}{A String specifying the year to which data relates.}

\item{also_include}{A String vector specifying the year to which data relates.}
}
\value{
A tibble.
}
\description{
This function updates an inputted tibble of population data for a profiled region by subsetting
the data and renaming the variables.
}
\details{

}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}},\code{\link[dplyr]{reexports}}

Other spatial functions.: \code{\link{pop_sum_15.24}},
  \code{\link{spatial_adjust_partial_in_custom_area}},
  \code{\link{spatial_adjust_population_by_included_fraction}},
  \code{\link{spatial_cases_coordinates}},
  \code{\link{spatial_create_agent_coordinates_tb}},
  \code{\link{spatial_merge_areas_attributes}},
  \code{\link{spatial_one_year_projections_15_24}},
  \code{\link{spatial_population_growth}},
  \code{\link{spatial_profile_by_sa2}},
  \code{\link{spatial_summarise_population_counts_0to29}},
  \code{\link{spatial_vic_pop_growth_lga}}
}
\concept{spatial functions.}
