% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_summarise_population_counts_0to29.R
\name{spatial_summarise_population_counts_0to29}
\alias{spatial_summarise_population_counts_0to29}
\title{Produces a summary of population counts for 0-29 year old females, males and persons.}
\usage{
spatial_summarise_population_counts_0to29(population_sf, groupvar,
  unitname, t0, t1)
}
\arguments{
\item{population_sf}{A SF object of data to be summarised.}

\item{groupvar}{A String specifying the variable to group data by.}

\item{unitname}{A String specifying the name of the data resolution unit.}

\item{t0}{A String specifying the baseline year.}

\item{t1}{A String specifying the follow up year.}
}
\value{
Returns a tibble.
}
\description{
Updates an inputted tibble with summaries by the unit of resolution (currently SA2 or LGA).
}
\details{
Note that group by action uses Mean rather than Sum function : it is therefore
assumed that inputted data will not require summing.
}
\examples{

}
\seealso{
Other spatial functions.: \code{\link{pop_sum_15.24}},
  \code{\link{spatial_adjust_partial_in_custom_area}},
  \code{\link{spatial_cases_coordinates}},
  \code{\link{spatial_create_agent_coordinates_tb}},
  \code{\link{spatial_merge_areas_attributes}},
  \code{\link{spatial_one_year_projections_15_24}},
  \code{\link{spatial_population_growth}},
  \code{\link{spatial_profile_by_sa2}},
  \code{\link{spatial_select_rename_age_sex}},
  \code{\link{spatial_vic_pop_growth_lga}}
}
\concept{spatial functions.}
