% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_vic_pop_growth.R
\name{spatial_vic_pop_growth_lga}
\alias{spatial_vic_pop_growth_lga}
\title{Read in Victorian 2016 to 2021 population growth data}
\usage{

  spatial_vic_pop_growth_lga(vic_pop_growth_by_age_lga_t0 = rfwn.setup::data_get(data_lookup_tb
  = aus_spatial_lookup_tb, lookup_reference = "aus_lga_vic_att_ppr_2016",
  lookup_variable = "name", target_variable = "source_reference") \%>\%
  purrr::pluck("y2016"),
  vic_pop_growth_by_age_lga_t1 = rfwn.setup::data_get(data_lookup_tb =
  aus_spatial_lookup_tb, lookup_reference = "aus_lga_vic_att_ppr_2016",
  lookup_variable = "name", target_variable = "source_reference") \%>\%
  purrr::pluck("y2021"), t0 = "2016", t1 = "2021")
}
\arguments{
\item{vic_pop_growth_by_age_lga_t0}{PARAM_DESCRIPTION, Default: rfwn.setup::data_get(data_lookup_tb = aus_spatial_lookup_tb,
lookup_reference = "aus_lga_vic_att_ppr_2016", lookup_variable = "name",
target_variable = "source_reference") %>% purrr::pluck("y2016")}

\item{vic_pop_growth_by_age_lga_t1}{PARAM_DESCRIPTION, Default: rfwn.setup::data_get(data_lookup_tb = aus_spatial_lookup_tb,
lookup_reference = "aus_lga_vic_att_ppr_2016", lookup_variable = "name",
target_variable = "source_reference") %>% purrr::pluck("y2021"), Default: '2016'}

\item{t0}{A String specifying the baseline year. This must match up with the years
specified in the datasource specified in vic_pop_growth_by_age_lga_t0 (spaced at
five year intervals: "2016", "2021", "2026", etc).}

\item{t1}{A String specifying the baseline year. This must match up with the years
specified in the datasource specified in vic_pop_growth_by_age_lga_t1 (spaced at
five year intervals: "2016", "2021", "2026", etc), Default: '2021'}
}
\value{
A tibble.
}
\description{
This function creates a tibble of projected population growth by Victorian LGA
between 2016 and 2021.
}
\details{


DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rfwn.setup]{data_get}}
 \code{\link[purrr]{pluck}}
 \code{\link[dplyr]{select}},\code{\link[dplyr]{join}}

Other spatial functions.: \code{\link{pop_sum_15.24}},
  \code{\link{spatial_adjust_partial_in_custom_area}},
  \code{\link{spatial_adjust_population_by_included_fraction}},
  \code{\link{spatial_cases_coordinates}},
  \code{\link{spatial_create_agent_coordinates_tb}},
  \code{\link{spatial_intersect_profiled_resolution_units}},
  \code{\link{spatial_merge_areas_attributes}},
  \code{\link{spatial_one_year_projections_15_24}},
  \code{\link{spatial_population_growth}},
  \code{\link{spatial_profile_by_resolution_and_update_counts}},
  \code{\link{spatial_profile_by_sa2}},
  \code{\link{spatial_select_rename_age_sex}},
  \code{\link{spatial_summarise_population_counts_0to29}}
}
\concept{spatial functions.}
