% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_st_data.R
\name{st_data}
\alias{st_data}
\alias{st_data-ready4_env}
\alias{st_data,ready4_env-method}
\title{st_data}
\usage{
st_data(x)

\S4method{st_data}{ready4_env}(x)
}
\arguments{
\item{x}{An object of class ready4_env}
}
\description{
S4 Generic function to get the value of the slot st_data

Get the value of the slot st_data for S4 objects of class ready4_env
}
