% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_st_data.R
\name{st_data<-}
\alias{st_data<-}
\alias{st_data<--ready4_env}
\title{st_data<-}
\usage{
st_data(x) <- value

\S4method{st_data}{ready4_env}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_env}
}
\description{
S4 Generic function to set the value of the slot st_data

Set the value of the slot st_data for S4 objects of class ready4_env
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
