% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_env.R, R/gnrc_st_data.R
\name{st_data-ready4_env}
\alias{st_data-ready4_env}
\alias{st_data}
\title{st_data}
\arguments{
\item{x}{An object of class ready4_env}

\item{x}{An object}
}
\description{
Get the value of the slot st_data for S4 objects of class ready4_env

S4 Generic function to get the value of the slot st_data
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
