% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_st_envir.R
\name{st_envir<-}
\alias{st_envir<-}
\title{st_envir<-}
\arguments{
\item{x}{An object}
}
\description{
S4 Generic function to set the value of the slot st_envir
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
