% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_territory_sa2_boundaries.R
\name{state_territory_sa1_pop_counts_projs}
\alias{state_territory_sa1_pop_counts_projs}
\title{state_territory_sa1_pop_counts_projs}
\usage{
state_territory_sa1_pop_counts_projs(s_t)
}
\arguments{
\item{s_t}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready.utils]{data_get}}
 \code{\link[dplyr]{filter}}
}
