% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_territory_sa2_boundaries.R
\name{state_territory_sa2_pop_counts_0_30}
\alias{state_territory_sa2_pop_counts_0_30}
\alias{state_territory_sa1_pop_counts_projs}
\title{Get population counts from age 0-30 by SA2 for a given state / territory.}
\usage{
state_territory_sa2_pop_counts_0_30(s_t)

state_territory_sa1_pop_counts_projs(s_t)
}
\arguments{
\item{s_t}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\section{Functions}{
\itemize{
\item \code{state_territory_sa1_pop_counts_projs}: Calculate ...
}}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready.data]{data_get}}
 \code{\link[dplyr]{filter}}
}
