% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_subset.R
\name{subset_sf_by_feature}
\alias{subset_sf_by_feature}
\title{Subset simple features object by feature}
\usage{
subset_sf_by_feature(profiled_sf, group_by_var)
}
\arguments{
\item{profiled_sf}{Profiled (a simple features object)}

\item{group_by_var}{PARAM_DESCRIPTION}
}
\description{
subset_sf_by_feature() is a Subset function that subsets and object. Specifically, this function implements an algorithm to subset simple features object by feature. The function is called for its side effects and does not return a value.
}
\keyword{internal}
