% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_subset.R
\name{subset_sf_ls_by_common_vars}
\alias{subset_sf_ls_by_common_vars}
\title{Subset simple features object list by common vars}
\usage{
subset_sf_ls_by_common_vars(sf_ls)
}
\arguments{
\item{sf_ls}{Simple features object list (a list of simple features objects)}
}
\description{
subset_sf_ls_by_common_vars() is a Subset function that subsets and object. Specifically, this function implements an algorithm to subset simple features object list by common vars. The function is called for its side effects and does not return a value.
}
\keyword{internal}
