% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suburb_based_effective_catchment.R
\name{suburb_based_effective_catch_excl_outliers}
\alias{suburb_based_effective_catch_excl_outliers}
\title{suburb_based_effective_catch_excl_outliers}
\usage{
suburb_based_effective_catch_excl_outliers(outliers,
  suburb_based_effective_catch_inc_outliers_sf)
}
\arguments{
\item{outliers}{PARAM_DESCRIPTION}

\item{suburb_based_effective_catch_inc_outliers_sf}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{pull}},\code{\link[dplyr]{filter}}
 \code{\link[stringr]{str_detect}}
 \code{\link[stats]{setNames}}
 \code{\link[purrr]{prepend}},\code{\link[purrr]{reduce}}
 \code{\link[sf]{st_as_sf}}
}
