% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update_counts_by_area.R
\name{transform_sfx_to_pfx}
\alias{transform_sfx_to_pfx}
\title{transform_sfx_to_pfx}
\usage{
transform_sfx_to_pfx(data_tb, suffix)
}
\arguments{
\item{data_tb}{PARAM_DESCRIPTION}

\item{suffix}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{funs}}
}
