% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sim_data_env.R
\name{sum_at_diff_funs}
\alias{sum_at_diff_funs}
\title{sum_at_diff_funs
FUNCTION_DESCRIPTION}
\usage{
sum_at_diff_funs(data_sf, var_list, funs_list, group_by)
}
\arguments{
\item{data_sf}{PARAM_DESCRIPTION}

\item{var_list}{PARAM_DESCRIPTION}

\item{funs_list}{PARAM_DESCRIPTION}

\item{group_by}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
sum_at_diff_funs
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map2}},\code{\link[purrr]{reduce}}
 \code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise_all}},\code{\link[dplyr]{select}},\code{\link[dplyr]{reexports}}
 \code{\link[rlang]{sym}}
 \code{\link[sf]{st_join}}
}
