% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update_counts_by_area.R
\name{sum_pop_by_multiple_groups_sf}
\alias{sum_pop_by_multiple_groups_sf}
\title{sum_pop_by_multiple_groups_sf}
\usage{
sum_pop_by_multiple_groups_sf(profiled_sf, group_by_var, age_sex_var_name,
  data_year, age_sex_pop_resolution, tot_pop_resolution,
  popl_var_prefix = "")
}
\arguments{
\item{profiled_sf}{PARAM_DESCRIPTION}

\item{group_by_var}{PARAM_DESCRIPTION}

\item{age_sex_var_name}{PARAM_DESCRIPTION}

\item{data_year}{PARAM_DESCRIPTION}

\item{age_sex_pop_resolution}{PARAM_DESCRIPTION}

\item{tot_pop_resolution}{PARAM_DESCRIPTION}

\item{popl_var_prefix}{PARAM_DESCRIPTION, Default: ''}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
