% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update_counts_by_area.R
\name{sum_updated_pop_by_grp}
\alias{sum_updated_pop_by_grp}
\title{sum_updated_pop_by_grp}
\usage{
sum_updated_pop_by_grp(profiled_sf, grp_var_name, nse_objs_ls,
  suff_to_pref = FALSE, top_level = FALSE)
}
\arguments{
\item{profiled_sf}{PARAM_DESCRIPTION}

\item{grp_var_name}{PARAM_DESCRIPTION}

\item{nse_objs_ls}{PARAM_DESCRIPTION}

\item{suff_to_pref}{PARAM_DESCRIPTION, Default: FALSE}

\item{top_level}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{st_geometry}}
 \code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{funs}},\code{\link[dplyr]{select_all}},\code{\link[dplyr]{bind}},\code{\link[dplyr]{join}}
 \code{\link[rlang]{sym}}
}
