% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_temporal_max.R
\name{temporal_max}
\alias{temporal_max}
\alias{temporal_max-ready4_macro}
\alias{temporal_max,ready4_macro-method}
\title{temporal_max}
\usage{
temporal_max(x)

\S4method{temporal_max}{ready4_macro}(x)
}
\arguments{
\item{x}{An object of class ready4_macro}
}
\description{
S4 Generic function to get the value of the slot temporal_max

Get the value of the slot temporal_max for S4 objects of class ready4_macro
}
