% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_temporal_min.R
\name{temporal_min}
\alias{temporal_min}
\alias{temporal_min-ready4_macro}
\title{temporal_min}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_macro}
}
\description{
S4 Generic function to get the value of the slot temporal_min

Get the value of the slot temporal_min for S4 objects of class ready4_macro
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
