% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_temporal_min.R
\name{temporal_min<-}
\alias{temporal_min<-}
\alias{temporal_min<--ready4_macro}
\alias{temporal_min<-,ready4_macro-method}
\title{temporal_min<-}
\usage{
temporal_min(x) <- value

\S4method{temporal_min}{ready4_macro}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_macro}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot temporal_min

Set the value of the slot temporal_min for S4 objects of class ready4_macro
}
