% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_agent_areas}
\alias{transform_agent_areas}
\title{Transform agent areas}
\usage{
transform_agent_areas(
  agent_areas_tb,
  area_var_nm_1L_chr = "Suburb",
  match_var_nm_1L_chr = character(0),
  match_value_xx = NULL,
  title_case_1L_lgl = T
)
}
\arguments{
\item{agent_areas_tb}{Agent areas (a tibble)}

\item{area_var_nm_1L_chr}{Area variable name (a character vector of length one), Default: 'Suburb'}

\item{match_var_nm_1L_chr}{Match variable name (a character vector of length one), Default: character(0)}

\item{match_value_xx}{Match value (an output object of multiple potential types), Default: NULL}

\item{title_case_1L_lgl}{Title case (a logical vector of length one), Default: T}
}
\value{
Agent areas (a tibble)
}
\description{
transform_agent_areas() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform agent areas. The function returns Agent areas (a tibble).
}
\keyword{internal}
