% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_circles_to_bands}
\alias{transform_circles_to_bands}
\title{Transform circles to bands}
\usage{
transform_circles_to_bands(geomc_dist_circles_ls)
}
\arguments{
\item{geomc_dist_circles_ls}{Geometric distance circles (a list)}
}
\value{
Bands (a list)
}
\description{
transform_circles_to_bands() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform circles to bands. Function argument geomc_dist_circles_ls specifies the object to be updated. The function returns Bands (a list).
}
\keyword{internal}
