% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_multiple_vals}
\alias{transform_multiple_vals}
\title{Transform multiple values}
\usage{
transform_multiple_vals(
  distribution_chr,
  dstr_param_1_dbl,
  dstr_param_2_dbl,
  dstr_param_3_dbl,
  transformation_chr
)
}
\arguments{
\item{distribution_chr}{Distribution (a character vector)}

\item{dstr_param_1_dbl}{Distribution parameter 1 (a double vector)}

\item{dstr_param_2_dbl}{Distribution parameter 2 (a double vector)}

\item{dstr_param_3_dbl}{Distribution parameter 3 (a double vector)}

\item{transformation_chr}{Transformation (a character vector)}
}
\value{
Transformed vas (a numeric)
}
\description{
transform_multiple_vals() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform multiple values. Function argument distribution_chr specifies the object to be updated. Argument dstr_param_1_dbl provides the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
