% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_polyline_to_sf}
\alias{transform_polyline_to_sf}
\title{Transform polyline to simple features object}
\usage{
transform_polyline_to_sf(
  polyline_xx,
  crs_1L_dbl,
  mode_of_transport_1L_chr,
  travel_time_hours_1L_dbl
)
}
\arguments{
\item{polyline_xx}{Polyline (an output object of multiple potential types)}

\item{crs_1L_dbl}{Coordinates reference system (a double vector of length one)}

\item{mode_of_transport_1L_chr}{Mode of transport (a character vector of length one)}

\item{travel_time_hours_1L_dbl}{Travel time hours (a double vector of length one)}
}
\value{
Geometry (a simple features object)
}
\description{
transform_polyline_to_sf() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform polyline to simple features object. Function argument polyline_xx specifies the object to be updated. Argument crs_1L_dbl provides the object to be updated. The function returns Geometry (a simple features object).
}
\keyword{internal}
