% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_sf_ls}
\alias{transform_sf_ls}
\title{Transform simple features object list}
\usage{
transform_sf_ls(sf_ls)
}
\arguments{
\item{sf_ls}{Simple features object list (a list of simple features objects)}
}
\value{
Transformed (a list of simple features objects)
}
\description{
transform_sf_ls() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform simple features object list. Function argument sf_ls specifies the object to be updated. The function returns Transformed (a list of simple features objects).
}
\keyword{internal}
