% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_shape_of_popl_predns_ds}
\alias{transform_shape_of_popl_predns_ds}
\title{Transform shape of population preductions dataset}
\usage{
transform_shape_of_popl_predns_ds(popl_predns_ds, prefix_1L_chr = "Persons.")
}
\arguments{
\item{popl_predns_ds}{PARAM_DESCRIPTION}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'Persons.'}
}
\value{
Reshaped (a tibble)
}
\description{
transform_shape_of_popl_predns_ds() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform shape of population preductions dataset. Function argument popl_predns_ds specifies the object to be updated. Argument prefix_1L_chr provides the object to be updated. The function returns Reshaped (a tibble).
}
\keyword{internal}
