% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_sp_local_r4_to_local_proc_r4}
\alias{transform_sp_local_r4_to_local_proc_r4}
\title{Transform sp local ready4 S4 to local process}
\usage{
transform_sp_local_r4_to_local_proc_r4(
  x,
  import_chr_vec,
  raw_data_dir_chr,
  save_lgl
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{import_chr_vec}{PARAM_DESCRIPTION}

\item{raw_data_dir_chr}{Raw data directory (a character vector)}

\item{save_lgl}{Save (a logical vector)}
}
\description{
transform_sp_local_r4_to_local_proc_r4() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform sp local ready4 s4 to local process ready4 s4. Function argument x specifies the object to be updated. Argument import_chr_vec provides the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
