% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_to_simpler_sf}
\alias{transform_to_simpler_sf}
\title{Transform to simpler simple features object}
\usage{
transform_to_simpler_sf(sf, crs_dbl = NULL)
}
\arguments{
\item{sf}{Simple features object (a simple features object)}

\item{crs_dbl}{Coordinates reference system (a double vector), Default: NULL}
}
\value{
Polygons (a simple features object)
}
\description{
transform_to_simpler_sf() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform to simpler simple features object. Function argument sf specifies the object to be updated. Argument crs_dbl provides the object to be updated. The function returns Polygons (a simple features object).
}
\keyword{internal}
