% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_tt_polygon_to_sf}
\alias{transform_tt_polygon_to_sf}
\title{Transform tt polygon to}
\usage{
transform_tt_polygon_to_sf(
  tt_polyline,
  mode_of_transport,
  travel_time_hours,
  crs
)
}
\arguments{
\item{tt_polyline}{PARAM_DESCRIPTION}

\item{mode_of_transport}{PARAM_DESCRIPTION}

\item{travel_time_hours}{PARAM_DESCRIPTION}

\item{...}{Additional arguments}
}
\description{
transform_tt_polygon_to_sf() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform tt polygon to simple features object. Function argument tt_polyline specifies the object to be updated. Argument mode_of_transport provides the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
