% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_value}
\alias{transform_value}
\title{Transform value}
\usage{
transform_value(
  distribution_1L_chr,
  dstr_param_1_1L_dbl,
  dstr_param_2_1L_dbl,
  dstr_param_3_1L_dbl,
  transformation_1L_chr
)
}
\arguments{
\item{distribution_1L_chr}{Distribution (a character vector of length one)}

\item{dstr_param_1_1L_dbl}{Distribution parameter 1 (a double vector of length one)}

\item{dstr_param_2_1L_dbl}{Distribution parameter 2 (a double vector of length one)}

\item{dstr_param_3_1L_dbl}{Distribution parameter 3 (a double vector of length one)}

\item{transformation_1L_chr}{Transformation (a character vector of length one)}
}
\value{
Transformed value length one (a numeric)
}
\description{
transform_value() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform value. The function is called for its side effects and does not return a value.
}
\keyword{internal}
