% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_travel_mode.R
\name{travel_mode}
\alias{travel_mode}
\alias{travel_mode-ready4_micro}
\title{travel_mode}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_micro}
}
\description{
S4 Generic function to get the value of the slot travel_mode

Get the value of the slot travel_mode for S4 objects of class ready4_micro
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
