% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_travel_mode.R
\name{travel_mode<-}
\alias{travel_mode<-}
\alias{travel_mode<--ready4_micro}
\alias{travel_mode<-,ready4_micro-method}
\title{travel_mode<-}
\usage{
travel_mode(x) <- value

\S4method{travel_mode}{ready4_micro}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_micro}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot travel_mode

Set the value of the slot travel_mode for S4 objects of class ready4_micro
}
