% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_travel_time_mins_cuts.R
\name{travel_time_mins_cuts}
\alias{travel_time_mins_cuts}
\alias{travel_time_mins_cuts-ready4_micro}
\alias{travel_time_mins_cuts,ready4_micro-method}
\title{travel_time_mins_cuts}
\usage{
travel_time_mins_cuts(x)

\S4method{travel_time_mins_cuts}{ready4_micro}(x)
}
\arguments{
\item{x}{An object of class ready4_micro}
}
\description{
S4 Generic function to get the value of the slot travel_time_mins_cuts

Get the value of the slot travel_time_mins_cuts for S4 objects of class ready4_micro
}
