% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_travel_time_mins_cuts.R
\name{travel_time_mins_cuts<-}
\alias{travel_time_mins_cuts<-}
\alias{travel_time_mins_cuts<--ready4_micro}
\title{travel_time_mins_cuts<-}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_micro}
}
\description{
S4 Generic function to set the value of the slot travel_time_mins_cuts

Set the value of the slot travel_time_mins_cuts for S4 objects of class ready4_micro
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
