% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_travel_time_mins_cuts.R
\name{travel_time_mins_cuts<-}
\alias{travel_time_mins_cuts<-}
\alias{travel_time_mins_cuts<--ready4_micro}
\alias{travel_time_mins_cuts<-,ready4_micro-method}
\title{travel_time_mins_cuts<-}
\usage{
travel_time_mins_cuts(x) <- value

\S4method{travel_time_mins_cuts}{ready4_micro}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_micro}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot travel_time_mins_cuts

Set the value of the slot travel_time_mins_cuts for S4 objects of class ready4_micro
}
