% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_union.R
\name{union_one_travel_time_band_across_sites}
\alias{union_one_travel_time_band_across_sites}
\title{Union one travel time band across sites}
\usage{
union_one_travel_time_band_across_sites(
  time_band_ref,
  one_cluster_time_bands_list
)
}
\arguments{
\item{time_band_ref}{PARAM_DESCRIPTION}

\item{one_cluster_time_bands_list}{PARAM_DESCRIPTION}
}
\description{
union_one_travel_time_band_across_sites() is an Union function that unions objects. Specifically, this function implements an algorithm to union one travel time band across sites. The function is called for its side effects and does not return a value.
}
\keyword{internal}
