% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_pop_by_inc_area}
\alias{update_pop_by_inc_area}
\title{Update pop by include area}
\usage{
update_pop_by_inc_area(
  profiled_sf,
  sp_unit,
  data_year,
  concept,
  age_sex_var_name = NULL,
  age_sex_pop_resolution = NULL,
  tot_pop_col = NULL,
  popl_var_prefix
)
}
\arguments{
\item{profiled_sf}{Profiled (a simple features object)}

\item{sp_unit}{PARAM_DESCRIPTION}

\item{data_year}{PARAM_DESCRIPTION}

\item{concept}{PARAM_DESCRIPTION}

\item{age_sex_var_name}{PARAM_DESCRIPTION, Default: NULL}

\item{age_sex_pop_resolution}{PARAM_DESCRIPTION, Default: NULL}

\item{tot_pop_col}{PARAM_DESCRIPTION, Default: NULL}

\item{popl_var_prefix}{PARAM_DESCRIPTION}
}
\value{
Profiled (a simple features object)
}
\description{
update_pop_by_inc_area() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update pop by include area. Function argument profiled_sf specifies the object to be updated. Argument sp_unit provides the object to be updated. The function returns Profiled (a simple features object).
}
\keyword{internal}
