% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_counts_by_area.R
\name{update_pop_count_by_areas}
\alias{update_pop_count_by_areas}
\title{Adjust population counts by fraction of spatial unit included in a profiled area.}
\usage{
update_pop_count_by_areas(profiled_sf, group_by_var, age_sex_var_name,
  data_year, age_sex_pop_resolution, tot_pop_resolution)
}
\arguments{
\item{profiled_sf}{A simple features object comprised of SA2s.}

\item{group_by_var}{A string specifying "SA1" or "SA2" as the resolution of the
returned object,}

\item{tot_pop_sf}{A simple features object comprised of SA1s}
}
\value{
A simple features object.
}
\description{
This function:
 -
 -
}
\details{
Need to review if removal of linestrings is appropriate.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{join}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}},\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{select_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{funs}},\code{\link[dplyr]{summarise_all}}
 \code{\link[sf]{st_geometry}},\code{\link[sf]{geos_measures}}
}
