% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_popl_by_group}
\alias{update_popl_by_group}
\title{Update population by group}
\usage{
update_popl_by_group(
  profiled_sf,
  data_year_1L_chr,
  dynamic_var_nm_1L_chr,
  dynamic_var_rsl_1L_chr,
  group_by_var_1L_chr,
  reference_var_rsl_1L_chr,
  featured_var_pfx_1L_chr = "",
  reference_vals_chr
)
}
\arguments{
\item{profiled_sf}{Profiled (a simple features object)}

\item{data_year_1L_chr}{Data year (a character vector of length one)}

\item{dynamic_var_nm_1L_chr}{Dynamic variable name (a character vector of length one)}

\item{dynamic_var_rsl_1L_chr}{Dynamic variable resolution (a character vector of length one)}

\item{group_by_var_1L_chr}{Group by variable (a character vector of length one)}

\item{reference_var_rsl_1L_chr}{Reference variable resolution (a character vector of length one)}

\item{featured_var_pfx_1L_chr}{Featured variable prefix (a character vector of length one), Default: ''}

\item{reference_vals_chr}{Reference values (a character vector)}
}
\value{
Profiled (a simple features object)
}
\description{
update_popl_by_group() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update population by group. The function returns Profiled (a simple features object).
}
\keyword{internal}
