% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_popl_by_incld_area}
\alias{update_popl_by_incld_area}
\title{Update population by included area}
\usage{
update_popl_by_incld_area(
  profiled_sf,
  concept_1L_chr,
  data_year_1L_chr,
  featured_var_pfx_1L_chr,
  spatial_unit_1L_chr,
  dynamic_var_nm_1L_chr = NULL,
  dynamic_var_rsl_1L_chr = NULL,
  reference_var_nm_1L_chr = NULL
)
}
\arguments{
\item{profiled_sf}{Profiled (a simple features object)}

\item{concept_1L_chr}{Concept (a character vector of length one)}

\item{data_year_1L_chr}{Data year (a character vector of length one)}

\item{featured_var_pfx_1L_chr}{Featured variable prefix (a character vector of length one)}

\item{spatial_unit_1L_chr}{Spatial unit (a character vector of length one)}

\item{dynamic_var_nm_1L_chr}{Dynamic variable name (a character vector of length one), Default: NULL}

\item{dynamic_var_rsl_1L_chr}{Dynamic variable resolution (a character vector of length one), Default: NULL}

\item{reference_var_nm_1L_chr}{Reference variable name (a character vector of length one), Default: NULL}
}
\value{
Profiled (a simple features object)
}
\description{
update_popl_by_incld_area() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update population by included area. The function returns Profiled (a simple features object).
}
\keyword{internal}
