% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_sf_boundary_descr}
\alias{update_sf_boundary_descr}
\title{Update simple features object boundary descr}
\usage{
update_sf_boundary_descr(look_up_ref, one_cluster_up_to_xmin_list)
}
\arguments{
\item{look_up_ref}{PARAM_DESCRIPTION}

\item{one_cluster_up_to_xmin_list}{PARAM_DESCRIPTION}
}
\value{
NA ()
}
\description{
update_sf_boundary_descr() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update simple features object boundary descr. Function argument look_up_ref specifies the object to be updated. Argument one_cluster_up_to_xmin_list provides the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
