% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_sp_data_list}
\alias{update_sp_data_list}
\title{Update sp data list}
\usage{
update_sp_data_list(sp_data_list, input_ls, profiled_area_bands_list)
}
\arguments{
\item{sp_data_list}{PARAM_DESCRIPTION}

\item{input_ls}{Input (a list)}

\item{profiled_area_bands_list}{PARAM_DESCRIPTION}
}
\value{
NA ()
}
\description{
update_sp_data_list() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update sp data list. Function argument sp_data_list specifies the object to be updated. Argument input_ls provides the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
