% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_sp_local_proc_r4}
\alias{update_sp_local_proc_r4}
\title{Update sp local process}
\usage{
update_sp_local_proc_r4(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\description{
update_sp_local_proc_r4() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update sp local process ready4 s4. Function argument x specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
