% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_use_coord_lup.R
\name{use_coord_lup<-}
\alias{use_coord_lup<-}
\alias{use_coord_lup<--ready4_profiled_area}
\title{use_coord_lup<-}
\usage{
use_coord_lup(x) <- value

\S4method{use_coord_lup}{ready4_profiled_area}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_profiled_area}
}
\description{
S4 Generic function to set the value of the slot use_coord_lup

Set the value of the slot use_coord_lup for S4 objects of class ready4_profiled_area
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
