% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_profiled_area.R,
%   R/gnrc_use_coord_lup.R
\name{use_coord_lup-ready4_profiled_area}
\alias{use_coord_lup-ready4_profiled_area}
\alias{use_coord_lup}
\title{use_coord_lup}
\arguments{
\item{x}{An object of class ready4_profiled_area}

\item{x}{An object}
}
\description{
Get the value of the slot use_coord_lup for S4 objects of class ready4_profiled_area

S4 Generic function to get the value of the slot use_coord_lup
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
