% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_use_coord_lup.R
\name{use_coord_lup<-}
\alias{use_coord_lup<-}
\alias{use_coord_lup<--ready4_profiled_area}
\alias{use_coord_lup<-,ready4_profiled_area-method}
\title{use_coord_lup<-}
\usage{
use_coord_lup(x) <- value

\S4method{use_coord_lup}{ready4_profiled_area}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_profiled_area}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot use_coord_lup

Set the value of the slot use_coord_lup for S4 objects of class ready4_profiled_area
}
