% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_validate.R
\name{validate_popl_predns_incld}
\alias{validate_popl_predns_incld}
\title{Validate population preductions included}
\usage{
validate_popl_predns_incld(data_items_chr, data_lookup_tb, key_var_1L_chr)
}
\arguments{
\item{data_items_chr}{Data items (a character vector)}

\item{data_lookup_tb}{Data lookup (a tibble)}

\item{key_var_1L_chr}{Key variable (a character vector of length one)}
}
\value{
Contains population preductions (a logical vector of length one)
}
\description{
validate_popl_predns_incld() is a Validate function that validates that an object conforms to required criteria. Specifically, this function implements an algorithm to validate population preductions included. The function returns Contains population preductions (a logical vector of length one).
}
\keyword{internal}
