% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_abbreviations_lup.R
\name{validate_ready4_sp_abbreviations_lup}
\alias{validate_ready4_sp_abbreviations_lup}
\title{Validate Readyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations.}
\usage{
validate_ready4_sp_abbreviations_lup(x)
}
\arguments{
\item{x}{An unvalidated instance of the Readyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations.}
}
\value{
A prototpe for Readyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations.
}
\description{
Validate an instance of the Readyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table for spatial data abbreviations.
}
