% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_data_pack_lup.R
\name{validate_ready4_sp_data_pack_lup}
\alias{validate_ready4_sp_data_pack_lup}
\title{Validate Readyforwhatsnext S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).}
\usage{
validate_ready4_sp_data_pack_lup(x)
}
\arguments{
\item{x}{An unvalidated instance of the Readyforwhatsnext S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).}
}
\value{
A prototpe for Readyforwhatsnext S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).
}
\description{
Validate an instance of the Readyforwhatsnext S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).
}
