% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_import_lup.R
\name{validate_ready4_sp_import_lup}
\alias{validate_ready4_sp_import_lup}
\title{Validate Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.}
\usage{
validate_ready4_sp_import_lup(x)
}
\arguments{
\item{x}{An unvalidated instance of the Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.}
}
\value{
A prototpe for Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.
}
\description{
Validate an instance of the Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.
}
