% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_resolution_lup.R
\name{validate_ready4_sp_resolution_lup}
\alias{validate_ready4_sp_resolution_lup}
\title{Validate Readyforwhatsnext S3 class for tibble object lookup table of the relative resolutions of different spatial objects.}
\usage{
validate_ready4_sp_resolution_lup(x)
}
\arguments{
\item{x}{An unvalidated instance of the Readyforwhatsnext S3 class for tibble object lookup table of the relative resolutions of different spatial objects.}
}
\value{
A prototpe for Readyforwhatsnext S3 class for tibble object lookup table of the relative resolutions of different spatial objects.
}
\description{
Validate an instance of the Readyforwhatsnext S3 class for tibble object lookup table of the relative resolutions of different spatial objects.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of the relative resolutions of different spatial objects.
}
