% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_site_coord_lup.R
\name{validate_ready4_sp_site_coord_lup}
\alias{validate_ready4_sp_site_coord_lup}
\title{Validate Readyforwhatsnext S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.}
\usage{
validate_ready4_sp_site_coord_lup(x)
}
\arguments{
\item{x}{An unvalidated instance of the Readyforwhatsnext S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.}
}
\value{
A prototpe for Readyforwhatsnext S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.
}
\description{
Validate an instance of the Readyforwhatsnext S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.
}
