% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_starter_sf_lup.R
\name{validate_ready4_sp_starter_sf_lup}
\alias{validate_ready4_sp_starter_sf_lup}
\title{Validate Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.}
\usage{
validate_ready4_sp_starter_sf_lup(x)
}
\arguments{
\item{x}{An unvalidated instance of the Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.}
}
\value{
A prototpe for Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
\description{
Validate an instance of the Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
