% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_vicinity_identifiers.R
\name{validate_vicinity_identifiers}
\alias{validate_vicinity_identifiers}
\title{validate vicinity identifiers ready4 submodule class for tibble object lookup table of unique feature identifiers used for different spatial objects.}
\usage{
validate_vicinity_identifiers(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 submodule class for tibble object lookup table of unique feature identifiers used for different spatial objects.}
}
\value{
A prototpe for ready4 submodule class for tibble object lookup table of unique feature identifiers used for different spatial objects.
}
\description{
Validate an instance of the ready4 submodule class for tibble object lookup table of unique feature identifiers used for different spatial objects.
}
\details{
ready4 submodule for tibble object lookup table of unique feature identifiers used for different spatial objects.
}
\keyword{internal}
