% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_vicinity_mapes.R
\name{validate_vicinity_mapes}
\alias{validate_vicinity_mapes}
\title{validate vicinity mean absolute prediction errors ready4 submodule class for tibble object that stores spatial simulation parameters relating to Mean Absolute Prediction Errors.}
\usage{
validate_vicinity_mapes(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 submodule class for tibble object that stores spatial simulation parameters relating to Mean Absolute Prediction Errors.}
}
\value{
A prototpe for ready4 submodule class for tibble object that stores spatial simulation parameters relating to Mean Absolute Prediction Errors.
}
\description{
Validate an instance of the ready4 submodule class for tibble object that stores spatial simulation parameters relating to Mean Absolute Prediction Errors.
}
\details{
ready4 submodule for tibble object that stores spatial simulation parameters relating to Mean Absolute Prediction Errors.
}
\keyword{internal}
