% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_vicinity_parameters.R
\name{validate_vicinity_parameters}
\alias{validate_vicinity_parameters}
\title{validate vicinity parameters ready4 submodule class for tibble object that stores simulation structural parameters relating to the spatial environment.}
\usage{
validate_vicinity_parameters(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 submodule class for tibble object that stores simulation structural parameters relating to the spatial environment.}
}
\value{
A prototpe for ready4 submodule class for tibble object that stores simulation structural parameters relating to the spatial environment.
}
\description{
Validate an instance of the ready4 submodule class for tibble object that stores simulation structural parameters relating to the spatial environment.
}
\details{
ready4 submodule for tibble object that stores simulation structural parameters relating to the spatial environment.
}
\keyword{internal}
