% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_vicinity_points.R
\name{validate_vicinity_points}
\alias{validate_vicinity_points}
\title{validate vicinity points ready4 submodule class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.}
\usage{
validate_vicinity_points(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 submodule class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.}
}
\value{
A prototpe for ready4 submodule class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.
}
\description{
Validate an instance of the ready4 submodule class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.
}
\details{
ready4 submodule for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.
}
\keyword{internal}
