% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_vicinity_processed.R
\name{validate_vicinity_processed}
\alias{validate_vicinity_processed}
\title{validate vicinity processed ready4 S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).}
\usage{
validate_vicinity_processed(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).}
}
\value{
A prototpe for ready4 S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).
}
\description{
Validate an instance of the ready4 S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).
}
\details{
ready4 S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data).
}
\keyword{internal}
