% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_vicinity_template.R
\name{validate_vicinity_template}
\alias{validate_vicinity_template}
\title{validate vicinity template ready4 S3 class for tibble object lookup table for base file used in creation of certain spatial objects.}
\usage{
validate_vicinity_template(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 S3 class for tibble object lookup table for base file used in creation of certain spatial objects.}
}
\value{
A prototpe for ready4 S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
\description{
Validate an instance of the ready4 S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
\details{
ready4 S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
\keyword{internal}
