% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_vicinity_values.R
\name{validate_vicinity_values}
\alias{validate_vicinity_values}
\title{validate vicinity values ready4 S3 class for tibble object that stores simulation parameter values for each iteration.}
\usage{
validate_vicinity_values(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 S3 class for tibble object that stores simulation parameter values for each iteration.}
}
\value{
A prototpe for ready4 S3 class for tibble object that stores simulation parameter values for each iteration.
}
\description{
Validate an instance of the ready4 S3 class for tibble object that stores simulation parameter values for each iteration.
}
\details{
ready4 S3 class for tibble object that stores simulation parameter values for each iteration.
}
\keyword{internal}
