% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_vicinity_abbreviations.R
\name{vicinity_abbreviations}
\alias{vicinity_abbreviations}
\alias{make_pt_vicinity_abbreviations}
\alias{is_vicinity_abbreviations}
\title{ready4 S3 class for tibble object lookup table for spatial data abbreviations.}
\usage{
vicinity_abbreviations(x = make_pt_vicinity_abbreviations())

make_pt_vicinity_abbreviations(
  long_name_chr = character(0),
  short_name_chr = character(0),
  type_chr = character(0)
)

is_vicinity_abbreviations(x)
}
\arguments{
\item{x}{An object of any type}

\item{long_name_chr}{Long name (a character vector), Default: character(0)}

\item{short_name_chr}{Short name (a character vector), Default: character(0)}

\item{type_chr}{Type (a character vector), Default: character(0)}
}
\value{
A validated instance of the ready4 S3 class for tibble object lookup table for spatial data abbreviations.

A prototype for ready4 S3 class for tibble object lookup table for spatial data abbreviations.

A logical value, TRUE if a valid instance of the ready4 S3 class for tibble object lookup table for spatial data abbreviations.
}
\description{
Create a new valid instance of the ready4 S3 class for tibble object lookup table for spatial data abbreviations.
}
\details{
ready4 S3 class for tibble object lookup table for spatial data abbreviations.
}
