% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_vicinity_identifiers.R
\name{vicinity_identifiers}
\alias{vicinity_identifiers}
\alias{make_pt_vicinity_identifiers}
\alias{is_vicinity_identifiers}
\title{ready4 submodule class for tibble object lookup table of unique feature identifiers used for different spatial objects.}
\usage{
vicinity_identifiers(x = make_pt_vicinity_identifiers())

make_pt_vicinity_identifiers(
  spatial_unit_chr = character(0),
  year_chr = character(0),
  var_name_chr = character(0)
)

is_vicinity_identifiers(x)
}
\arguments{
\item{x}{An object of any type}

\item{spatial_unit_chr}{Spatial unit (a character vector), Default: character(0)}

\item{year_chr}{Year (a character vector), Default: character(0)}

\item{var_name_chr}{Variable name (a character vector), Default: character(0)}
}
\value{
A validated instance of the ready4 submodule class for tibble object lookup table of unique feature identifiers used for different spatial objects.

A prototype for ready4 submodule class for tibble object lookup table of unique feature identifiers used for different spatial objects.

A logical value, TRUE if a valid instance of the ready4 submodule class for tibble object lookup table of unique feature identifiers used for different spatial objects.
}
\description{
Create a new valid instance of the ready4 submodule class for tibble object lookup table of unique feature identifiers used for different spatial objects.
}
\details{
ready4 submodule for tibble object lookup table of unique feature identifiers used for different spatial objects.
}
