% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_vicinity_parameters.R
\name{vicinity_parameters}
\alias{vicinity_parameters}
\alias{make_pt_vicinity_parameters}
\alias{is_vicinity_parameters}
\title{ready4 S3 class for tibble object that stores simulation structural parameters relating to the spatial environment.}
\usage{
vicinity_parameters(x = make_pt_vicinity_parameters())

make_pt_vicinity_parameters(
  param_name_chr = character(0),
  deterministic_val_dbl = numeric(0),
  distribution_chr = character(0),
  dstr_param_1_dbl = numeric(0),
  dstr_param_2_dbl = numeric(0),
  dstr_param_3_dbl = numeric(0),
  dstr_param_4_dbl = numeric(0),
  transformation_chr = character(0),
  use_in_chr = character(0),
  source_chr = character(0)
)

is_vicinity_parameters(x)
}
\arguments{
\item{x}{An object of any type}

\item{param_name_chr}{Parameter name (a character vector), Default: character(0)}

\item{deterministic_val_dbl}{Deterministic value (a double vector), Default: numeric(0)}

\item{distribution_chr}{Distribution (a character vector), Default: character(0)}

\item{dstr_param_1_dbl}{Distribution parameter 1 (a double vector), Default: numeric(0)}

\item{dstr_param_2_dbl}{Distribution parameter 2 (a double vector), Default: numeric(0)}

\item{dstr_param_3_dbl}{Distribution parameter 3 (a double vector), Default: numeric(0)}

\item{dstr_param_4_dbl}{Distribution parameter 4 (a double vector), Default: numeric(0)}

\item{transformation_chr}{Transformation (a character vector), Default: character(0)}

\item{use_in_chr}{Use in (a character vector), Default: character(0)}

\item{source_chr}{Source (a character vector), Default: character(0)}
}
\value{
A validated instance of the ready4 S3 class for tibble object that stores simulation structural parameters relating to the spatial environment.

A prototype for ready4 S3 class for tibble object that stores simulation structural parameters relating to the spatial environment.

A logical value, TRUE if a valid instance of the ready4 S3 class for tibble object that stores simulation structural parameters relating to the spatial environment.
}
\description{
Create a new valid instance of the ready4 S3 class for tibble object that stores simulation structural parameters relating to the spatial environment.
}
\details{
ready4 S3 class for tibble object that stores simulation structural parameters relating to the spatial environment.
}
