% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_vicinity_points.R
\name{vicinity_points}
\alias{vicinity_points}
\alias{make_pt_vicinity_points}
\alias{is_vicinity_points}
\title{ready4 S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.}
\usage{
vicinity_points(x = make_pt_vicinity_points())

make_pt_vicinity_points(
  service_type_chr = character(0),
  cluster_name_chr = character(0),
  service_name_chr = character(0),
  lat_dbl = numeric(0),
  lng_dbl = numeric(0)
)

is_vicinity_points(x)
}
\arguments{
\item{x}{An object of any type}

\item{service_type_chr}{Service type (a character vector), Default: character(0)}

\item{cluster_name_chr}{Cluster name (a character vector), Default: character(0)}

\item{service_name_chr}{Service name (a character vector), Default: character(0)}

\item{lat_dbl}{Lat (a double vector), Default: numeric(0)}

\item{lng_dbl}{Longitude (a double vector), Default: numeric(0)}
}
\value{
A validated instance of the ready4 S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.

A prototype for ready4 S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.

A logical value, TRUE if a valid instance of the ready4 S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.
}
\description{
Create a new valid instance of the ready4 S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.
}
\details{
ready4 S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes.
}
