% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_vicinity_raw.R
\name{vicinity_raw}
\alias{vicinity_raw}
\alias{make_pt_vicinity_raw}
\alias{is_vicinity_raw}
\title{ready4 S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.}
\usage{
vicinity_raw(x = make_pt_vicinity_raw())

make_pt_vicinity_raw(
  file_type_chr = character(0),
  file_name_chr = character(0),
  data_repo_chr = character(0),
  data_repo_ui_chr = character(0),
  data_repo_db_ui_chr = character(0),
  data_repo_file_ext_chr = character(0),
  data_repo_save_type_chr = character(0),
  local_file_src_chr = character(0),
  path_to_make_script_chr = character(0),
  download_url_chr = character(0),
  inc_file_main_chr = character(0),
  inc_fls_to_rename_ls = list(),
  new_nms_for_inc_fls_ls = list(),
  name = character(0),
  country_chr = character(0),
  area_type_chr = character(0),
  area_bndy_yr_chr = character(0),
  region_chr = character(0),
  data_type_chr = character(0),
  main_feature_chr = character(0),
  year_chr = character(0),
  year_start_chr = character(0),
  year_end_chr = character(0),
  uid_chr = character(0),
  add_boundaries_chr = list()
)

is_vicinity_raw(x)
}
\arguments{
\item{x}{An object of any type}

\item{file_type_chr}{File type (a character vector), Default: character(0)}

\item{file_name_chr}{File name (a character vector), Default: character(0)}

\item{data_repo_chr}{Data repository (a character vector), Default: character(0)}

\item{data_repo_ui_chr}{Data repository user interface (a character vector), Default: character(0)}

\item{data_repo_db_ui_chr}{Data repository database user interface (a character vector), Default: character(0)}

\item{data_repo_file_ext_chr}{Data repository file extension (a character vector), Default: character(0)}

\item{data_repo_save_type_chr}{Data repository save type (a character vector), Default: character(0)}

\item{local_file_src_chr}{Local file source (a character vector), Default: character(0)}

\item{path_to_make_script_chr}{Path to make script (a character vector), Default: character(0)}

\item{download_url_chr}{Download url (a character vector), Default: character(0)}

\item{inc_file_main_chr}{Include file main (a character vector), Default: character(0)}

\item{inc_fls_to_rename_ls}{Include files to rename (a list), Default: list()}

\item{new_nms_for_inc_fls_ls}{New names for include files (a list), Default: list()}

\item{name}{PARAM_DESCRIPTION, Default: character(0)}

\item{country_chr}{Country (a character vector), Default: character(0)}

\item{area_type_chr}{Area type (a character vector), Default: character(0)}

\item{area_bndy_yr_chr}{Area boundary year (a character vector), Default: character(0)}

\item{region_chr}{Region (a character vector), Default: character(0)}

\item{data_type_chr}{Data type (a character vector), Default: character(0)}

\item{main_feature_chr}{Main feature (a character vector), Default: character(0)}

\item{year_chr}{Year (a character vector), Default: character(0)}

\item{year_start_chr}{Year start (a character vector), Default: character(0)}

\item{year_end_chr}{Year end (a character vector), Default: character(0)}

\item{uid_chr}{Unique identifier (a character vector), Default: character(0)}

\item{add_boundaries_chr}{Add boundaries (a character vector), Default: list()}
}
\value{
A validated instance of the ready4 S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.

A prototype for ready4 S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.

A logical value, TRUE if a valid instance of the ready4 S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.
}
\description{
Create a new valid instance of the ready4 S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.
}
\details{
ready4 S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.
}
