% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_vicinity_resolutions.R
\name{vicinity_resolutions}
\alias{vicinity_resolutions}
\alias{make_pt_vicinity_resolutions}
\alias{is_vicinity_resolutions}
\title{ready4 S3 class for tibble object lookup table of the relative resolutions of different spatial objects.}
\usage{
vicinity_resolutions(x = make_pt_vicinity_resolutions())

make_pt_vicinity_resolutions(
  parent_area_chr = character(0),
  boundary_year_dbl = numeric(0),
  area_type_chr = character(0),
  area_count_chr_dbl = numeric(0),
  complete_lgl = logical(0),
  summed_area_dbl = numeric(0),
  mean_size_dbl = numeric(0)
)

is_vicinity_resolutions(x)
}
\arguments{
\item{x}{An object of any type}

\item{parent_area_chr}{Parent area (a character vector), Default: character(0)}

\item{boundary_year_dbl}{Boundary year (a double vector), Default: numeric(0)}

\item{area_type_chr}{Area type (a character vector), Default: character(0)}

\item{area_count_chr_dbl}{Area count character vector (a double vector), Default: numeric(0)}

\item{complete_lgl}{Complete (a logical vector), Default: logical(0)}

\item{summed_area_dbl}{Summed area (a double vector), Default: numeric(0)}

\item{mean_size_dbl}{Mean size (a double vector), Default: numeric(0)}
}
\value{
A validated instance of the ready4 S3 class for tibble object lookup table of the relative resolutions of different spatial objects.

A prototype for ready4 S3 class for tibble object lookup table of the relative resolutions of different spatial objects.

A logical value, TRUE if a valid instance of the ready4 S3 class for tibble object lookup table of the relative resolutions of different spatial objects.
}
\description{
Create a new valid instance of the ready4 S3 class for tibble object lookup table of the relative resolutions of different spatial objects.
}
\details{
ready4 S3 class for tibble object lookup table of the relative resolutions of different spatial objects.
}
