% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_vicinity_template.R
\name{vicinity_template}
\alias{vicinity_template}
\alias{make_pt_vicinity_template}
\alias{is_vicinity_template}
\title{ready4 S3 class for tibble object lookup table for base file used in creation of certain spatial objects.}
\usage{
vicinity_template(x = make_pt_vicinity_template())

make_pt_vicinity_template(
  country_chr = character(0),
  area_type_chr = character(0),
  area_bndy_yr_chr = character(0),
  starter_sf = character(0),
  subdivision_chr = character(0)
)

is_vicinity_template(x)
}
\arguments{
\item{x}{An object of any type}

\item{country_chr}{Country (a character vector), Default: character(0)}

\item{area_type_chr}{Area type (a character vector), Default: character(0)}

\item{area_bndy_yr_chr}{Area boundary year (a character vector), Default: character(0)}

\item{starter_sf}{Starter (a simple features object), Default: character(0)}

\item{subdivision_chr}{Subdivision (a character vector), Default: character(0)}
}
\value{
A validated instance of the ready4 S3 class for tibble object lookup table for base file used in creation of certain spatial objects.

A prototype for ready4 S3 class for tibble object lookup table for base file used in creation of certain spatial objects.

A logical value, TRUE if a valid instance of the ready4 S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
\description{
Create a new valid instance of the ready4 S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
\details{
ready4 S3 class for tibble object lookup table for base file used in creation of certain spatial objects.
}
