% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_vicinity_values.R
\name{vicinity_values}
\alias{vicinity_values}
\alias{make_pt_vicinity_values}
\alias{is_vicinity_values}
\title{ready4 submodule class for tibble object that stores simulation parameter values for each iteration.}
\usage{
vicinity_values(x = make_pt_vicinity_values())

make_pt_vicinity_values(
  param_name_chr = character(0),
  iteration_1_dbl = numeric(0)
)

is_vicinity_values(x)
}
\arguments{
\item{x}{An object of any type}

\item{param_name_chr}{Parameter name (a character vector), Default: character(0)}

\item{iteration_1_dbl}{Iteration 1 (a double vector), Default: numeric(0)}
}
\value{
A validated instance of the ready4 submodule class for tibble object that stores simulation parameter values for each iteration.

A prototype for ready4 submodule class for tibble object that stores simulation parameter values for each iteration.

A logical value, TRUE if a valid instance of the ready4 submodule class for tibble object that stores simulation parameter values for each iteration.
}
\description{
Create a new valid instance of the ready4 submodule class for tibble object that stores simulation parameter values for each iteration.
}
\details{
ready4 submodule for tibble object that stores simulation parameter values for each iteration.
}
