% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_VicinityLocal.R
\docType{class}
\name{VicinityLocal-class}
\alias{VicinityLocal-class}
\alias{VicinityLocal}
\title{VicinityLocal}
\description{
Object defining data to be saved in local directory.
}
\section{Slots}{

\describe{
\item{\code{a_VicinityLookup}}{(an instance of the VicinityLookup class)}

\item{\code{merge_itms_chr}}{Merge items (a character vector)}

\item{\code{raw_fls_dir_1L_chr}}{Raw files directory (a character vector of length one)}

\item{\code{pkg_1L_chr}}{Package (a character vector of length one)}

\item{\code{overwrite_1L_lgl}}{Overwrite (a logical vector of length one)}

\item{\code{write_1L_lgl}}{Write (a logical vector of length one)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

