% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_VicinityMacro.R
\docType{class}
\name{VicinityMacro-class}
\alias{VicinityMacro-class}
\alias{VicinityMacro}
\title{VicinityMacro}
\description{
Macro level context
}
\section{Slots}{

\describe{
\item{\code{a_VicinityLookup}}{(an instance of the VicinityLookup class)}

\item{\code{global_region_chr}}{Global region (a character vector)}

\item{\code{country_chr}}{Country (a character vector)}

\item{\code{country_bndy_yr_dbl}}{Country boundary year (a double vector)}

\item{\code{crs_dbl}}{Coordinates reference system (a double vector)}

\item{\code{temporal_min_dtm}}{Temporal minimum (a date vector)}

\item{\code{temporal_max_dtm}}{Temporal maximum (a date vector)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

