% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_VicinityMesoArea.R
\docType{class}
\name{VicinityMesoArea-class}
\alias{VicinityMesoArea-class}
\alias{VicinityMesoArea}
\title{VicinityMesoArea}
\description{
Meso level context - area
}
\section{Slots}{

\describe{
\item{\code{area_type_chr}}{Area type (a character vector)}

\item{\code{area_chr}}{Area (a character vector)}

\item{\code{area_bndy_yr_dbl}}{Area boundary year (a double vector)}

\item{\code{region_type_chr}}{Region type (a character vector)}

\item{\code{region_chr}}{Region (a character vector)}

\item{\code{region_bndy_yr_dbl}}{Region boundary year (a double vector)}

\item{\code{a_VicinityLookup}}{(an instance of the VicinityLookup class)}

\item{\code{global_region_chr}}{Global region (a character vector)}

\item{\code{country_chr}}{Country (a character vector)}

\item{\code{country_bndy_yr_dbl}}{Country boundary year (a double vector)}

\item{\code{crs_dbl}}{Coordinates reference system (a double vector)}

\item{\code{temporal_min_dtm}}{Temporal minimum (a date vector)}

\item{\code{temporal_max_dtm}}{Temporal maximum (a date vector)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

