% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_VicinityMicro.R
\docType{class}
\name{VicinityMicro-class}
\alias{VicinityMicro-class}
\alias{VicinityMicro}
\title{VicinityMicro}
\description{
Micro level context
}
\section{Slots}{

\describe{
\item{\code{geom_dist_km_cuts_dbl}}{Geometry distance kilometre cuts (a double vector)}

\item{\code{travel_time_mins_cuts_dbl}}{Travel time minimums cuts (a double vector)}

\item{\code{travel_mode_chr}}{Travel mode (a character vector)}

\item{\code{area_type_chr}}{Area type (a character vector)}

\item{\code{area_chr}}{Area (a character vector)}

\item{\code{area_bndy_yr_dbl}}{Area boundary year (a double vector)}

\item{\code{region_type_chr}}{Region type (a character vector)}

\item{\code{region_chr}}{Region (a character vector)}

\item{\code{region_bndy_yr_dbl}}{Region boundary year (a double vector)}

\item{\code{a_VicinityLookup}}{(an instance of the VicinityLookup class)}

\item{\code{global_region_chr}}{Global region (a character vector)}

\item{\code{country_chr}}{Country (a character vector)}

\item{\code{country_bndy_yr_dbl}}{Country boundary year (a double vector)}

\item{\code{crs_dbl}}{Crs (a double vector)}

\item{\code{temporal_min_dtm}}{Temporal minimum (a date vector)}

\item{\code{temporal_max_dtm}}{Temporal maximum (a date vector)}
}}

