% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_VicinityProfile.R
\docType{class}
\name{VicinityProfile-class}
\alias{VicinityProfile-class}
\alias{VicinityProfile}
\title{VicinityProfile}
\description{
Information to create a profiled area object
}
\section{Slots}{

\describe{
\item{\code{features_chr}}{Features (a character vector)}

\item{\code{use_coord_lup_lgl}}{Use coord lookup table (a logical vector)}

\item{\code{geomc_dist_limit_km_dbl}}{Geometric distance limit kilometre (a double vector)}

\item{\code{drive_time_limit_mins_dbl}}{Drive time limit minimums (a double vector)}

\item{\code{nbr_bands_dbl}}{Number bands (a double vector)}

\item{\code{data_year_chr}}{Data year (a character vector)}

\item{\code{data_ymds_dtm}}{Data ymds (a date vector)}

\item{\code{geom_dist_km_cuts_dbl}}{Geometry distance kilometre cuts (a double vector)}

\item{\code{travel_time_mins_cuts_dbl}}{Travel time minimums cuts (a double vector)}

\item{\code{travel_mode_chr}}{Travel mode (a character vector)}

\item{\code{area_type_chr}}{Area type (a character vector)}

\item{\code{area_chr}}{Area (a character vector)}

\item{\code{area_bndy_yr_dbl}}{Area boundary year (a double vector)}

\item{\code{region_type_chr}}{Region type (a character vector)}

\item{\code{region_chr}}{Region (a character vector)}

\item{\code{region_bndy_yr_dbl}}{Region boundary year (a double vector)}

\item{\code{a_VicinityLookup}}{(an instance of the VicinityLookup class)}

\item{\code{global_region_chr}}{Global region (a character vector)}

\item{\code{country_chr}}{Country (a character vector)}

\item{\code{country_bndy_yr_dbl}}{Country boundary year (a double vector)}

\item{\code{crs_dbl}}{Crs (a double vector)}

\item{\code{temporal_min_dtm}}{Temporal minimum (a date vector)}

\item{\code{temporal_max_dtm}}{Temporal maximum (a date vector)}
}}

