% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_att_tb}
\alias{write_att_tb}
\title{Write attribute tibble}
\usage{
write_att_tb(
  att_tb,
  object_nm_1L_chr,
  processed_fls_dir_1L_chr,
  overwrite_1L_lgl
)
}
\arguments{
\item{att_tb}{Attribute (a tibble)}

\item{object_nm_1L_chr}{Object name (a character vector of length one)}

\item{processed_fls_dir_1L_chr}{Processed files directory (a character vector of length one)}

\item{overwrite_1L_lgl}{Overwrite (a logical vector of length one)}
}
\value{
No return value, called for side effects.
}
\description{
write_att_tb() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write attribute tibble. The function is called for its side effects and does not return a value.
}
\keyword{internal}
