% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_attr_tb}
\alias{write_attr_tb}
\title{Write attr}
\usage{
write_attr_tb(attr_tb, obj_name, processed_dir, overwrite_lgl)
}
\arguments{
\item{attr_tb}{Attr (a tibble)}

\item{obj_name}{PARAM_DESCRIPTION}

\item{processed_dir}{PARAM_DESCRIPTION}

\item{overwrite_lgl}{Overwrite (a logical vector)}
}
\description{
write_attr_tb() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write attr tibble. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
