% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_dirs_for_imp}
\alias{write_dirs_for_imp}
\title{Write directories for import}
\usage{
write_dirs_for_imp(paths_chr)
}
\arguments{
\item{paths_chr}{Paths (a character vector)}
}
\value{
No return value, called for side effects.
}
\description{
write_dirs_for_imp() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write directories for import. The function is called for its side effects and does not return a value.
}
\keyword{internal}
