% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_fls_and_mk_sngl_row_data_lup}
\alias{write_fls_and_mk_sngl_row_data_lup}
\title{Write files and make sngl row data}
\usage{
write_fls_and_mk_sngl_row_data_lup(
  x,
  merge_with,
  pckg_name,
  raw_data_dir,
  processed_dir,
  crs_nbr_vec = NA_real_,
  overwrite_lgl = F
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{merge_with}{PARAM_DESCRIPTION}

\item{pckg_name}{PARAM_DESCRIPTION}

\item{raw_data_dir}{PARAM_DESCRIPTION}

\item{processed_dir}{PARAM_DESCRIPTION}

\item{crs_nbr_vec}{PARAM_DESCRIPTION, Default: NA}

\item{overwrite_lgl}{Overwrite (a logical vector), Default: F}
}
\description{
write_fls_and_mk_sngl_row_data_lup() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write files and make sngl row data lookup table. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
