% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_fls_for_imp}
\alias{write_fls_for_imp}
\title{Write files for import}
\usage{
write_fls_for_imp(
  x,
  data_lookup_ref,
  lookup_variable,
  directory_path,
  overwrite_lgl = F
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{data_lookup_ref}{PARAM_DESCRIPTION}

\item{lookup_variable}{PARAM_DESCRIPTION}

\item{directory_path}{PARAM_DESCRIPTION}

\item{overwrite_lgl}{Overwrite (a logical vector), Default: F}
}
\description{
write_fls_for_imp() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write files for import. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
