% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_fls_from_imp_and_upd_r4}
\alias{write_fls_from_imp_and_upd_r4}
\title{Write files from import and update}
\usage{
write_fls_from_imp_and_upd_r4(x, processed_dir_chr, crs_nbr_vec)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{processed_dir_chr}{Processed directory (a character vector)}

\item{crs_nbr_vec}{PARAM_DESCRIPTION}
}
\description{
write_fls_from_imp_and_upd_r4() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write files from import and update ready4 s4. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
