% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_fls_from_local_imp}
\alias{write_fls_from_local_imp}
\title{Write files from local import}
\usage{
write_fls_from_local_imp(x, raw_data_dir_chr, save_lgl)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{raw_data_dir_chr}{Raw data directory (a character vector)}

\item{save_lgl}{Save (a logical vector)}
}
\description{
write_fls_from_local_imp() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write files from local import. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
