% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_fls_from_sp_imp_and_upd_imp_ls}
\alias{write_fls_from_sp_imp_and_upd_imp_ls}
\title{Write files from sp import and update import}
\usage{
write_fls_from_sp_imp_and_upd_imp_ls(x, crs_nbr_vec, return_r4_lgl = T)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{crs_nbr_vec}{PARAM_DESCRIPTION}

\item{return_r4_lgl}{Return ready4 S4 (a logical vector), Default: T}
}
\description{
write_fls_from_sp_imp_and_upd_imp_ls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write files from sp import and update import list. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
