% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_procsd_geom_imp}
\alias{write_procsd_geom_imp}
\title{Write procsd geom import}
\usage{
write_procsd_geom_imp(
  x,
  import_this_ls,
  path_to_starter_sf_chr,
  merge_with_chr_vec,
  crs_nbr_vec,
  overwrite_lgl
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{import_this_ls}{Import this (a list)}

\item{path_to_starter_sf_chr}{Path to starter simple features object (a character vector)}

\item{merge_with_chr_vec}{PARAM_DESCRIPTION}

\item{crs_nbr_vec}{PARAM_DESCRIPTION}

\item{overwrite_lgl}{Overwrite (a logical vector)}
}
\description{
write_procsd_geom_imp() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write procsd geom import. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
