% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_procsd_imp_xx}
\alias{write_procsd_imp_xx}
\title{Write procsd import}
\usage{
write_procsd_imp_xx(
  x,
  import_this_ls,
  path_to_starter_sf_chr,
  merge_with,
  pckg_name,
  raw_data_dir,
  processed_dir,
  crs_nbr_vec = NA_real_,
  overwrite_lgl = F
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{import_this_ls}{Import this (a list)}

\item{path_to_starter_sf_chr}{Path to starter simple features object (a character vector)}

\item{merge_with}{PARAM_DESCRIPTION}

\item{pckg_name}{PARAM_DESCRIPTION}

\item{raw_data_dir}{PARAM_DESCRIPTION}

\item{processed_dir}{PARAM_DESCRIPTION}

\item{crs_nbr_vec}{PARAM_DESCRIPTION, Default: NA}

\item{overwrite_lgl}{Overwrite (a logical vector), Default: F}
}
\description{
write_procsd_imp_xx() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write procsd import output object of multiple potential types. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
