% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_raw_data_from_sp_local_r4}
\alias{write_raw_data_from_sp_local_r4}
\title{Write raw data from sp local}
\usage{
write_raw_data_from_sp_local_r4(x, return_r4_lgl)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{return_r4_lgl}{Return ready4 S4 (a logical vector)}
}
\description{
write_raw_data_from_sp_local_r4() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write raw data from sp local ready4 s4. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
