% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_raw_format_dir}
\alias{write_raw_format_dir}
\title{Write raw format directory}
\usage{
write_raw_format_dir(data_type_chr, raw_data_dir)
}
\arguments{
\item{data_type_chr}{Data type (a character vector)}

\item{raw_data_dir}{PARAM_DESCRIPTION}
}
\description{
write_raw_format_dir() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write raw format directory. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
