% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attr_to_sf.R
\name{add_attr_list_to_sf}
\alias{add_attr_list_to_sf}
\title{add_attr_list_to_sf}
\usage{
add_attr_list_to_sf(x, y, lookup_tb_r4)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{lookup_tb_r4}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready4utils]{data_get}}
 \code{\link[ready4s4]{sp_data_pack_lup}}
}
