% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_attr_list_to_sf}
\alias{add_attr_list_to_sf}
\title{Add attr list to}
\usage{
add_attr_list_to_sf(x, y, lookup_tb_r4)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{lookup_tb_r4}{Lookup (a ready4 S4 collection of tibbles)}
}
\description{
add_attr_list_to_sf() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add attr list to simple features object. Function argument x specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
