% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add_attr_to_data_pack.R
\name{add_attr_tb_to_data_pack_lup}
\alias{add_attr_tb_to_data_pack_lup}
\title{add_attr_tb_to_data_pack_lup}
\usage{
add_attr_tb_to_data_pack_lup(
  data_pack_lup,
  attr_tb,
  object_name,
  area_type,
  area_bound_yr,
  region,
  year,
  year_start,
  year_end,
  main_feature
)
}
\arguments{
\item{data_pack_lup}{PARAM_DESCRIPTION}

\item{attr_tb}{PARAM_DESCRIPTION}

\item{object_name}{PARAM_DESCRIPTION}

\item{area_type}{PARAM_DESCRIPTION}

\item{area_bound_yr}{PARAM_DESCRIPTION}

\item{region}{PARAM_DESCRIPTION}

\item{year}{PARAM_DESCRIPTION}

\item{year_start}{PARAM_DESCRIPTION}

\item{year_end}{PARAM_DESCRIPTION}

\item{main_feature}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
 \code{\link[dplyr]{bind}}
}
