% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attr_to_data_pack.R
\name{add_attr_tb_to_data_pack_lup_from_arg_list}
\alias{add_attr_tb_to_data_pack_lup_from_arg_list}
\title{add_attr_tb_to_data_pack_lup_from_arg_list}
\usage{
add_attr_tb_to_data_pack_lup_from_arg_list(x, y)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
