% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_attr_to_sf}
\alias{add_attr_to_sf}
\title{Add attr to}
\usage{
add_attr_to_sf(area_sf, attr_data_tb, attr_data_desc)
}
\arguments{
\item{area_sf}{Area (a simple features object)}

\item{attr_data_tb}{Attr data (a tibble)}

\item{attr_data_desc}{PARAM_DESCRIPTION}
}
\value{
NA ()
}
\description{
add_attr_to_sf() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add attr to simple features object. Function argument area_sf specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
