% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add_attr_to_sf.R
\name{add_attr_to_sf}
\alias{add_attr_to_sf}
\title{add_attr_to_sf}
\usage{
add_attr_to_sf(area_sf, attr_data_tb, attr_data_desc)
}
\arguments{
\item{area_sf}{PARAM_DESCRIPTION}

\item{attr_data_tb}{PARAM_DESCRIPTION}

\item{attr_data_desc}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{join}}
 \code{\link[stringr]{str_detect}}
 \code{\link[sf]{st_as_sf}}
}
