% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attr_to_data_pack.R
\name{add_attribute_to_data_pack}
\alias{add_attribute_to_data_pack}
\title{add_attribute_to_data_pack}
\usage{
add_attribute_to_data_pack(combined_ste_ppr_ls, object_name_stub)
}
\arguments{
\item{combined_ste_ppr_ls}{PARAM_DESCRIPTION}

\item{object_name_stub}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
 \code{\link[stringr]{str_sub}}
}
