% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_attribute_to_data_pack_from_tb}
\alias{add_attribute_to_data_pack_from_tb}
\title{Add attribute to data pack from}
\usage{
add_attribute_to_data_pack_from_tb(attr_tb, object_name)
}
\arguments{
\item{attr_tb}{Attr (a tibble)}

\item{object_name}{PARAM_DESCRIPTION}
}
\description{
add_attribute_to_data_pack_from_tb() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add attribute to data pack from tibble. Function argument attr_tb specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
