% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make_sp_data_list.R
\name{add_dynamic_sp_vars_to_sf}
\alias{add_dynamic_sp_vars_to_sf}
\title{add_dynamic_sp_vars_to_sf}
\usage{
add_dynamic_sp_vars_to_sf(dynamic_sp_vars_sf, pop_attr_sf,
  age_sex_pop_resolution, age_sex_var_name, popl_var_prefix, data_year,
  crs_nbr_vec)
}
\arguments{
\item{dynamic_sp_vars_sf}{PARAM_DESCRIPTION}

\item{pop_attr_sf}{PARAM_DESCRIPTION}

\item{age_sex_pop_resolution}{PARAM_DESCRIPTION}

\item{age_sex_var_name}{PARAM_DESCRIPTION}

\item{popl_var_prefix}{PARAM_DESCRIPTION}

\item{data_year}{PARAM_DESCRIPTION}

\item{crs_nbr_vec}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
 \code{\link[rlang]{sym}}
}
