% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_intersect_sfs_keep_counts.R
\name{add_kmsq_area_by_group}
\alias{add_kmsq_area_by_group}
\title{add_kmsq_area_by_group}
\usage{
add_kmsq_area_by_group(
  sf,
  group_by_var,
  feature_nm,
  prefix = "whl_",
  suffix = "_area"
)
}
\arguments{
\item{sf}{PARAM_DESCRIPTION}

\item{group_by_var}{PARAM_DESCRIPTION}

\item{feature_nm}{PARAM_DESCRIPTION}

\item{prefix}{PARAM_DESCRIPTION, Default: 'whl_'}

\item{suffix}{PARAM_DESCRIPTION, Default: '_area'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}}
 \code{\link[rlang]{sym}}
 \code{\link[sf]{geos_combine}},\code{\link[sf]{st_geometry}}
}
