% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_names}
\alias{add_names}
\title{Add names}
\usage{
add_names(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\description{
add_names() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add names. Function argument x specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
