% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add_attr_to_data_pack.R
\name{add_ppr_ls_to_data_pack_lup}
\alias{add_ppr_ls_to_data_pack_lup}
\title{add_ppr_ls_to_data_pack_lup}
\usage{
add_ppr_ls_to_data_pack_lup(x, y)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
