% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_ppr_to_data_pack_lup}
\alias{add_ppr_to_data_pack_lup}
\title{Add ppr to data pack}
\usage{
add_ppr_to_data_pack_lup(
  data_pack_lup,
  combined_ste_ppr_ls,
  object_name_stub,
  area_type,
  area_bound_yr,
  region
)
}
\arguments{
\item{data_pack_lup}{Data pack (a lookup table)}

\item{combined_ste_ppr_ls}{Combined ste ppr (a list)}

\item{object_name_stub}{PARAM_DESCRIPTION}

\item{area_type}{PARAM_DESCRIPTION}

\item{area_bound_yr}{PARAM_DESCRIPTION}

\item{region}{PARAM_DESCRIPTION}
}
\description{
add_ppr_to_data_pack_lup() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add ppr to data pack lookup table. Function argument data_pack_lup specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
