% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_sp_resolution}
\alias{add_sp_resolution}
\title{Add sp resolution}
\usage{
add_sp_resolution(lookup_tbs_r4, processed_dir)
}
\arguments{
\item{lookup_tbs_r4}{Lookup tibbles (a ready4 S4)}

\item{processed_dir}{PARAM_DESCRIPTION}
}
\description{
add_sp_resolution() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add sp resolution. Function argument lookup_tbs_r4 specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
