% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_profiled_area.R,
%   R/gnrc_area_bound_year.R
\name{area_bound_year-ready4_profiled_area}
\alias{area_bound_year-ready4_profiled_area}
\alias{area_bound_year}
\title{area_bound_year}
\arguments{
\item{x}{An object of class ready4_profiled_area}

\item{x}{An object}
}
\description{
Get the value of the slot area_bound_year for S4 objects of class ready4_profiled_area

S4 Generic function to get the value of the slot area_bound_year
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
