% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prevalence.R
\name{calc_change_in_prevalence}
\alias{calc_change_in_prevalence}
\title{calc_change_in_prevalence}
\usage{
calc_change_in_prevalence(sp_data_sf, col_ind, t0_prev, tx_prev)
}
\arguments{
\item{sp_data_sf}{PARAM_DESCRIPTION}

\item{col_ind}{PARAM_DESCRIPTION}

\item{t0_prev}{PARAM_DESCRIPTION}

\item{tx_prev}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
 \code{\link[rlang]{sym}}
 \code{\link[stringr]{str_sub}}
}
