% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sp_data.R
\name{check_if_ppr}
\alias{check_if_ppr}
\title{check_if_ppr}
\usage{
check_if_ppr(data_name_item, data_lookup_tb, pop_projs_str)
}
\arguments{
\item{data_name_item}{PARAM_DESCRIPTION}

\item{data_lookup_tb}{PARAM_DESCRIPTION}

\item{pop_projs_str}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
 \code{\link[ready4utils]{data_get}}
 \code{\link[stringr]{str_detect}}
 \code{\link[magrittr]{extract}}
}
