% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aus_land_boundary.R
\name{create_australia_land_boundary}
\alias{create_australia_land_boundary}
\title{create_australia_land_boundary}
\usage{
create_australia_land_boundary(state_territories = NULL,
  aus_boundary_sf = ready.utils::data_get(data_lookup_tb =
  aus_spatial_lookup_tb, lookup_reference = "aus_boundary_phns_sf",
  lookup_variable = "name", target_variable = "source_reference"))
}
\arguments{
\item{state_territories}{PARAM_DESCRIPTION}

\item{aus_boundary_sf}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Create a boundary SF of Australia
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}}
 \code{\link[sf]{geos_combine}}
}
