% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{data_import_get_dir_paths}
\alias{data_import_get_dir_paths}
\title{data_import_get_dir_paths}
\usage{
data_import_get_dir_paths(x, destination_directory, data_lookup_ref,
  lookup_variable, directory_sub_divs)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{destination_directory}{PARAM_DESCRIPTION}

\item{data_lookup_ref}{PARAM_DESCRIPTION}

\item{lookup_variable}{PARAM_DESCRIPTION}

\item{directory_sub_divs}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}},\code{\link[purrr]{accumulate}}
}
