% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_import_data.R
\name{get_sngl_path_for_imp}
\alias{get_sngl_path_for_imp}
\title{get_sngl_path_for_imp}
\usage{
get_sngl_path_for_imp(downloaded_data_tb, lookup_reference, data_directory)
}
\arguments{
\item{downloaded_data_tb}{PARAM_DESCRIPTION}

\item{lookup_reference}{PARAM_DESCRIPTION}

\item{data_directory}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
 \code{\link[dplyr]{select}}
 \code{\link[ready4fun]{get_from_lup}}
}
