% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_import_data.R
\name{data_import_make_directories}
\alias{data_import_make_directories}
\title{data_import_make_directories}
\usage{
data_import_make_directories(directory_paths)
}
\arguments{
\item{directory_paths}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
}
