% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{data_import_non_shape_items}
\alias{data_import_non_shape_items}
\title{data_import_non_shape_items}
\usage{
data_import_non_shape_items(path_str, x)
}
\arguments{
\item{path_str}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_sub}},\code{\link[stringr]{str_c}}
 \code{\link[stringi]{stri_locate_all}}
 \code{\link[purrr]{map}}
 \code{\link[tibble]{as_tibble}},\code{\link[tibble]{deprecated}}
 \code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{filter}}
 \code{\link[readxl]{read_excel}}
 \code{\link[stats]{setNames}}
}
