% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_import_data.R
\name{get_menu_of_type_nms_for_imp}
\alias{get_menu_of_type_nms_for_imp}
\title{get_menu_of_type_nms_for_imp}
\usage{
get_menu_of_type_nms_for_imp(x, lookup_ref)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{lookup_ref}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}},\code{\link[dplyr]{pull}}
}
