% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_profiled_area.R, R/gnrc_data_year.R
\name{data_year-ready4_profiled_area}
\alias{data_year-ready4_profiled_area}
\alias{data_year}
\title{data_year}
\arguments{
\item{x}{An object of class ready4_profiled_area}

\item{x}{An object}
}
\description{
Get the value of the slot data_year for S4 objects of class ready4_profiled_area

S4 Generic function to get the value of the slot data_year
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
