% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_by_yearly_age_sex.R
\name{gen_demog_features}
\alias{gen_demog_features}
\alias{by_sex_for_an_age}
\alias{by_age_sex_for_a_year}
\alias{demographic_compound_growth_rate}
\alias{col_pair_growth_rate}
\alias{future_pop_from_comp_g_r}
\alias{gen_age_sex_estimates_tx}
\title{Calculate age/sex populations for each year / age of interest.}
\usage{
gen_demog_features(profiled_sf, years, age_lower, age_upper, param_tb,
  it_nbr, acgr = TRUE, age_by_year = FALSE, drop_bands = TRUE,
  intervals = 5, month = "07", day = "01", included_prefix = "")

by_sex_for_an_age(profiled_sf, age, age_lower, age_upper, t0_date,
  age_bands_ref, intervals, included_age_bands_num_all, included_prefix)

by_age_sex_for_a_year(profiled_sf, age_lower, age_upper, t0_date,
  age_bands_ref, intervals, included_age_bands_num_all, included_prefix)

demographic_compound_growth_rate(t0_pop, t1_pop, n_periods, age_sex_band,
  param_tb, it_nbr)

col_pair_growth_rate(profiled_sf, list_element, included_cols_pairs,
  included_time_intervals_start_end, param_tb, it_nbr)

future_pop_from_comp_g_r(cgr, base_year_pop, nyears)

gen_age_sex_estimates_tx(profiled_sf, ymwd_step)
}
\arguments{
\item{profiled_sf}{PARAM_DESCRIPTION}

\item{years}{Numeric vector of the years for which age / sex counts should be estimated.}

\item{age_lower}{Numeric, start of included age range.}

\item{age_upper}{Numeric, end of included age range.}

\item{param_tb}{PARAM_DESCRIPTION}

\item{it_nbr}{PARAM_DESCRIPTION}

\item{intervals}{Numeric, the years included in each age group in the source sf, Default: 5}

\item{age}{PARAM_DESCRIPTION}

\item{t0_date}{PARAM_DESCRIPTION}

\item{age_bands_ref}{PARAM_DESCRIPTION}

\item{included_age_bands_num_all}{PARAM_DESCRIPTION}

\item{t0_pop}{A numeric vector of population counts in the baseline year.}

\item{t1_pop}{A numeric vector of population counts in the follow-up year.}

\item{n_periods}{Number of years between baseline and follow-up.}

\item{age_sex_band}{PARAM_DESCRIPTION}

\item{list_element}{Numeric, index value for look-up.}

\item{included_cols_pairs}{A list ....}

\item{included_time_intervals_start_end}{A list ....}

\item{cgr}{Numeric, cumulative annual growth rate.}

\item{base_year_pop}{Numeric, population count in base year.}

\item{nyears}{Number of years between baseline and follow-up.}

\item{ymwd_step}{PARAM_DESCRIPTION}

\item{included_features}{XXXXXXXXXXXXX}

\item{profiled_sf}{PARAM_DESCRIPTION}

\item{age_lower}{PARAM_DESCRIPTION}

\item{age_upper}{PARAM_DESCRIPTION}

\item{intervals}{PARAM_DESCRIPTION}

\item{profiled_sf}{PARAM_DESCRIPTION}

\item{age_lower}{PARAM_DESCRIPTION}

\item{age_upper}{PARAM_DESCRIPTION}

\item{t0_date}{PARAM_DESCRIPTION}

\item{age_bands_ref}{PARAM_DESCRIPTION}

\item{intervals}{PARAM_DESCRIPTION}

\item{included_age_bands_num_all}{PARAM_DESCRIPTION}

\item{profiled_sf}{PARAM_DESCRIPTION}

\item{param_tb}{PARAM_DESCRIPTION}

\item{it_nbr}{PARAM_DESCRIPTION}

\item{profiled_sf}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS

DETAILS

DETAILS

DETAILS

DETAILS

DETAILS
}
\section{Functions}{
\itemize{
\item \code{by_sex_for_an_age}: Calculates age populations for each sex.

\item \code{by_age_sex_for_a_year}: Calculates age and sex populations for a specific year.

\item \code{demographic_compound_growth_rate}: Calculates the compound annual growth rate between two periods.

\item \code{col_pair_growth_rate}: Adds columns with the compound annual growth rate between two periods to inputted sf.

\item \code{future_pop_from_comp_g_r}: Calculates a projected future population using the annual compound growth rate.

\item \code{gen_age_sex_estimates_tx}: Calculates grouped age / sex counts for additional years.
}}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}},\code{\link[purrr]{flatten}},\code{\link[purrr]{modify}},\code{\link[purrr]{reduce}},\code{\link[purrr]{prepend}}
 \code{\link[stringr]{str_subset}},\code{\link[stringr]{str_sub}}

\code{\link[stringr]{str_sub}},\code{\link[stringr]{case}}
 \code{\link[purrr]{reduce}},\code{\link[purrr]{prepend}},\code{\link[purrr]{pluck}}
 \code{\link[dplyr]{mutate}}
 \code{\link[rlang]{sym}}

\code{\link[purrr]{reduce}},\code{\link[purrr]{prepend}}

\code{\link[purrr]{reduce}},\code{\link[purrr]{prepend}},\code{\link[purrr]{pluck}}
 \code{\link[dplyr]{mutate}}
 \code{\link[rlang]{sym}}
 \code{\link[stringr]{str_sub}},\code{\link[stringr]{str_replace}},\code{\link[stringr]{case}}

\code{\link[dplyr]{summarise_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{funs}},\code{\link[dplyr]{select_all}},\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{pull}}
 \code{\link[stringr]{str_sub}},\code{\link[stringr]{str_replace}}
 \code{\link[purrr]{pluck}},\code{\link[purrr]{map}},\code{\link[purrr]{reduce}},\code{\link[purrr]{prepend}}
 \code{\link[lubridate]{ymd}},\code{\link[lubridate]{period}},\code{\link[lubridate]{interval}}
}
