% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_by_yearly_age_sex.R
\name{gen_demog_features}
\alias{gen_demog_features}
\alias{by_sex_for_an_age}
\alias{by_age_sex_for_a_year}
\alias{demographic_compound_growth_rate}
\alias{col_pair_growth_rate}
\alias{future_pop_from_comp_g_r}
\alias{add_extra_year_pop_totals}
\title{Calculate age/sex populations for each year / age of interest.}
\usage{
gen_demog_features(profiled_sf, years, age0, age1, param_tb, it_nbr,
  acgr = TRUE, age_by_year = FALSE, drop_bands = TRUE,
  intervals = 5, month = "07", day = "01")

by_sex_for_an_age(profiled_sf, age, age0, age1, t0_date, age_bands_ref,
  intervals, included_age_bands_num_all)

by_age_sex_for_a_year(profiled_sf, age0, age1, t0_date, age_bands_ref,
  intervals, included_age_bands_num_all)

demographic_compound_growth_rate(t0_pop, t1_pop, n_periods, age_sex_band,
  param_tb, it_nbr)

col_pair_growth_rate(profiled_sf, list_element, included_cols_pairs,
  included_time_intervals_start_end, param_tb, it_nbr)

future_pop_from_comp_g_r(cgr, base_year_pop, nyears)

add_extra_year_pop_totals(profiled_sf, year, included_time_intervals,
  included_cols_pairs)
}
\arguments{
\item{profiled_sf}{PARAM_DESCRIPTION}

\item{years}{Numeric vector of the years for which age / sex counts should be estimated.}

\item{age0}{Numeric, start of included age range.}

\item{age1}{Numeric, end of included age range.}

\item{intervals}{Numeric, the years included in each age group in the source sf, Default: 5}

\item{age}{Numeric, ...}

\item{included_age_bands_num_all}{A list ....}

\item{t0_pop}{A numeric vector of population counts in the baseline year.}

\item{t1_pop}{A numeric vector of population counts in the follow-up year.}

\item{n_periods}{Number of years between baseline and follow-up.}

\item{list_element}{Numeric, index value for look-up.}

\item{included_cols_pairs}{A list ....}

\item{included_time_intervals_start_end}{A list ....}

\item{cgr}{Numeric, cumulative annual growth rate.}

\item{base_year_pop}{Numeric, population count in base year.}

\item{nyears}{Number of years between baseline and follow-up.}

\item{year}{Numeric, ....}

\item{included_time_intervals}{A list, ....}

\item{year}{Numeric, ....}

\item{included_age_bands_num_all}{A list ....}

\item{year}{Numeric, ....}

\item{included_cols_pairs}{A list, ....}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\section{Functions}{
\itemize{
\item \code{by_sex_for_an_age}: Calculates age populations for each sex.

\item \code{by_age_sex_for_a_year}: Calculates age and sex populations for a specific year.

\item \code{demographic_compound_growth_rate}: Calculates the compound annual growth rate between two periods.

\item \code{col_pair_growth_rate}: Adds columns with the compound annual growth rate between two periods to inputted sf.

\item \code{future_pop_from_comp_g_r}: Calculates a projected future population using the annual compound growth rate.

\item \code{add_extra_year_pop_totals}: Calculates grouped age / sex counts for additional years.
}}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 demographic_by_yearly_age_sex(profiled_sf = spatial_vic_pop_growth_lga(vic_pop_growth_by_age_lga_t0 = ready.data::data_get(data_lookup_tb = aus_spatial_lookup_tb, lookup_reference = "vic_pop_growth_by_age_lga_2016_tb"),
 vic_pop_growth_by_age_lga_t1 = ready.data::data_get(data_lookup_tb = aus_spatial_lookup_tb, lookup_reference = "vic_pop_growth_by_age_lga_2031_tb"),
 t0 ="2016",
 t1 ="2031"),
 years = c(2016,2019, 2031, 2025),
 age0 = 4,
 age1 = 18)
 }
}
}
\seealso{
\code{\link[purrr]{map}},\code{\link[purrr]{flatten}},\code{\link[purrr]{modify}},\code{\link[purrr]{reduce}},\code{\link[purrr]{prepend}}
 \code{\link[stringr]{str_subset}},\code{\link[stringr]{str_sub}}
}
