% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_growth_projs_slice.R
\name{demographic_growth_projs_slice}
\alias{demographic_growth_projs_slice}
\title{demographic_growth_projs_slice}
\usage{
demographic_growth_projs_slice(demographic.projection.errors.partial,
  demographic.data.partial, namestub)
}
\arguments{
\item{demographic.projection.errors.partial}{PARAM_DESCRIPTION}

\item{demographic.data.partial}{PARAM_DESCRIPTION}

\item{namestub}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Creates a tibble......
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map2}}
 \code{\link[stats]{setNames}}
 \code{\link[dplyr]{bind}}
}
