% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_projections.R
\name{demographic_projections}
\alias{demographic_projections}
\title{demographic_projections}
\usage{
demographic_projections(naive, n.replications)
}
\arguments{
\item{naive}{A logical value, denoting whether using a naive modelling approach (assumes
previous growth rates continue into future) or a theory based model.}

\item{n.replications}{A positive integer.}
}
\value{
A list.
}
\description{
Forecast population groth with prediction intervals.
}
\details{
Function implements calculations based on:
http://www.jstor.org.ezp.lib.unimelb.edu.au/stable/41110880?seq=1#page_scan_tab_contents
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
 \code{\link[dplyr]{summarise}},\code{\link[dplyr]{slice}},\code{\link[dplyr]{bind}},\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{pull}}
 \code{\link[purrr]{map2}}
 \code{\link[stats]{setNames}}
}
