% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_range_of_projections.R
\name{demographic_range_of_projections}
\alias{demographic_range_of_projections}
\title{demographic_range_of_projections}
\usage{
demographic_range_of_projections(mean.abs.error,
  three.std.devs.from.mean.error, n.five.year.predictions, n.replications)
}
\arguments{
\item{mean.abs.error}{A .....}

\item{three.std.devs.from.mean.error}{A .....}

\item{n.five.year.predictions}{A positive integer.}

\item{n.replications}{A positive integer.}
}
\value{
A tibble.
}
\description{
Estimate a vector of demographic projections.
}
\details{
Implements formula from http://www.healthstrategy.com/cgi-bin/datagen.pl
Converts absolute mean error into +/- range.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[invgamma]{invgamma}}
}
