% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_data_pack.R
\name{export_data_pack_lup}
\alias{export_data_pack_lup}
\title{export_data_pack_lup}
\usage{
export_data_pack_lup(lookup_tbs_r4, tb_data_type = "Geometry",
  template_ls = NULL, pckg_name)
}
\arguments{
\item{lookup_tbs_r4}{PARAM_DESCRIPTION}

\item{tb_data_type}{PARAM_DESCRIPTION, Default: 'Geometry'}

\item{template_ls}{PARAM_DESCRIPTION, Default: NULL}

\item{pckg_name}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map2}},\code{\link[purrr]{reduce}}
 \code{\link[ready4utils]{data_get}}
 \code{\link[dplyr]{mutate}}
}
