% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_pack.R
\name{export_starter_sf}
\alias{export_starter_sf}
\title{export_starter_sf}
\usage{
export_starter_sf(lookup_tbs_r4, boundary_ls, processed_dir, merge_with)
}
\arguments{
\item{lookup_tbs_r4}{PARAM_DESCRIPTION}

\item{boundary_ls}{PARAM_DESCRIPTION}

\item{processed_dir}{PARAM_DESCRIPTION}

\item{merge_with}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{geos_binary_ops}},\code{\link[sf]{geos_measures}}
 \code{\link[dplyr]{mutate}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{pull}}
 \code{\link[units]{set_units}}
 \code{\link[tibble]{add_row}}
 \code{\link[ready4s4]{sp_starter_sf_lup}},\code{\link[ready4s4]{sp_import_lup}},\code{\link[ready4s4]{sp_starter_sf_lup<-}}
}
