% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_data_pack.R
\name{export_uid_lup}
\alias{export_uid_lup}
\title{export_uid_lup}
\usage{
export_uid_lup(lookup_tbs_r4)
}
\arguments{
\item{lookup_tbs_r4}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{add_row}}
 \code{\link[dplyr]{pull}}
}
