% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_area_within_xkm_of_points.R
\name{gen_distance_based_bands}
\alias{gen_distance_based_bands}
\title{gen_distance_based_bands
FUNCTION_DESCRIPTION}
\usage{
gen_distance_based_bands(distance_km_outer, nbr_distance_bands,
  service_cluster_tb, aus_stt_sf)
}
\arguments{
\item{distance_km_outer}{PARAM_DESCRIPTION}

\item{nbr_distance_bands}{PARAM_DESCRIPTION}

\item{service_cluster_tb}{PARAM_DESCRIPTION}

\item{aus_stt_sf}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
gen_distance_based_bands
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{pull}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{arrange}}
 \code{\link[purrr]{map}}
 \code{\link[stats]{setNames}}
}
