% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update_counts_by_area.R
\name{make_nse_objs_ls}
\alias{make_nse_objs_ls}
\title{make_nse_objs_ls}
\usage{
make_nse_objs_ls(
  sp_unit,
  concept,
  tot_pop_col = NULL,
  grouping_1 = NULL,
  data_year,
  popl_var_prefix
)
}
\arguments{
\item{sp_unit}{PARAM_DESCRIPTION}

\item{concept}{PARAM_DESCRIPTION}

\item{tot_pop_col}{PARAM_DESCRIPTION, Default: NULL}

\item{grouping_1}{PARAM_DESCRIPTION, Default: NULL}

\item{data_year}{PARAM_DESCRIPTION}

\item{popl_var_prefix}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
make_nse_objs_ls
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
