% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_gen_val_envir.R
\name{gen_val_envir}
\alias{gen_val_envir}
\title{gen_val_envir}
\usage{
gen_val_envir(params_struc_tb, it_nbr)
}
\arguments{
\item{params_struc_tb}{PARAM_DESCRIPTION}

\item{it_nbr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Generate a set of parameter values by iteration for the mean absolute prediction error of demographic projections.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
