% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_gen_val_envir.R
\name{gen_val_envir_sgl}
\alias{gen_val_envir_sgl}
\title{gen_val_envir_sgl}
\usage{
gen_val_envir_sgl(distribution, dist_param_1, dist_param_2, dist_param_3,
  transformation)
}
\arguments{
\item{distribution}{PARAM_DESCRIPTION}

\item{dist_param_1}{PARAM_DESCRIPTION}

\item{dist_param_2}{PARAM_DESCRIPTION}

\item{dist_param_3}{PARAM_DESCRIPTION}

\item{transformation}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
