% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_gen.R
\name{gen_val_envir_vect}
\alias{gen_val_envir_vect}
\title{Gen value envir vect}
\usage{
gen_val_envir_vect(
  distribution,
  dist_param_1,
  dist_param_2,
  dist_param_3,
  transformation
)
}
\arguments{
\item{distribution}{PARAM_DESCRIPTION}

\item{dist_param_1}{PARAM_DESCRIPTION}

\item{dist_param_2}{PARAM_DESCRIPTION}

\item{dist_param_3}{PARAM_DESCRIPTION}

\item{transformation}{PARAM_DESCRIPTION}
}
\description{
gen_val_envir_vect() is a Gen function that generates values for a specified varaible or variables. Specifically, this function implements an algorithm to gen value envir vect. The function is called for its side effects and does not return a value.
}
\keyword{internal}
