% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_gen_par_vals.R
\name{gen_val_mape}
\alias{gen_val_mape}
\title{gen_val_mape}
\usage{
gen_val_mape(params_struc_tb, it_nbr, jt_dist)
}
\arguments{
\item{params_struc_tb}{PARAM_DESCRIPTION}

\item{it_nbr}{PARAM_DESCRIPTION}

\item{jt_dist}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Generate a set of parameter values by iteration for the mean absolute prediction error of demographic projections.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}}
 \code{\link[mc2d]{pert}}
 \code{\link[tidyr]{gather}},\code{\link[tidyr]{unite}}
}
