% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_profiled_area.R,
%   R/gnrc_geom_dist_limit_km.R
\name{geom_dist_limit_km<--ready4_profiled_area}
\alias{geom_dist_limit_km<--ready4_profiled_area}
\alias{geom_dist_limit_km<-}
\title{geom_dist_limit_km<-}
\arguments{
\item{x}{An object of class ready4_profiled_area}

\item{x}{An object}
}
\description{
Set the value of the slot geom_dist_limit_km for S4 objects of class ready4_profiled_area

S4 Generic function to set the value of the slot geom_dist_limit_km
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
