% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_geom_dist_limit_km.R
\name{geom_dist_limit_km}
\alias{geom_dist_limit_km}
\alias{geom_dist_limit_km-ready4_profiled_area}
\title{geom_dist_limit_km}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_profiled_area}
}
\description{
S4 Generic function to get the value of the slot geom_dist_limit_km

Get the value of the slot geom_dist_limit_km for S4 objects of class ready4_profiled_area
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
