% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add_sp_resolution.R
\name{get_area_sqkm_sf}
\alias{get_area_sqkm_sf}
\title{get_area_sqkm_sf}
\usage{
get_area_sqkm_sf(data_sf)
}
\arguments{
\item{data_sf}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{summarise}},\code{\link[dplyr]{pull}}
 \code{\link[sf]{geos_measures}}
 \code{\link[units]{set_units}}
}
