% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sp_data.R
\name{get_closest_year}
\alias{get_closest_year}
\title{get_closest_year}
\usage{
get_closest_year(data_lookup_tb, incl_main_ft_vec, target_year,
  target_area = NULL, find_closest = "abs")
}
\arguments{
\item{data_lookup_tb}{PARAM_DESCRIPTION}

\item{incl_main_ft_vec}{PARAM_DESCRIPTION}

\item{target_year}{PARAM_DESCRIPTION}

\item{target_area}{PARAM_DESCRIPTION, Default: NULL}

\item{find_closest}{PARAM_DESCRIPTION, Default: 'abs'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{pull}}
 \code{\link[purrr]{map}}
}
