% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_get_data.R
\name{get_data}
\alias{get_data}
\alias{get_data,ready4_sp_data_pack_lup-method}
\title{Get data}
\usage{
\S4method{get_data}{ready4_sp_data_pack_lup}(x, col_chr = "main_feature", value_chr)
}
\arguments{
\item{x}{Primary input object - see Usage section for allowable signatures}

\item{col_chr}{PARAM_DESCRIPTION, Default: 'main_feature'}

\item{value_chr}{PARAM_DESCRIPTION}
}
\description{
Get data referenced in a lookup table.
}
