% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_data_year_chr}
\alias{get_data_year_chr}
\title{Get data year}
\usage{
get_data_year_chr(data_ymdhms)
}
\arguments{
\item{data_ymdhms}{PARAM_DESCRIPTION}
}
\description{
get_data_year_chr() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get data year character vector. Function argument data_ymdhms specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
