% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sim_data_env.R
\name{get_group_by_var}
\alias{get_group_by_var}
\title{get_group_by_var}
\usage{
get_group_by_var(profile_unit, data_unit, group_at_profile_unit = TRUE,
  group_by_lookup_tb)
}
\arguments{
\item{profile_unit}{PARAM_DESCRIPTION}

\item{data_unit}{PARAM_DESCRIPTION}

\item{group_at_profile_unit}{PARAM_DESCRIPTION, Default: TRUE}

\item{group_by_lookup_tb}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready.utils]{data_get}}
}
