% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_group_by_var_from_pai}
\alias{get_group_by_var_from_pai}
\title{Get group by var from pai}
\usage{
get_group_by_var_from_pai(pa_r4)
}
\arguments{
\item{pa_r4}{Pa (a ready4 S4)}
}
\value{
NA ()
}
\description{
get_group_by_var_from_pai() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get group by var from pai. Function argument pa_r4 specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
