% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sp_data_list.R
\name{get_group_by_var_from_pai}
\alias{get_group_by_var_from_pai}
\title{get_group_by_var_from_pai}
\usage{
get_group_by_var_from_pai(pa_r4)
}
\arguments{
\item{pa_r4}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready4s4]{sp_uid_lup}},\code{\link[ready4s4]{lookup_tb}},\code{\link[ready4s4]{data_year}},\code{\link[ready4s4]{use_coord_lup}},\code{\link[ready4s4]{area_type}},\code{\link[ready4s4]{geom_dist_limit_km}}
 \code{\link[dplyr]{filter}}
}
