% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get_sp_data.R
\name{get_highest_res}
\alias{get_highest_res}
\title{get_highest_res}
\usage{
get_highest_res(options_vec, year, resolution_lup_r3)
}
\arguments{
\item{options_vec}{PARAM_DESCRIPTION}

\item{year}{PARAM_DESCRIPTION}

\item{resolution_lup_r3}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
