% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_data_pack.R
\name{get_import_chr_vec}
\alias{get_import_chr_vec}
\title{get_import_chr_vec}
\usage{
get_import_chr_vec(lookup_tbs_r4, data_type_chr)
}
\arguments{
\item{lookup_tbs_r4}{PARAM_DESCRIPTION}

\item{data_type_chr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{pull}}
}
