% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_included_yrs}
\alias{get_included_yrs}
\title{Get included years}
\usage{
get_included_yrs(geometry_sf, pfx_1L_chr = "y2")
}
\arguments{
\item{geometry_sf}{Geometry (a simple features object)}

\item{pfx_1L_chr}{Prefix (a character vector of length one), Default: 'y2'}
}
\value{
Years (a double vector)
}
\description{
get_included_yrs() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get included years. Function argument geometry_sf specifies the where to look for the required object. The function returns Years (a double vector).
}
\keyword{internal}
