% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_max_or_min_yr_of_sf}
\alias{get_max_or_min_yr_of_sf}
\title{Get maximum or minimum year of simple features object}
\usage{
get_max_or_min_yr_of_sf(geometry_sf, max_1L_lgl = T, pfx_1L_chr = "y2")
}
\arguments{
\item{geometry_sf}{Geometry (a simple features object)}

\item{max_1L_lgl}{Maximum (a logical vector of length one), Default: T}

\item{pfx_1L_chr}{Prefix (a character vector of length one), Default: 'y2'}
}
\value{
Year (a double vector of length one)
}
\description{
get_max_or_min_yr_of_sf() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get maximum or minimum year of simple features object. The function returns Year (a double vector of length one).
}
\keyword{internal}
