% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_pack.R
\name{get_merge_sf_str}
\alias{get_merge_sf_str}
\title{get_merge_sf_str}
\usage{
get_merge_sf_str(lookup_r4, sp_import_r3_slice, processed_dir = NULL)
}
\arguments{
\item{lookup_r4}{PARAM_DESCRIPTION}

\item{sp_import_r3_slice}{PARAM_DESCRIPTION}

\item{processed_dir}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{pull}}
 \code{\link[purrr]{pluck}},\code{\link[purrr]{map}}
 \code{\link[ready4utils]{data_get}}
 \code{\link[ready4s4]{sp_import_lup}},\code{\link[ready4s4]{sp_data_pack_lup}}
 \code{\link[stringr]{str_detect}}
}
