% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sim_data_env.R
\name{get_model_end_ymdhs}
\alias{get_model_end_ymdhs}
\title{get_model_end_ymdhs}
\usage{
get_model_end_ymdhs(input_data)
}
\arguments{
\item{input_data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[lubridate]{period}}
}
