% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_name_from_path_chr}
\alias{get_name_from_path_chr}
\title{Get name from path}
\usage{
get_name_from_path_chr(path_str, with_ext = TRUE)
}
\arguments{
\item{path_str}{PARAM_DESCRIPTION}

\item{with_ext}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
get_name_from_path_chr() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get name from path character vector. Function argument path_str specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
