% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_import_data.R
\name{get_name_from_path_chr}
\alias{get_name_from_path_chr}
\title{get_name_from_path_chr}
\usage{
get_name_from_path_chr(path_str, with_ext = TRUE)
}
\arguments{
\item{path_str}{PARAM_DESCRIPTION}

\item{with_ext}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_sub}}
 \code{\link[stringi]{stri_locate_all}}
}
