% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_popl_var_prefix}
\alias{get_popl_var_prefix}
\title{Get popl var prefix}
\usage{
get_popl_var_prefix(
  age_sex_pop_resolution,
  tot_pop_resolution = NULL,
  data_year
)
}
\arguments{
\item{age_sex_pop_resolution}{PARAM_DESCRIPTION}

\item{tot_pop_resolution}{PARAM_DESCRIPTION, Default: NULL}

\item{data_year}{PARAM_DESCRIPTION}
}
\description{
get_popl_var_prefix() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get popl var prefix. Function argument age_sex_pop_resolution specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
