% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_pack.R
\name{get_r_import_path_chr}
\alias{get_r_import_path_chr}
\title{get_r_import_path_chr}
\usage{
get_r_import_path_chr(r_data_dir_chr, name_chr, data_type_chr)
}
\arguments{
\item{r_data_dir_chr}{PARAM_DESCRIPTION}

\item{name_chr}{PARAM_DESCRIPTION}

\item{data_type_chr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
