% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_intersect_sfs_keep_counts.R
\name{get_res_specific_vars}
\alias{get_res_specific_vars}
\title{get_res_specific_vars}
\usage{
get_res_specific_vars(var_names, data_type, data_year, popl_var_prefix)
}
\arguments{
\item{var_names}{PARAM_DESCRIPTION}

\item{data_type}{PARAM_DESCRIPTION}

\item{data_year}{PARAM_DESCRIPTION}

\item{popl_var_prefix}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
