% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_resolution_hierarchy}
\alias{get_resolution_hierarchy}
\title{Get resolution hierarchy}
\usage{
get_resolution_hierarchy(data_year, resolution_lup_r3, whole_area = TRUE)
}
\arguments{
\item{data_year}{PARAM_DESCRIPTION}

\item{resolution_lup_r3}{Resolution (a ready4 S3 extension of lookup table)}

\item{whole_area}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
get_resolution_hierarchy() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get resolution hierarchy. Function argument data_year specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
