% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get_sp_data.R
\name{get_spatial_data_list}
\alias{get_spatial_data_list}
\title{get_spatial_data_list}
\usage{
get_spatial_data_list(
  input_ls,
  sub_div_unit = NULL,
  require_year_match = TRUE,
  excl_diff_bound_yr = TRUE
)
}
\arguments{
\item{input_ls}{PARAM_DESCRIPTION}

\item{sub_div_unit}{PARAM_DESCRIPTION, Default: NULL}

\item{require_year_match}{PARAM_DESCRIPTION, Default: TRUE}

\item{excl_diff_bound_yr}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Create a list of SF objects constructed from specified data types at highest or requested level of resolution.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_sub}}
 \code{\link[purrr]{map}},\code{\link[purrr]{map2}},\code{\link[purrr]{prepend}}
 \code{\link[stats]{setNames}}
 \code{\link[ready4utils]{data_get}}
}
