% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sp_data.R
\name{get_spatial_data_list}
\alias{get_spatial_data_list}
\title{get_spatial_data_list}
\usage{
get_spatial_data_list(at_highest_res, data_year, model_end_year = NULL,
  at_specified_res = NULL, country = "Australia", state = NULL,
  require_year_match = TRUE, excl_diff_bound_yr = TRUE, pop_projs_str)
}
\arguments{
\item{at_highest_res}{PARAM_DESCRIPTION}

\item{data_year}{PARAM_DESCRIPTION}

\item{model_end_year}{PARAM_DESCRIPTION, Default: NULL}

\item{at_specified_res}{PARAM_DESCRIPTION, Default: NULL}

\item{country}{PARAM_DESCRIPTION, Default: 'Australia'}

\item{state}{PARAM_DESCRIPTION, Default: NULL}

\item{require_year_match}{PARAM_DESCRIPTION, Default: TRUE}

\item{excl_diff_bound_yr}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Create a list of SF objects constructed from specified data types at highest or requested level of resolution.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_sub}}
 \code{\link[purrr]{map}},\code{\link[purrr]{map2}},\code{\link[purrr]{prepend}}
 \code{\link[stats]{setNames}}
}
