% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get_sp_data.R
\name{get_spatial_data_names}
\alias{get_spatial_data_names}
\title{get_spatial_data_names}
\usage{
get_spatial_data_names(input_ls, sub_div_unit = NULL,
  require_year_match = TRUE, excl_diff_bound_yr = TRUE)
}
\arguments{
\item{input_ls}{PARAM_DESCRIPTION}

\item{sub_div_unit}{PARAM_DESCRIPTION, Default: NULL}

\item{require_year_match}{PARAM_DESCRIPTION, Default: TRUE}

\item{excl_diff_bound_yr}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[lubridate]{year}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{pull}}
 \code{\link[stringr]{str_length}}
 \code{\link[purrr]{pluck}},\code{\link[purrr]{map}},\code{\link[purrr]{map2}},\code{\link[purrr]{flatten}},\code{\link[purrr]{reduce}}
 \code{\link[ready4utils]{data_get}}
}
