% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sim_data_env.R
\name{get_starter_sf_for_profiled_area}
\alias{get_starter_sf_for_profiled_area}
\title{get_starter_sf_for_profiled_area}
\usage{
get_starter_sf_for_profiled_area(profiled_area_input, group_by_var)
}
\arguments{
\item{profiled_area_input}{PARAM_DESCRIPTION}

\item{group_by_var}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready4s4]{lookup_tb}},\code{\link[ready4s4]{sp_starter_sf_lup}},\code{\link[ready4s4]{country}},\code{\link[ready4s4]{area_type}},\code{\link[ready4s4]{use_coord_lup}},\code{\link[ready4s4]{crs_nbr}},\code{\link[ready4s4]{features}}
 \code{\link[dplyr]{filter}}
 \code{\link[ready4utils]{data_get}}
 \code{\link[sf]{st_crs<-}}
 \code{\link[rlang]{sym}}
}
