% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make_sp_data_list.R
\name{get_starter_sf_for_profiled_area}
\alias{get_starter_sf_for_profiled_area}
\title{get_starter_sf_for_profiled_area}
\usage{
get_starter_sf_for_profiled_area(pa_r4, group_by_var)
}
\arguments{
\item{pa_r4}{PARAM_DESCRIPTION}

\item{group_by_var}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}
 \code{\link[ready4utils]{data_get}}
 \code{\link[sf]{st_crs<-}}
 \code{\link[rlang]{sym}}
}
