% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_sys_data_tbs_ls}
\alias{get_sys_data_tbs_ls}
\title{Get sys data tibbles}
\usage{
get_sys_data_tbs_ls()
}
\description{
get_sys_data_tbs_ls() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get sys data tibbles list. The function is called for its side effects and does not return a value.
}
\keyword{internal}
