% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_global_region.R
\name{global_region}
\alias{global_region}
\alias{global_region-ready4_macro}
\alias{global_region,ready4_macro-method}
\title{global_region}
\usage{
global_region(x)

\S4method{global_region}{ready4_macro}(x)
}
\arguments{
\item{x}{An object of class ready4_macro}
}
\description{
S4 Generic function to get the value of the slot global_region

Get the value of the slot global_region for S4 objects of class ready4_macro
}
