% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_sfs_keep_counts.R
\name{intersect_lon_lat_sfs}
\alias{intersect_lon_lat_sfs}
\title{intersect_lon_lat_sfs}
\usage{
intersect_lon_lat_sfs(sf_1, sf_2, crs_nbr_vec)
}
\arguments{
\item{sf_1}{PARAM_DESCRIPTION}

\item{sf_2}{PARAM_DESCRIPTION}

\item{crs_nbr_vec}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{geos_binary_ops}},\code{\link[sf]{st_transform}}
}
