% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_sfs_keep_counts.R
\name{intersect_sfs_keep_counts}
\alias{intersect_sfs_keep_counts}
\title{intersect_sfs_keep_counts}
\usage{
intersect_sfs_keep_counts(profiled_sf, profiled_colref = NA,
  profiled_rowref = NA, attribute_sf, attribute_unit, data_type,
  data_year, popl_var_prefix = NULL)
}
\arguments{
\item{profiled_sf}{PARAM_DESCRIPTION}

\item{profiled_colref}{PARAM_DESCRIPTION, Default: NA}

\item{profiled_rowref}{PARAM_DESCRIPTION, Default: NA}

\item{attribute_sf}{PARAM_DESCRIPTION}

\item{attribute_unit}{PARAM_DESCRIPTION}

\item{data_type}{PARAM_DESCRIPTION}

\item{data_year}{PARAM_DESCRIPTION}

\item{popl_var_prefix}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
intersect_sfs_keep_counts
}
\details{
This function:

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}},\code{\link[dplyr]{pull}},\code{\link[dplyr]{slice}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{funs}},\code{\link[dplyr]{filter}}
 \code{\link[stringr]{str_subset}}
 \code{\link[rlang]{sym}}
 \code{\link[sf]{geos_measures}},\code{\link[sf]{geos_binary_ops}}
 \code{\link[units]{set_units}}
}
