% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_intersect.R
\name{intersect_sfs_keep_counts}
\alias{intersect_sfs_keep_counts}
\title{Intersect sfs keep counts}
\usage{
intersect_sfs_keep_counts(
  profiled_sf,
  profiled_colref = NA,
  profiled_rowref = NA,
  attribute_sf,
  attribute_unit,
  data_type,
  data_year,
  crs_nbr_vec,
  popl_var_prefix = NULL
)
}
\arguments{
\item{profiled_sf}{Profiled (a simple features object)}

\item{profiled_colref}{PARAM_DESCRIPTION, Default: NA}

\item{profiled_rowref}{PARAM_DESCRIPTION, Default: NA}

\item{attribute_sf}{Attribute (a simple features object)}

\item{attribute_unit}{PARAM_DESCRIPTION}

\item{data_type}{PARAM_DESCRIPTION}

\item{data_year}{PARAM_DESCRIPTION}

\item{crs_nbr_vec}{PARAM_DESCRIPTION}

\item{popl_var_prefix}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
Profiled (a simple features object)
}
\description{
intersect_sfs_keep_counts() is an Intersect function that gets the intersection between two or more data objects. Specifically, this function implements an algorithm to intersect sfs keep counts. The function returns Profiled (a simple features object).
}
\keyword{internal}
