% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_sfs_update_counts.R
\name{intersect_sfs_update_counts}
\alias{intersect_sfs_update_counts}
\title{intersect_sfs_update_counts
Create a simple features object of two intersecting areas and adjust population counts by
fraction of spatial unit included in a profiled area.}
\usage{
intersect_sfs_update_counts(profiled_sf, profiled_colref = NA,
  profiled_rowref = NA, sp_data_list, tot_pop_resolution,
  age_sex_pop_resolution, group_by_var, age_sex_counts_grouped_by,
  data_year)
}
\arguments{
\item{profiled_sf}{PARAM_DESCRIPTION}

\item{profiled_colref}{PARAM_DESCRIPTION, Default: NA}

\item{profiled_rowref}{PARAM_DESCRIPTION, Default: NA}

\item{sp_data_list}{PARAM_DESCRIPTION}

\item{tot_pop_resolution}{PARAM_DESCRIPTION}

\item{age_sex_pop_resolution}{PARAM_DESCRIPTION}

\item{group_by_var}{PARAM_DESCRIPTION}

\item{age_sex_counts_grouped_by}{PARAM_DESCRIPTION}

\item{data_year}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
intersect_sfs_update_counts
Create a simple features object of two intersecting areas and adjust population counts by
fraction of spatial unit included in a profiled area.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{funs}}
}
