% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add_attr_to_sf.R
\name{make_attr_data_xx}
\alias{make_attr_data_xx}
\title{make_attr_data_xx}
\usage{
make_attr_data_xx(lookup_tb_r4, lookup_ref, starter_sf)
}
\arguments{
\item{lookup_tb_r4}{PARAM_DESCRIPTION}

\item{lookup_ref}{PARAM_DESCRIPTION}

\item{starter_sf}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready4utils]{data_get}}
 \code{\link[stats]{setNames}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{pull}}
}
