% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_pack.R
\name{make_data_pack_sngl}
\alias{make_data_pack_sngl}
\title{make_data_pack_sngl}
\usage{
make_data_pack_sngl(x, merge_with, pckg_name, raw_data_dir, processed_dir)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{merge_with}{PARAM_DESCRIPTION}

\item{pckg_name}{PARAM_DESCRIPTION}

\item{raw_data_dir}{PARAM_DESCRIPTION}

\item{processed_dir}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready4s4]{ready4_lookup}},\code{\link[ready4s4]{sp_import_lup<-}}
 \code{\link[dplyr]{pull}}
 \code{\link[purrr]{map2}}
}
