% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_intersect_sfs_keep_counts.R
\name{make_each_uid_a_poly_sf}
\alias{make_each_uid_a_poly_sf}
\title{make_each_uid_a_poly_sf}
\usage{
make_each_uid_a_poly_sf(sf, uid_chr)
}
\arguments{
\item{sf}{PARAM_DESCRIPTION}

\item{uid_chr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{pull}},\code{\link[dplyr]{summarise_all}},\code{\link[dplyr]{nth}}
 \code{\link[sf]{valid}},\code{\link[sf]{sf}},\code{\link[sf]{st_geometry}},\code{\link[sf]{geos_combine}},\code{\link[sf]{sfc}}
 \code{\link[rlang]{sym}}
 \code{\link[purrr]{map}},\code{\link[purrr]{reduce}}
}
