% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_import_object.R
\name{make_data_packs}
\alias{make_data_packs}
\title{make_import_object}
\usage{
make_data_packs(x, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Generic function to make import object
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
