% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_nse_objs_ls}
\alias{make_nse_objs_ls}
\title{Make non-standard evaluation objects list}
\usage{
make_nse_objs_ls(
  concept_1L_chr,
  spatial_unit_1L_chr,
  reference_var_nm_1L_chr = NULL,
  grouping_var_1L_chr = NULL,
  data_year_1L_chr,
  featured_var_pfx_1L_chr
)
}
\arguments{
\item{concept_1L_chr}{Concept (a character vector of length one)}

\item{spatial_unit_1L_chr}{Spatial unit (a character vector of length one)}

\item{reference_var_nm_1L_chr}{Reference variable name (a character vector of length one), Default: NULL}

\item{grouping_var_1L_chr}{Grouping variable (a character vector of length one), Default: NULL}

\item{data_year_1L_chr}{Data year (a character vector of length one)}

\item{featured_var_pfx_1L_chr}{Featured variable prefix (a character vector of length one)}
}
\value{
Non-standard evaluation objects (a list)
}
\description{
make_nse_objs_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make non-standard evaluation objects list. The function returns Non-standard evaluation objects (a list).
}
\keyword{internal}
