% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make_sp_data_list.R
\name{make_profiled_area_objs}
\alias{make_profiled_area_objs}
\title{make_profiled_area_objs}
\usage{
make_profiled_area_objs(pa_r4)
}
\arguments{
\item{pa_r4}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready4utils]{data_get}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{pull}}
 \code{\link[rlang]{sym}}
 \code{\link[sf]{st_transform}}
}
