% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sp_data_list.R
\name{make_profiled_area_objs}
\alias{make_profiled_area_objs}
\title{make_profiled_area_objs}
\usage{
make_profiled_area_objs(pa_r4)
}
\arguments{
\item{pa_r4}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready4s4]{use_coord_lup}},\code{\link[ready4s4]{lookup_tb}},\code{\link[ready4s4]{sp_starter_sf_lup}},\code{\link[ready4s4]{country}},\code{\link[ready4s4]{area_type}},\code{\link[ready4s4]{sp_site_coord_lup}},\code{\link[ready4s4]{features}},\code{\link[ready4s4]{geom_dist_limit_km}},\code{\link[ready4s4]{nbr_bands}},\code{\link[ready4s4]{crs_nbr}},\code{\link[ready4s4]{drive_time_limmit_mins}}
 \code{\link[ready4utils]{data_get}}
 \code{\link[dplyr]{filter}},\code{\link[dplyr]{pull}}
 \code{\link[rlang]{sym}}
 \code{\link[sf]{st_transform}}
}
