% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_sp_import_lup.R
\name{make_pt_ready4_sp_import_lup}
\alias{make_pt_ready4_sp_import_lup}
\title{Make prototype Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.}
\usage{
make_pt_ready4_sp_import_lup(
  file_type = character(0),
  file_name = character(0),
  data_repo = character(0),
  data_repo_ui = character(0),
  data_repo_db_ui = character(0),
  data_repo_file_ext = character(0),
  data_repo_save_type = character(0),
  local_file_src = character(0),
  make_script_src = character(0),
  download_url = character(0),
  inc_file_main = character(0),
  inc_files_to_rename = list(),
  new_names_for_inc_files = list(),
  name = character(0),
  country = character(0),
  area_type = character(0),
  area_bound_yr = character(0),
  region = character(0),
  data_type = character(0),
  main_feature = character(0),
  year = character(0),
  year_start = character(0),
  year_end = character(0),
  uid = character(0),
  add_boundaries = list()
)
}
\arguments{
\item{file_type}{PARAM_DESCRIPTION, Default: character(0)}

\item{file_name}{PARAM_DESCRIPTION, Default: character(0)}

\item{data_repo}{PARAM_DESCRIPTION, Default: character(0)}

\item{data_repo_ui}{PARAM_DESCRIPTION, Default: character(0)}

\item{data_repo_db_ui}{PARAM_DESCRIPTION, Default: character(0)}

\item{data_repo_file_ext}{PARAM_DESCRIPTION, Default: character(0)}

\item{data_repo_save_type}{PARAM_DESCRIPTION, Default: character(0)}

\item{local_file_src}{PARAM_DESCRIPTION, Default: character(0)}

\item{make_script_src}{PARAM_DESCRIPTION, Default: character(0)}

\item{download_url}{PARAM_DESCRIPTION, Default: character(0)}

\item{inc_file_main}{PARAM_DESCRIPTION, Default: character(0)}

\item{inc_files_to_rename}{PARAM_DESCRIPTION, Default: list()}

\item{new_names_for_inc_files}{PARAM_DESCRIPTION, Default: list()}

\item{name}{PARAM_DESCRIPTION, Default: character(0)}

\item{country}{PARAM_DESCRIPTION, Default: character(0)}

\item{area_type}{PARAM_DESCRIPTION, Default: character(0)}

\item{area_bound_yr}{PARAM_DESCRIPTION, Default: character(0)}

\item{region}{PARAM_DESCRIPTION, Default: character(0)}

\item{data_type}{PARAM_DESCRIPTION, Default: character(0)}

\item{main_feature}{PARAM_DESCRIPTION, Default: character(0)}

\item{year}{PARAM_DESCRIPTION, Default: character(0)}

\item{year_start}{PARAM_DESCRIPTION, Default: character(0)}

\item{year_end}{PARAM_DESCRIPTION, Default: character(0)}

\item{add_boundaries}{PARAM_DESCRIPTION, Default: list()}

\item{...}{Additional arguments, Default: character(0)}
}
\value{
A prototype for Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.
}
\description{
Create a new prototype for the Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.
}
