% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_data_pack.R
\name{make_raw_format_dir_str}
\alias{make_raw_format_dir_str}
\title{make_raw_format_dir_str}
\usage{
make_raw_format_dir_str(raw_data_dir, category)
}
\arguments{
\item{raw_data_dir}{PARAM_DESCRIPTION}

\item{category}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
