% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sim_data_env.R
\name{make_sim_data_env}
\alias{make_sim_data_env}
\title{make_sim_data_env
Make a simulation data input for the environment.}
\usage{
make_sim_data_env(profiled_area_type, profiled_area, age_lower, age_upper,
  project_for_year, env_str_par_tb, nbr_its, deterministic)
}
\arguments{
\item{profiled_area_type}{PARAM_DESCRIPTION}

\item{profiled_area}{PARAM_DESCRIPTION}

\item{age_lower}{PARAM_DESCRIPTION}

\item{age_upper}{PARAM_DESCRIPTION}

\item{env_str_par_tb}{PARAM_DESCRIPTION}

\item{nbr_its}{PARAM_DESCRIPTION}

\item{age_upper}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
make_sim_data_env
Make a simulation data input for the environment.
}
\details{
DETAILS
}
