% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sp_data_list.R
\name{make_sp_data_list}
\alias{make_sp_data_list}
\title{make_sp_data_list}
\usage{
make_sp_data_list(input_ls, sub_div_units_vec)
}
\arguments{
\item{input_ls}{PARAM_DESCRIPTION}

\item{sub_div_units_vec}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}},\code{\link[purrr]{transpose}},\code{\link[purrr]{prepend}}
 \code{\link[stats]{setNames}}
}
