% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_sfs_keep_counts.R
\name{make_valid_new_sf}
\alias{make_valid_new_sf}
\title{make_valid_new_sf}
\usage{
make_valid_new_sf(sf)
}
\arguments{
\item{sf}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[lwgeom]{valid}}
 \code{\link[sf]{st_cast}}
}
