% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get_sp_data.R
\name{make_year_filter_logic_vec}
\alias{make_year_filter_logic_vec}
\title{make_year_filter_logic_vec}
\usage{
make_year_filter_logic_vec(data_tb, included_years_vec)
}
\arguments{
\item{data_tb}{PARAM_DESCRIPTION}

\item{included_years_vec}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map2}}
}
