% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_year_vec}
\alias{make_year_vec}
\title{Make year vec}
\usage{
make_year_vec(input_ls)
}
\arguments{
\item{input_ls}{Input (a list)}
}
\description{
make_year_vec() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make year vec. The function is called for its side effects and does not return a value.
}
\keyword{internal}
