% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_manufacture.R
\name{manufacture.vicinity_parameters}
\alias{manufacture.vicinity_parameters}
\alias{manufacture,vicinity_parameters-method}
\alias{manufacture.vicinity_points}
\alias{manufacture,vicinity_points-method}
\alias{manufacture.vicinity_processed}
\alias{manufacture,vicinity_processed-method}
\alias{manufacture.vicinity_raw}
\alias{manufacture,vicinity_raw-method}
\title{Manufacture a new object}
\usage{
manufacture.vicinity_parameters(
  x,
  y_vicinity_mapes = NULL,
  n_its_int = integer(0),
  joint_dstr_1L_lgl = T,
  what_1L_chr = "values"
)

\S4method{manufacture}{vicinity_parameters}(
  x,
  y_vicinity_mapes = NULL,
  n_its_int = integer(0),
  joint_dstr_1L_lgl = T,
  what_1L_chr = "values"
)

manufacture.vicinity_points(
  x,
  bands_1L_dbl = numeric(0),
  crs_nbr_dbl = numeric(0),
  land_sf = NULL,
  metres_1L_dbl = numeric(0),
  service_1L_chr = character(0),
  time_min_1L_dbl = numeric(0),
  time_max_1L_dbl = numeric(0),
  time_steps_1L_dbl = numeric(0),
  travel_mode_1L_chr = "car",
  type_1L_chr = "single",
  what_1L_chr = "geometric"
)

\S4method{manufacture}{vicinity_points}(
  x,
  bands_1L_dbl = numeric(0),
  crs_nbr_dbl = numeric(0),
  land_sf = NULL,
  metres_1L_dbl = numeric(0),
  service_1L_chr = character(0),
  time_min_1L_dbl = numeric(0),
  time_max_1L_dbl = numeric(0),
  time_steps_1L_dbl = numeric(0),
  travel_mode_1L_chr = "car",
  type_1L_chr = "single",
  what_1L_chr = "geometric"
)

manufacture.vicinity_processed(
  x,
  approximation_1L_chr = "abs",
  main_incld_feature_chr = character(0),
  target_area_1L_chr = character(0),
  target_year_1L_chr = character(0),
  what_1L_chr = "closest years"
)

\S4method{manufacture}{vicinity_processed}(
  x,
  approximation_1L_chr = "abs",
  main_incld_feature_chr = character(0),
  target_area_1L_chr = character(0),
  target_year_1L_chr = character(0),
  what_1L_chr = "closest years"
)

manufacture.vicinity_raw(
  x,
  args_ls = NULL,
  crs_nbr_dbl = NA_real_,
  fn = function(x, ...) {
     NULL
 },
  match_value_xx = NULL,
  match_var_nm_1L_chr = character(0),
  merge_itms_chr = character(0),
  overwrite_1L_lgl = F,
  package_1L_chr = character(0),
  processed_fls_dir_1L_chr = character(0),
  raw_fls_dir_1L_chr = character(0),
  sub_dirs_chr = character(0),
  what_1L_chr = "output"
)

\S4method{manufacture}{vicinity_raw}(
  x,
  args_ls = NULL,
  crs_nbr_dbl = NA_real_,
  fn = function(x, ...) {
     NULL
 },
  match_value_xx = NULL,
  match_var_nm_1L_chr = character(0),
  merge_itms_chr = character(0),
  overwrite_1L_lgl = F,
  package_1L_chr = character(0),
  processed_fls_dir_1L_chr = character(0),
  raw_fls_dir_1L_chr = character(0),
  sub_dirs_chr = character(0),
  what_1L_chr = "output"
)
}
\arguments{
\item{x}{An instance of ready4 S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import.}

\item{y_vicinity_mapes}{PARAM_DESCRIPTION, Default: NULL}

\item{n_its_int}{N its (an integer vector), Default: integer(0)}

\item{joint_dstr_1L_lgl}{Joint distribution (a logical vector of length one), Default: T}

\item{what_1L_chr}{What (a character vector of length one), Default: 'output'}

\item{bands_1L_dbl}{Bands (a double vector of length one), Default: numeric(0)}

\item{crs_nbr_dbl}{Coordinates reference system number (a double vector), Default: NA}

\item{land_sf}{Land (a simple features object), Default: NULL}

\item{metres_1L_dbl}{Metres (a double vector of length one), Default: numeric(0)}

\item{service_1L_chr}{Service (a character vector of length one), Default: character(0)}

\item{time_min_1L_dbl}{Time minimum (a double vector of length one), Default: numeric(0)}

\item{time_max_1L_dbl}{Time maximum (a double vector of length one), Default: numeric(0)}

\item{time_steps_1L_dbl}{Time steps (a double vector of length one), Default: numeric(0)}

\item{travel_mode_1L_chr}{Travel mode (a character vector of length one), Default: 'car'}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'single'}

\item{approximation_1L_chr}{Approximation (a character vector of length one), Default: 'abs'}

\item{main_incld_feature_chr}{Main included feature (a character vector), Default: character(0)}

\item{target_area_1L_chr}{Target area (a character vector of length one), Default: character(0)}

\item{target_year_1L_chr}{Target year (a character vector of length one), Default: character(0)}

\item{args_ls}{Arguments (a list), Default: NULL}

\item{fn}{Function (a function), Default: function(x, ...) {
NULL
}}

\item{match_value_xx}{Match value (an output object of multiple potential types), Default: NULL}

\item{match_var_nm_1L_chr}{Match variable name (a character vector of length one), Default: character(0)}

\item{merge_itms_chr}{Merge items (a character vector), Default: character(0)}

\item{overwrite_1L_lgl}{Overwrite (a logical vector of length one), Default: F}

\item{package_1L_chr}{Package (a character vector of length one), Default: character(0)}

\item{processed_fls_dir_1L_chr}{Processed files directory (a character vector of length one), Default: character(0)}

\item{raw_fls_dir_1L_chr}{Raw files directory (a character vector of length one), Default: character(0)}

\item{sub_dirs_chr}{Sub directories (a character vector), Default: character(0)}
}
\value{
Object (an output object of multiple potential types)

Object (an output object of multiple potential types)

Object (an output object of multiple potential types)

Object (an output object of multiple potential types)
}
\description{
manufacture.vicinity_parameters() is a manufacture method that manufactures a novel R object using data contained in an instance of a class. This method is implemented for the ready4 S3 class for tibble object that stores simulation structural parameters relating to the spatial environment. The function returns Object (an output object of multiple potential types).

manufacture.vicinity_points() is a manufacture method that manufactures a novel R object using data contained in an instance of a class. This method is implemented for the ready4 S3 class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes. The function returns Object (an output object of multiple potential types).

manufacture.vicinity_processed() is a manufacture method that manufactures a novel R object using data contained in an instance of a class. This method is implemented for the ready4 S3 class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data). The function returns Object (an output object of multiple potential types).

manufacture.vicinity_raw() is a manufacture method that manufactures a novel R object using data contained in an instance of a class. This method is implemented for the ready4 S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import. The function returns Object (an output object of multiple potential types).
}
