% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_nbr_bands.R
\name{nbr_bands}
\alias{nbr_bands}
\alias{nbr_bands-ready4_profiled_area}
\title{nbr_bands}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_profiled_area}
}
\description{
S4 Generic function to get the value of the slot nbr_bands

Get the value of the slot nbr_bands for S4 objects of class ready4_profiled_area
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
