% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make_isochrs.R
\name{make_isochrs_for_1_srvc}
\alias{make_isochrs_for_1_srvc}
\title{make_isochrs_for_1_srvc}
\usage{
make_isochrs_for_1_srvc(
  cluster_tb,
  service,
  time_min,
  time_max,
  nbr_time_steps
)
}
\arguments{
\item{cluster_tb}{PARAM_DESCRIPTION}

\item{service}{PARAM_DESCRIPTION}

\item{time_min}{PARAM_DESCRIPTION}

\item{time_max}{PARAM_DESCRIPTION}

\item{nbr_time_steps}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}},\code{\link[dplyr]{pull}}
}
