% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_area_within_xmin_drive_of_points.R
\name{one_service_time_bands}
\alias{one_service_time_bands}
\title{one_service_time_bands}
\usage{
one_service_time_bands(look_up_ref, one_cluster_travel_time_sf_list)
}
\arguments{
\item{look_up_ref}{PARAM_DESCRIPTION}

\item{one_cluster_travel_time_sf_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{pluck}},\code{\link[purrr]{map}}
 \code{\link[dplyr]{pull}},\code{\link[dplyr]{filter}}
 \code{\link[stats]{setNames}}
 \code{\link[stringr]{str_replace}}
}
