% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_par_vals.R
\name{par_vals}
\alias{par_vals}
\alias{par_vals-ready4_env}
\title{par_vals}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_env}
}
\description{
S4 Generic function to get the value of the slot par_vals

Get the value of the slot par_vals for S4 objects of class ready4_env
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
