% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_data_pack.R
\name{write_procsd_geom_imp}
\alias{write_procsd_geom_imp}
\title{write_procsd_geom_imp}
\usage{
write_procsd_geom_imp(
  x,
  import_this_ls,
  path_to_starter_sf_chr,
  merge_with_chr_vec,
  crs_nbr_vec,
  overwrite_lgl
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{import_this_ls}{PARAM_DESCRIPTION}

\item{path_to_starter_sf_chr}{PARAM_DESCRIPTION}

\item{merge_with_chr_vec}{PARAM_DESCRIPTION}

\item{crs_nbr_vec}{PARAM_DESCRIPTION}

\item{overwrite_lgl}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ready4use]{assert_single_row_tb}}
 \code{\link[purrr]{reduce}}
 \code{\link[sf]{st_geometry_type}},\code{\link[sf]{geos_measures}}
 \code{\link[dplyr]{mutate}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{pull}}
 \code{\link[units]{set_units}}
}
