% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_epi_for_area.R
\name{project_epi_for_area}
\alias{project_epi_for_area}
\title{project_epi_for_area
Project epidemiology for area at specified point in time}
\usage{
project_epi_for_area(state_territory, profiled_area, age_lower, age_upper,
  project_for_year, env_str_par_tb, nbr_its, deterministic)
}
\arguments{
\item{state_territory}{PARAM_DESCRIPTION}

\item{profiled_area}{PARAM_DESCRIPTION}

\item{age_lower}{PARAM_DESCRIPTION}

\item{age_upper}{PARAM_DESCRIPTION}

\item{env_str_par_tb}{PARAM_DESCRIPTION}

\item{nbr_its}{PARAM_DESCRIPTION}

\item{age_upper}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
project_epi_for_area
Project epidemiology for area at specified point in time
}
\details{
DETAILS
}
