% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rready_space.R
\docType{package}
\name{ready.space-package}
\alias{ready.space}
\alias{ready.space-package}
\title{ready.space: Package of functions for modelling in space and time with the readfyforwhatsnext suite of resilience packages.}
\description{
ready.space is a package of functions that creates the spatial and temporal environment for the readyforwhatsnext simulation toolkit.
}
\details{
It has three main goals:
\itemize{
\item Goal 1.
\item Goal 2
\item Goal 3.
}
To learn more about ready.space, start with the vignettes:
`browseVignettes(package = "ready.space")`
}
\section{First section}{

\describe{
\item{firstitem}
}
}

\section{Second section}{

Some text.
\describe{
\item{moretext}
}
}

\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au} (0000-0001-7407-9194)

Other contributors:
\itemize{
  \item Glen Wiesner \email{Glen.Wiesner@vu.edu.au} (0000-0002-0071-130X) [reviewer]
  \item Alexandra Parker \email{Alex.Parker@vu.edu.au} (0000-0002-2398-6306) [reviewer]
  \item Cathrine Mihalopoulos \email{cathy.mihalopoulos@deakin.edu.au} (0000-0002-7127-9462) [reviewer]
  \item Jonathan Karnon \email{Jonathan.Karnon@flinders.edu.au} (0000-0003-3220-2099) [reviewer]
  \item Petra Plencnerova \email{Petra.Plencnerova@vu.edu.au} (0000-0001-9698-9084) [reviewer]
  \item Orygen [copyright holder, funder]
  \item VicHealth [funder]
  \item Victoria University [funder]
}

}
