% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_env.R
\docType{class}
\name{ready4_env}
\alias{ready4_env}
\title{ready4_env}
\description{
An S4 class to represent Spatiotemporal environment
}
\section{Slots}{

\describe{
\item{\code{st_data}}{list}

\item{\code{env_sf}}{sf}

\item{\code{par_vals}}{tbl_df}
}}

