% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_ready4_lookup.R
\name{ready4_lookup}
\alias{ready4_lookup}
\title{ready4_lookup}
\usage{
ready4_lookup(sp_abbreviations_lup = ready4_sp_abbreviations_lup(),
  sp_import_lup = ready4_sp_import_lup(),
  sp_data_pack_lup = ready4_sp_data_pack_lup(),
  sp_resolution_lup = ready4_sp_resolution_lup(),
  sp_site_coord_lup = ready4_sp_site_coord_lup(),
  sp_starter_sf_lup = ready4_sp_starter_sf_lup(),
  sp_uid_lup = ready4_sp_uid_lup())
}
\arguments{
\item{sp_abbreviations_lup}{ready4_sp_abbreviations_lup, Default: ready4_sp_abbreviations_lup()}

\item{sp_import_lup}{ready4_sp_import_lup, Default: ready4_sp_import_lup()}

\item{sp_data_pack_lup}{ready4_sp_data_pack_lup, Default: ready4_sp_data_pack_lup()}

\item{sp_resolution_lup}{ready4_sp_resolution_lup, Default: ready4_sp_resolution_lup()}

\item{sp_site_coord_lup}{ready4_sp_site_coord_lup, Default: ready4_sp_site_coord_lup()}

\item{sp_starter_sf_lup}{ready4_sp_starter_sf_lup, Default: ready4_sp_starter_sf_lup()}

\item{sp_uid_lup}{ready4_sp_uid_lup, Default: ready4_sp_uid_lup()}
}
\value{
An S4 object of the ready4_lookup class
}
\description{
An S4 class to represent Look up tables to use throughout readyforwhatsnext suite

Create a new S4 object of the class:ready4_lookup
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[methods]{new}}
}
