% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_macro.R
\docType{class}
\name{ready4_macro}
\alias{ready4_macro}
\title{ready4_macro}
\description{
An S4 class to represent Macro level context
}
\section{Slots}{

\describe{
\item{\code{global_region}}{character}
}}

