% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_meso_area.R
\docType{class}
\name{ready4_meso_area}
\alias{ready4_meso_area}
\title{ready4_meso_area}
\description{
An S4 class to represent Meso level context - area
}
\section{Slots}{

\describe{
\item{\code{area_type}}{character}

\item{\code{region_type}}{character}

\item{\code{global_region}}{character}
}}

