% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_meso_region.R
\docType{class}
\name{ready4_meso_region}
\alias{ready4_meso_region}
\title{ready4_meso_region}
\description{
An S4 class to represent Meso level context - region
}
\section{Slots}{

\describe{
\item{\code{region_type}}{character}

\item{\code{global_region}}{character}
}}

