% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_micro.R
\docType{class}
\name{ready4_micro}
\alias{ready4_micro}
\title{ready4_micro}
\description{
An S4 class to represent Micro level context
}
\section{Slots}{

\describe{
\item{\code{geom_dist_km_cuts}}{numeric}

\item{\code{travel_time_mins_cuts}}{numeric}

\item{\code{travel_mode}}{character}

\item{\code{area_type}}{character}

\item{\code{area}}{character}

\item{\code{area_bound_year}}{numeric}

\item{\code{region_type}}{character}

\item{\code{region}}{character}

\item{\code{region_bound_year}}{numeric}

\item{\code{global_region}}{character}

\item{\code{country}}{character}

\item{\code{country_bound_year}}{numeric}

\item{\code{lookup_tb}}{ready4_lookup}

\item{\code{crs_nbr}}{numeric}

\item{\code{temporal_min}}{POSIXt}

\item{\code{temporal_max}}{POSIXt}
}}

