% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_micro.R
\docType{class}
\name{ready4_micro}
\alias{ready4_micro}
\title{ready4_micro}
\usage{
ready4_micro(
  area_type = NA_character_,
  area = NA_character_,
  area_bound_year = NA_real_,
  region_type = NA_character_,
  region = NA_character_,
  region_bound_year = NA_real_,
  global_region = NA_character_,
  country = NA_character_,
  country_bound_year = NA_real_,
  lookup_tb = ready4_lookup(),
  crs_nbr = NA_real_,
  temporal_min = .POSIXct(NA_character_),
  temporal_max = .POSIXct(NA_character_),
  geom_dist_km_cuts = NA_real_,
  travel_time_mins_cuts = NA_real_,
  travel_mode = NA_character_
)
}
\arguments{
\item{area_type}{character, Default: 'NA'}

\item{area}{character, Default: 'NA'}

\item{area_bound_year}{numeric, Default: NA}

\item{region_type}{character, Default: 'NA'}

\item{region}{character, Default: 'NA'}

\item{region_bound_year}{numeric, Default: NA}

\item{global_region}{character, Default: 'NA'}

\item{country}{character, Default: 'NA'}

\item{country_bound_year}{numeric, Default: NA}

\item{lookup_tb}{ready4_lookup, Default: ready4_lookup()}

\item{crs_nbr}{numeric, Default: NA}

\item{temporal_min}{POSIXt, Default: .POSIXct(NA_character_)}

\item{temporal_max}{POSIXt, Default: .POSIXct(NA_character_)}

\item{geom_dist_km_cuts}{numeric, Default: NA}

\item{travel_time_mins_cuts}{numeric, Default: NA}

\item{travel_mode}{character, Default: 'NA'}
}
\value{
An S4 object of the ready4_micro class
}
\description{
An S4 class to represent Micro level context

Create a new S4 object of the class:ready4_micro
}
\section{Slots}{

\describe{
\item{\code{geom_dist_km_cuts}}{numeric}

\item{\code{travel_time_mins_cuts}}{numeric}

\item{\code{travel_mode}}{character}

\item{\code{area_type}}{character}

\item{\code{area}}{character}

\item{\code{area_bound_year}}{numeric}

\item{\code{region_type}}{character}

\item{\code{region}}{character}

\item{\code{region_bound_year}}{numeric}

\item{\code{global_region}}{character}

\item{\code{country}}{character}

\item{\code{country_bound_year}}{numeric}

\item{\code{lookup_tb}}{ready4_lookup}

\item{\code{crs_nbr}}{numeric}

\item{\code{temporal_min}}{POSIXt}

\item{\code{temporal_max}}{POSIXt}
}}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[methods]{new}}
}
