% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_ready4_profiled_area.R
\name{ready4_profiled_area}
\alias{ready4_profiled_area}
\title{ready4_profiled_area}
\usage{
ready4_profiled_area(country = NA_character_)
}
\arguments{
\item{country}{character, Default: 'NA'}
}
\value{
An S4 object of the ready4_profiled_area class
}
\description{
An S4 class to represent Information to create a profiled area object

Create a new S4 object of the class:ready4_profiled_area
}
\section{Slots}{

\describe{
\item{\code{country}}{character}
}}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[methods]{new}}
}
