% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_profiled_area.R
\docType{class}
\name{ready4_profiled_area}
\alias{ready4_profiled_area}
\title{ready4_profiled_area}
\usage{
ready4_profiled_area(
  geom_dist_km_cuts = NA_real_,
  travel_time_mins_cuts = NA_real_,
  travel_mode = NA_character_,
  area_type = NA_character_,
  area = NA_character_,
  area_bound_year = NA_real_,
  region_type = NA_character_,
  region = NA_character_,
  region_bound_year = NA_real_,
  global_region = NA_character_,
  country = NA_character_,
  country_bound_year = NA_real_,
  lookup_tb = ready4_lookup(),
  crs_nbr = NA_real_,
  temporal_min = .POSIXct(NA_character_),
  temporal_max = .POSIXct(NA_character_),
  features = NA_character_,
  use_coord_lup = NA,
  geom_dist_limit_km = NA_real_,
  drive_time_limit_mins = NA_real_,
  nbr_bands = NA_real_,
  data_year = NA_character_,
  data_ymds = .POSIXct(NA_character_)
)
}
\arguments{
\item{geom_dist_km_cuts}{numeric, Default: NA}

\item{travel_time_mins_cuts}{numeric, Default: NA}

\item{travel_mode}{character, Default: 'NA'}

\item{area_type}{character, Default: 'NA'}

\item{area}{character, Default: 'NA'}

\item{area_bound_year}{numeric, Default: NA}

\item{region_type}{character, Default: 'NA'}

\item{region}{character, Default: 'NA'}

\item{region_bound_year}{numeric, Default: NA}

\item{global_region}{character, Default: 'NA'}

\item{country}{character, Default: 'NA'}

\item{country_bound_year}{numeric, Default: NA}

\item{lookup_tb}{ready4_lookup, Default: ready4_lookup()}

\item{crs_nbr}{numeric, Default: NA}

\item{temporal_min}{POSIXt, Default: .POSIXct(NA_character_)}

\item{temporal_max}{POSIXt, Default: .POSIXct(NA_character_)}

\item{features}{character, Default: 'NA'}

\item{use_coord_lup}{logical, Default: NA}

\item{geom_dist_limit_km}{numeric, Default: NA}

\item{drive_time_limit_mins}{numeric, Default: NA}

\item{nbr_bands}{numeric, Default: NA}

\item{data_year}{character, Default: 'NA'}

\item{data_ymds}{POSIXt, Default: .POSIXct(NA_character_)}
}
\value{
An S4 object of the ready4_profiled_area class
}
\description{
An S4 class to represent Information to create a profiled area object

Create a new S4 object of the class:ready4_profiled_area
}
\section{Slots}{

\describe{
\item{\code{features}}{character}

\item{\code{use_coord_lup}}{logical}

\item{\code{geom_dist_limit_km}}{numeric}

\item{\code{drive_time_limit_mins}}{numeric}

\item{\code{nbr_bands}}{numeric}

\item{\code{data_year}}{character}

\item{\code{data_ymds}}{POSIXt}

\item{\code{geom_dist_km_cuts}}{numeric}

\item{\code{travel_time_mins_cuts}}{numeric}

\item{\code{travel_mode}}{character}

\item{\code{area_type}}{character}

\item{\code{area}}{character}

\item{\code{area_bound_year}}{numeric}

\item{\code{region_type}}{character}

\item{\code{region}}{character}

\item{\code{region_bound_year}}{numeric}

\item{\code{global_region}}{character}

\item{\code{country}}{character}

\item{\code{country_bound_year}}{numeric}

\item{\code{lookup_tb}}{ready4_lookup}

\item{\code{crs_nbr}}{numeric}

\item{\code{temporal_min}}{POSIXt}

\item{\code{temporal_max}}{POSIXt}
}}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[methods]{new}}
}
