% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ready4space.R, R/ready4space-package.R
\docType{package}
\name{ready4space-package}
\alias{ready4space}
\alias{ready4space-package}
\title{ready4space: Model Spatial Attributes with Ready4}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for developing geospatial models for use with the ready4 youth mental health systems model (https://ready4-dev.github.io/ready4/). This development version of the ready4space package has been made available as part of the process of testing and documenting the package. It is currently highly unstable and is not yet recommended for use. If you have any questions, please contact the authors (matthew.hamilton@orygen.org.au).

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for developing geospatial models for use with the ready4 youth mental health systems model (https://ready4-dev.github.io/ready4/). This development version of the ready4space package has been made available as part of the process of testing and documenting the package. It is currently highly unstable and is not yet recommended for use. If you have any questions, please contact the authors (matthew.hamilton@orygen.org.au).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/ready4space/}
  \item \url{https://github.com/ready4-dev/ready4space}
  \item \url{https://ready4-dev.github.io/ready4/}
}


Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/ready4space/}
  \item \url{https://github.com/ready4-dev/ready4space}
  \item \url{https://ready4-dev.github.io/ready4/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Authors:
\itemize{
  \item Caroline Gao \email{caroline.gao@orygen.org.au} (\href{https://orcid.org/0000-0002-0987-2759}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item Headspace [funder]
  \item National Health and Medical Research Council [funder]
}

}
\keyword{internal}
