% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ready4space.R
\docType{package}
\name{ready4space-package}
\alias{ready4space}
\alias{ready4space-package}
\title{ready4space: Standardised Modeller Tools for Retrieving, Managing and
Synthesising Spatial Data in Projects Developed with the Ready4 Suite}
\description{
ready4space provides a set of classes and
methods for spatial data management tasks throughout the ready4 suite
of tools for mental health data synthesis and modelling projects.
This development version of the ready4space package has been made
available as part of the process of testing and documenting the
package. The tools contained in this development release automate a
number of tasks which MODIFY THE DIRECTORY STRUCTURE OF YOUR LOCAL
MACHINE.  Therefore you should only trial this software if you feel
confident that you understand what it does and have created a sandpit
area in which you can safely undertake testing. If you have any
questions, please contact the authors
(matthew.hamilton@orygen.org.au).
}
\details{
To learn more about ready4space, start with the vignettes:
\code{browseVignettes(package = "ready4space")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/ready4space/}
  \item \url{https://github.com/ready4-dev/ready4space}
  \item \url{https://ready4-dev.github.io/ready4/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Authors:
\itemize{
  \item Glen Wiesner \email{Glen.Wiesner@vu.edu.au} (\href{https://orcid.org/0000-0002-0071-130X}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item VicHealth [funder]
  \item Victoria University [funder]
}

}
