% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add_attr_to_sf.R
\name{recur_add_attr_to_sf}
\alias{recur_add_attr_to_sf}
\title{recur_add_attr_to_sf}
\usage{
recur_add_attr_to_sf(
  input_ls,
  sub_div_unit = NULL,
  area_unit,
  boundary_year,
  attribute_data
)
}
\arguments{
\item{input_ls}{PARAM_DESCRIPTION}

\item{sub_div_unit}{PARAM_DESCRIPTION, Default: NULL}

\item{area_unit}{PARAM_DESCRIPTION}

\item{boundary_year}{PARAM_DESCRIPTION}

\item{attribute_data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Merges a SF file with population data relating to areas described in that file.
}
\details{
Makes data transformations (variable name, changes from strings to factors) necessary
to ensure that merged objects are compatible.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}
 \code{\link[ready4utils]{data_get}}
 \code{\link[stringr]{str_detect}}
 \code{\link[rlang]{sym}}
 \code{\link[purrr]{map}},\code{\link[purrr]{prepend}},\code{\link[purrr]{reduce}}
 \code{\link[stats]{setNames}}
}
