% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_region_bound_year.R
\name{region_bound_year<-}
\alias{region_bound_year<-}
\alias{region_bound_year<--ready4_meso_region}
\title{region_bound_year<-}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_meso_region}
}
\description{
S4 Generic function to set the value of the slot region_bound_year

Set the value of the slot region_bound_year for S4 objects of class ready4_meso_region
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
