% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_remove.R
\name{remove_grouped_popl_vars}
\alias{remove_grouped_popl_vars}
\title{Remove grouped popl vars}
\usage{
remove_grouped_popl_vars(profiled_sf, popl_var_prefix)
}
\arguments{
\item{profiled_sf}{Profiled (a simple features object)}

\item{popl_var_prefix}{PARAM_DESCRIPTION}
}
\description{
remove_grouped_popl_vars() is a Remove function that edits an object, removing a specified element or elements. Specifically, this function implements an algorithm to remove grouped popl vars. Function argument profiled_sf specifies the object to be updated. Argument popl_var_prefix provides the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
