% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_area_within_xkm_of_cluster.R
\name{reorder_distance_list_by_cluster}
\alias{reorder_distance_list_by_cluster}
\title{reorder_distance_list_by_cluster}
\usage{
reorder_distance_list_by_cluster(look_up_ref, clusters_by_distance_list,
  distances_vec)
}
\arguments{
\item{look_up_ref}{PARAM_DESCRIPTION}

\item{clusters_by_distance_list}{PARAM_DESCRIPTION}

\item{distances_vec}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}},\code{\link[purrr]{pluck}}
 \code{\link[dplyr]{mutate}}
}
