% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_reshape_pop_projs.R
\name{reshape_pop_projs}
\alias{reshape_pop_projs}
\title{reshape_pop_projs}
\usage{
reshape_pop_projs(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}},\code{\link[tibble]{tibble}},\code{\link[tibble]{as_tibble}}
 \code{\link[dplyr]{mutate_all}}
 \code{\link[stringr]{str_trim}}
}
