% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{save_raw}
\alias{save_raw}
\alias{save_raw.ready4_sp_import_lup}
\title{save_raw}
\usage{
save_raw(x, ...)

\method{save_raw}{ready4_sp_import_lup}(x, required_data,
  destination_directory, overwrite_lgl = F)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{required_data}{PARAM_DESCRIPTION}

\item{destination_directory}{PARAM_DESCRIPTION}

\item{overwrite_lgl}{PARAM_DESCRIPTION, Default: FALSE}

\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
}
