% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{save_raw.ready4_sp_import_lup}
\alias{save_raw.ready4_sp_import_lup}
\title{save_raw}
\usage{
\method{save_raw}{ready4_sp_import_lup}(x, required_data,
  destination_directory)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{required_data}{PARAM_DESCRIPTION}

\item{destination_directory}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
}
