% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make_sp_data_list.R
\name{simplify_geoms_in_lup}
\alias{simplify_geoms_in_lup}
\title{simplify_geoms_in_lup}
\usage{
simplify_geoms_in_lup(lup_r4, r_data_dir, crs_nbr)
}
\arguments{
\item{lup_r4}{PARAM_DESCRIPTION}

\item{r_data_dir}{PARAM_DESCRIPTION}

\item{crs_nbr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{pull}}
 \code{\link[purrr]{map}}
}
