% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_simplify.R
\name{simplify_sf}
\alias{simplify_sf}
\title{Simplify}
\usage{
simplify_sf(sf, crs = NULL)
}
\arguments{
\item{sf}{Simple features object (a simple features object)}

\item{...}{Additional arguments, Default: NULL}
}
\description{
simplify_sf() is a Simplify function that simplifies an object. Specifically, this function implements an algorithm to simplify simple features object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
