% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_lookup.R,
%   R/gnrc_sp_abbreviations_lup.R
\name{sp_abbreviations_lup-ready4_lookup}
\alias{sp_abbreviations_lup-ready4_lookup}
\alias{sp_abbreviations_lup}
\title{sp_abbreviations_lup}
\arguments{
\item{x}{An object of class ready4_lookup}

\item{x}{An object}
}
\description{
Get the value of the slot sp_abbreviations_lup for S4 objects of class ready4_lookup

S4 Generic function to get the value of the slot sp_abbreviations_lup
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
