% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_sp_resolution_lup.R
\name{sp_resolution_lup}
\alias{sp_resolution_lup}
\alias{sp_resolution_lup-ready4_lookup}
\title{sp_resolution_lup}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_lookup}
}
\description{
S4 Generic function to get the value of the slot sp_resolution_lup

Get the value of the slot sp_resolution_lup for S4 objects of class ready4_lookup
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
