% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_lookup.R, R/gnrc_sp_starter_sf_lup.R
\name{sp_starter_sf_lup<--ready4_lookup}
\alias{sp_starter_sf_lup<--ready4_lookup}
\alias{sp_starter_sf_lup<-}
\title{sp_starter_sf_lup<-}
\arguments{
\item{x}{An object of class ready4_lookup}

\item{x}{An object}
}
\description{
Set the value of the slot sp_starter_sf_lup for S4 objects of class ready4_lookup

S4 Generic function to set the value of the slot sp_starter_sf_lup
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
