% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_sp_uid_lup.R
\name{sp_uid_lup<-}
\alias{sp_uid_lup<-}
\alias{sp_uid_lup<--ready4_lookup}
\title{sp_uid_lup<-}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_lookup}
}
\description{
S4 Generic function to set the value of the slot sp_uid_lup

Set the value of the slot sp_uid_lup for S4 objects of class ready4_lookup
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
