% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make_srvc_clstr_isochrs.R
\name{make_srvc_clstr_isochrs}
\alias{make_srvc_clstr_isochrs}
\title{make_srvc_clstr_isochrs}
\usage{
make_srvc_clstr_isochrs(
  distance_km,
  clusters_vec,
  clusters_tbs_list,
  land_boundary_sf,
  crs_nbr
)
}
\arguments{
\item{distance_km}{PARAM_DESCRIPTION}

\item{clusters_vec}{PARAM_DESCRIPTION}

\item{clusters_tbs_list}{PARAM_DESCRIPTION}

\item{land_boundary_sf}{PARAM_DESCRIPTION}

\item{crs_nbr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Calculate area within a specified distance of a cluster of points.

FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}},\code{\link[purrr]{pluck}}
 \code{\link[stats]{setNames}}
}
