% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_area_within_xkm_of_points.R
\name{spatial_area_within_xkm_of_points}
\alias{spatial_area_within_xkm_of_points}
\title{Plots a profiled geographic unit based on distancw from service location.}
\usage{
spatial_area_within_xkm_of_points(point_locations, land_sf, distance,
  crs_nbr)
}
\arguments{
\item{point_locations}{A tibble...}

\item{land_sf}{A SF object ....}

\item{distance}{A double....}
}
\value{
A simple features object.
}
\description{
This function:
  -
  -
}
\details{

}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{st_as_sf}},\code{\link[sf]{st_transform}},\code{\link[sf]{geos_unary}},\code{\link[sf]{geos_combine}},\code{\link[sf]{geos_binary_ops}},\code{\link[sf]{sf}}
}
\concept{plot functions.}
