% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spatial_area_within_xkm_of_points.R
\name{spatial_area_within_xkm_of_points}
\alias{spatial_area_within_xkm_of_points}
\title{spatial_area_within_xkm_of_points}
\usage{
spatial_area_within_xkm_of_points(point_locations, land_sf, distance,
  crs_nbr)
}
\arguments{
\item{point_locations}{PARAM_DESCRIPTION}

\item{land_sf}{PARAM_DESCRIPTION}

\item{distance}{PARAM_DESCRIPTION}

\item{crs_nbr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Generates a SF based on distance from service location.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{st_as_sf}},\code{\link[sf]{st_transform}},\code{\link[sf]{geos_unary}},\code{\link[sf]{geos_combine}},\code{\link[sf]{geos_binary_ops}},\code{\link[sf]{sf}}
}
