% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_area_within_xmin_drive_of_points.R
\name{spatial_area_within_xmin_drive_of_points}
\alias{spatial_area_within_xmin_drive_of_points}
\alias{one_cluster_one_service_travel_time}
\alias{one_service_time_bands}
\alias{union_one_travel_time_band_across_sites}
\alias{fix_names_up_to_sfs}
\alias{cluster_isochrones}
\title{Calculate area within specified drives timeof a cluster of points.}
\usage{
spatial_area_within_xmin_drive_of_points(long, lat, time_min, time_max,
  nbr_time_steps)

one_cluster_one_service_travel_time(cluster_tb, service, time_min,
  time_max, nbr_time_steps)

one_service_time_bands(look_up_ref, one_cluster_travel_time_sf_list)

union_one_travel_time_band_across_sites(time_band_ref,
  one_cluster_time_bands_list)

fix_names_up_to_sfs(look_up_ref, one_cluster_up_to_xmin_list)

cluster_isochrones(cluster_tbs_list, look_up_ref, time_min = 0,
  time_max = 60, nbr_time_steps = 5)
}
\arguments{
\item{long}{PARAM_DESCRIPTION}

\item{lat}{PARAM_DESCRIPTION}

\item{time_min}{PARAM_DESCRIPTION}

\item{time_max}{PARAM_DESCRIPTION}

\item{nbr_time_steps}{PARAM_DESCRIPTION}

\item{cluster_tb}{PARAM_DESCRIPTION}

\item{service}{PARAM_DESCRIPTION}

\item{look_up_ref}{PARAM_DESCRIPTION}

\item{one_cluster_travel_time_sf_list}{PARAM_DESCRIPTION}

\item{time_band_ref}{PARAM_DESCRIPTION}

\item{one_cluster_time_bands_list}{PARAM_DESCRIPTION}

\item{one_cluster_up_to_xmin_list}{PARAM_DESCRIPTION}

\item{cluster_tbs_list}{PARAM_DESCRIPTION}

\item{look_up_ref}{PARAM_DESCRIPTION}

\item{look_up_ref}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\section{Functions}{
\itemize{
\item \code{one_cluster_one_service_travel_time}: Calculates the .....

\item \code{one_service_time_bands}: Calculates the .....

\item \code{union_one_travel_time_band_across_sites}: Calculates the .....

\item \code{fix_names_up_to_sfs}: Calculates the .....

\item \code{cluster_isochrones}: Calculates the .....
}}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[osrm]{osrmIsochrone}}
 \code{\link[sf]{st_as_sf}}
 \code{\link[dplyr]{mutate}},\code{\link[dplyr]{arrange}}
}
