% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_area_within_xmin_drive_of_points.R
\name{spatial_area_within_xmin_drive_of_points}
\alias{spatial_area_within_xmin_drive_of_points}
\title{spatial_area_within_xmin_drive_of_points}
\usage{
spatial_area_within_xmin_drive_of_points(long, lat, time_min, time_max,
  nbr_time_steps)
}
\arguments{
\item{long}{PARAM_DESCRIPTION}

\item{lat}{PARAM_DESCRIPTION}

\item{time_min}{PARAM_DESCRIPTION}

\item{time_max}{PARAM_DESCRIPTION}

\item{nbr_time_steps}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Calculate area within specified drives timeof a cluster of points.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[osrm]{osrmIsochrone}}
 \code{\link[sf]{st_as_sf}}
 \code{\link[dplyr]{mutate}},\code{\link[dplyr]{arrange}}
}
