% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_create_agent_coordinates_tb.R
\name{spatial_create_agent_coordinates_tb}
\alias{spatial_create_agent_coordinates_tb}
\title{Creates a tibble of coordinates for synthetic incident cases.}
\usage{
spatial_create_agent_coordinates_tb(profiled_area_sf, disorder, year,
  case_type = "expected.incidence", person_type = "p", resolution_unit)
}
\arguments{
\item{profiled_area_sf}{A SF object corresponding to the geographic unit that is to be profiled.}

\item{disorder}{A String ...}

\item{year}{A String ...}

\item{case_type}{A String ...}

\item{person_type}{A String ...}

\item{resolution_unit}{A String ...}
}
\value{
A simple features object.
}
\description{
This function:
  -
  -
}
\details{

}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_sub}}
 \code{\link[sf]{st_geometry}}
 \code{\link[dplyr]{select}},\code{\link[dplyr]{pull}},\code{\link[dplyr]{filter}}
 \code{\link[rlang]{sym}}
 \code{\link[purrr]{map2}}

Other spatial functions.: \code{\link{pop_sum_15.24}},
  \code{\link{spatial_adjust_partial_in_custom_area}},
  \code{\link{spatial_adjust_population_by_included_fraction}},
  \code{\link{spatial_cases_coordinates}},
  \code{\link{spatial_merge_areas_attributes}},
  \code{\link{spatial_one_year_projections_15_24}},
  \code{\link{spatial_population_growth}},
  \code{\link{spatial_profile_by_sa2}},
  \code{\link{spatial_select_rename_age_sex}},
  \code{\link{spatial_summarise_population_counts_0to29}},
  \code{\link{spatial_vic_pop_growth_lga}}
}
\concept{spatial functions.}
