% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_sfs_update_counts.R
\name{intersect_sfs_update_counts}
\alias{intersect_sfs_update_counts}
\title{intersect_sfs_update_counts
Create a simple features object of two intersecting areas and adjust population counts by
fraction of spatial unit included in a profiled area.}
\usage{
intersect_sfs_update_counts(profiled_sf, profiled_colref = NA,
  profiled_rowref = NA, sp_data_list, tot_pop_resolution,
  age_sex_pop_resolution, group_by_var, age_sex_counts_grouped_by,
  data_year)
}
\arguments{
\item{profiled_sf}{A SF object corresponding to the geographic unit that is to be profiled.}

\item{profiled_colref}{A variable name from the profiled_sf object that is the basis for
subsetting the object if not all areas from the object are to be used.}

\item{group_by_var}{A string specifying "SA1", "SA2" or "LGA" as the resolution of the
returned object,}

\item{resolution_sa1s_sf}{A simple features object comprised of SA1s}

\item{resolution_sa2s_sf}{A simple features object comprised of SA2s.}

\item{profiled_rworef}{A value from the profiled_colref variable which, if specified, will
be used to identify a subset of the profiled_sf object.}

\item{resolution_sf}{A SF object corresponding to the geographic unit that is resolution
at which the profiled_sf object will be profiled.}
}
\value{
A simple features object.
}
\description{
intersect_sfs_update_counts
Create a simple features object of two intersecting areas and adjust population counts by
fraction of spatial unit included in a profiled area.
}
\details{

}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
