% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_st_data.R, R/s4_ready4_env.R
\name{st_data}
\alias{st_data}
\alias{st_data-ready4_env}
\title{st_data}
\arguments{
\item{x}{An object}

\item{x}{An object of class ready4_env}
}
\description{
S4 Generic function to get the value of the slot st_data

Get the value of the slot st_data for S4 objects of class ready4_env
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
