% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sim_data_env.R
\name{subset_sf_by_feature}
\alias{subset_sf_by_feature}
\title{subset_sf_by_feature}
\usage{
subset_sf_by_feature(profiled_sf, group_by_var)
}
\arguments{
\item{profiled_sf}{PARAM_DESCRIPTION}

\item{group_by_var}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
 \code{\link[dplyr]{pull}},\code{\link[dplyr]{filter}}
 \code{\link[rlang]{sym}}
 \code{\link[stats]{setNames}}
}
