% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attr_to_sf.R
\name{subset_sf_ls_by_common_vars}
\alias{subset_sf_ls_by_common_vars}
\title{subset_sf_ls_by_common_vars}
\usage{
subset_sf_ls_by_common_vars(sf_ls)
}
\arguments{
\item{sf_ls}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
 \code{\link[dplyr]{select}}
}
