% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_counts_by_area.R
\name{suffix_to_prefix}
\alias{suffix_to_prefix}
\title{suffix_to_prefix}
\usage{
suffix_to_prefix(data_tb, suffix)
}
\arguments{
\item{data_tb}{PARAM_DESCRIPTION}

\item{suffix}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{funs}}
}
