% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_circles_to_bands}
\alias{transform_circles_to_bands}
\title{Transform circles to bands}
\usage{
transform_circles_to_bands(geom_distance_circle_sfs_list)
}
\arguments{
\item{geom_distance_circle_sfs_list}{PARAM_DESCRIPTION}
}
\description{
transform_circles_to_bands() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform circles to bands. Function argument geom_distance_circle_sfs_list specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
