% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_sfx_to_pfx}
\alias{transform_sfx_to_pfx}
\title{Transform suffix to prefix}
\usage{
transform_sfx_to_pfx(data_tb, suffix_1L_chr)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{suffix_1L_chr}{Suffix (a character vector of length one)}
}
\value{
Data (a tibble)
}
\description{
transform_sfx_to_pfx() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform suffix to prefix. Function argument data_tb specifies the object to be updated. Argument suffix_1L_chr provides the object to be updated. The function returns Data (a tibble).
}
\keyword{internal}
