% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spatial_area_within_xmin_drive_of_points.R
\name{union_one_travel_time_band_across_sites}
\alias{union_one_travel_time_band_across_sites}
\title{union_one_travel_time_band_across_sites}
\usage{
union_one_travel_time_band_across_sites(time_band_ref,
  one_cluster_time_bands_list)
}
\arguments{
\item{time_band_ref}{PARAM_DESCRIPTION}

\item{one_cluster_time_bands_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}},\code{\link[purrr]{pluck}},\code{\link[purrr]{reduce}}
 \code{\link[sf]{geos_combine}}
 \code{\link[dplyr]{select}}
}
