% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_counts_by_area.R
\name{update_pop_by_inc_area}
\alias{update_pop_by_inc_area}
\title{update_pop_by_inc_area}
\usage{
update_pop_by_inc_area(profiled_sf, sp_unit, data_year, concept,
  age_sex_var_name = NULL, age_sex_pop_resolution = NULL,
  tot_pop_col = NULL, popl_var_prefix)
}
\arguments{
\item{profiled_sf}{PARAM_DESCRIPTION}

\item{sp_unit}{PARAM_DESCRIPTION}

\item{data_year}{PARAM_DESCRIPTION}

\item{concept}{PARAM_DESCRIPTION}

\item{age_sex_var_name}{PARAM_DESCRIPTION, Default: NULL}

\item{age_sex_pop_resolution}{PARAM_DESCRIPTION, Default: NULL}

\item{tot_pop_col}{PARAM_DESCRIPTION, Default: NULL}

\item{popl_var_prefix}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{vars}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{funs}},\code{\link[dplyr]{select_all}}
 \code{\link[rlang]{sym}}
 \code{\link[sf]{geos_measures}}
 \code{\link[units]{set_units}}
}
