% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_pop_count_by_areas}
\alias{update_pop_count_by_areas}
\title{Update pop count by areas}
\usage{
update_pop_count_by_areas(
  profiled_sf,
  group_by_var,
  age_sex_var_name,
  data_year,
  age_sex_pop_resolution,
  tot_pop_resolution,
  popl_var_prefix = ""
)
}
\arguments{
\item{profiled_sf}{Profiled (a simple features object)}

\item{group_by_var}{PARAM_DESCRIPTION}

\item{age_sex_var_name}{PARAM_DESCRIPTION}

\item{data_year}{PARAM_DESCRIPTION}

\item{age_sex_pop_resolution}{PARAM_DESCRIPTION}

\item{tot_pop_resolution}{PARAM_DESCRIPTION}

\item{popl_var_prefix}{PARAM_DESCRIPTION, Default: ''}
}
\value{
Profiled (a simple features object)
}
\description{
update_pop_count_by_areas() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update pop count by areas. Function argument profiled_sf specifies the object to be updated. Argument group_by_var provides the object to be updated. The function returns Profiled (a simple features object).
}
\keyword{internal}
