% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_counts_by_area.R
\name{update_pop_count_by_areas}
\alias{update_pop_count_by_areas}
\title{update_pop_count_by_areas}
\usage{
update_pop_count_by_areas(profiled_sf, group_by_var, age_sex_var_name,
  data_year, age_sex_pop_resolution, tot_pop_resolution,
  popl_var_prefix = "")
}
\arguments{
\item{profiled_sf}{PARAM_DESCRIPTION}

\item{group_by_var}{PARAM_DESCRIPTION}

\item{age_sex_var_name}{PARAM_DESCRIPTION}

\item{data_year}{PARAM_DESCRIPTION}

\item{age_sex_pop_resolution}{PARAM_DESCRIPTION}

\item{tot_pop_resolution}{PARAM_DESCRIPTION}

\item{popl_var_prefix}{PARAM_DESCRIPTION, Default: ''}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Adjust population counts by fraction of spatial unit included in a profiled area.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
