% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_updateAttrDataXx.R
\name{updateAttrDataXx}
\alias{updateAttrDataXx}
\title{updateAttrDataXx
S4 Generic function to update attribute data object.}
\usage{
updateAttrDataXx(x, ..., verbose = TRUE)

\S4method{updateAttrDataXx}{ready4_lookup}(
  x,
  attr_data_xx,
  alt_names_sf,
  area_names_var_str,
  region_short_long_vec
)
}
\arguments{
\item{x}{An object}

\item{attr_data_xx}{An object}

\item{alt_names_sf}{A simple features object}

\item{area_names_var_str}{A character string}

\item{region_short_long_vec}{A character vector}
}
\description{
Return input object unaltered if no method for child class defined.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
\seealso{
\code{\link[methods]{setGeneric}}
}
