% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_updateAttrDataXx.R
\name{updateAttrDataXx}
\alias{updateAttrDataXx}
\title{updateAttrDataXx
S4 Generic function to update attribute data object.}
\arguments{
\item{x}{An object}
}
\description{
updateAttrDataXx
S4 Generic function to update attribute data object.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1#'  }
}
}
\seealso{
\code{\link[methods]{setGeneric}}
}
