% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_vicinity.R, R/vicinity-package.R
\docType{package}
\name{vicinity-package}
\alias{vicinity-package}
\title{vicinity: Model Spatial Features of Health Systems}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Modules from the readyforwhatsnext youth mental health economic model (https://readyforwhatsnext.org) that can be used to develop geospatial models. Designed for use in models developed with the ready4 framework (https://www.ready4-dev.com/). This development version of the vicinity package has been made available as part of the process of testing and documenting the package. It is currently highly unstable and is not yet recommended for use. If you have any questions, please contact the authors (matthew.hamilton1@monash.edu).

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Modules from the readyforwhatsnext youth mental health economic model (https://readyforwhatsnext.org) that can be used to develop geospatial models. Designed for use in models developed with the ready4 framework (https://www.ready4-dev.com/). This development version of the vicinity package has been made available as part of the process of testing and documenting the package. It is currently highly unstable and is not yet recommended for use. If you have any questions, please contact the authors (matthew.hamilton1@monash.edu).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/vicinity/}
  \item \url{https://github.com/ready4-dev/vicinity}
  \item \url{https://ready4-dev.github.io/ready4/}
}


Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/vicinity/}
  \item \url{https://github.com/ready4-dev/vicinity}
  \item \url{https://ready4-dev.github.io/ready4/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton1@monash.edu} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID}) [funder]

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item Australian Government Research Training Program [funder]
}

}
\keyword{internal}
