% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_VicinityMesoRegion.R
\docType{class}
\name{VicinityMesoRegion-class}
\alias{VicinityMesoRegion-class}
\alias{VicinityMesoRegion}
\title{VicinityMesoRegion}
\description{
Meso level context - region
}
\section{Slots}{

\describe{
\item{\code{region_type_chr}}{Region type (a character vector)}

\item{\code{region_chr}}{Region (a character vector)}

\item{\code{region_bndy_yr_dbl}}{Region boundary year (a double vector)}

\item{\code{a_VicinityLookup}}{(an instance of the VicinityLookup class)}

\item{\code{global_region_chr}}{Global region (a character vector)}

\item{\code{country_chr}}{Country (a character vector)}

\item{\code{country_bndy_yr_dbl}}{Country boundary year (a double vector)}

\item{\code{crs_dbl}}{Coordinates reference system (a double vector)}

\item{\code{temporal_min_dtm}}{Temporal minimum (a date vector)}

\item{\code{temporal_max_dtm}}{Temporal maximum (a date vector)}
}}

