% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_raw_format_dir}
\alias{write_raw_format_dir}
\title{Write raw format directory}
\usage{
write_raw_format_dir(data_type_chr, raw_fls_dir_1L_chr)
}
\arguments{
\item{data_type_chr}{Data type (a character vector)}

\item{raw_fls_dir_1L_chr}{Raw files directory (a character vector of length one)}
}
\value{
Path to raw output directory (a character vector of length one)
}
\description{
write_raw_format_dir() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write raw format directory. The function returns Path to raw output directory (a character vector of length one).
}
\keyword{internal}
